Koshucode.Baala.Base.Code.Clause

Description

This module provides a container for tokens. No tokens in Koshucode are in a extent of multiple lines. CodeLine includes whole tokens in a line. You can represent tokenized source code as a list of CodeLine.

CodeClause

data CodeClause a

Tokens in clause.

Constructors

CodeClause

Fields

clauseLines :: [CodeLine a]

Source lines of clause

clauseTokens :: [a]

Source tokens of clause

splitClause :: Gather [(IndentSize, a)] [a]

Split lines into clause based on indent size.

>>> splitClause [(0, "a"), (2, "b"), (0, "c"), (1, "d")]
(["a","b"], [(0,"c"),(1,"d")])