Koshucode.Baala.Base.Code.Infix

Description

Convert infixed-operator trees into prefixed-operator trees.

Documentation

type InfixHeight = Either Int Int

Direction and height for binary splitting.

Left H

Splits first the left-most operator, and operator height is H. For example, if the operator : is of Left 5, an expression (a : b : c) is splitted into (: a (b : c)) and then (: a (: b c)).

Right H

Right-most splitting and height H. For example, if the operator . is of Right 3, an expression (a . b . c) is splitted into (. (a . b) c) and then (. (. a b) c). An expression (a . b : c) is into (: (a . b) c), and then (: (. a b) c). Symbols that is not a binary operator like b are of height 0. For that reason, heights of expression (a . b : c) are (0 3 0 5 0)

infixHeight

Arguments

:: Ord b

=> (a -> Maybe b)
Connection function

-> [(b, InfixHeight)]
Height table

-> a -> InfixHeight
Height function

Make the height function from a height table of operators.

>>> infixHeight Just [('.', Right 3), (':', Right 5)] <$> "1.2:3"
[Left 0, Right 3, Left 0, Right 5, Left 0]

infixToPrefix

Arguments

:: (Map a, Map a, Map a)
Prefix/infix/postfix converters.

-> (a -> InfixHeight)
Height function

-> Collect (InfixTree p a)
Tree grouping function

-> InfixMapper p a

-> InfixConv a (CodeTree p a)
Infix-to-prefix conversion.

Split branches in a given tree at infixed binary operators.

type InfixMapper p a = InfixConv a [InfixTree p a] -> InfixConv a (InfixTree p a)

Intermediate infix-to-prefix conversion.

type InfixConv a tree = tree -> InfixAmb a tree

Warning: This is only used in defined module.

Infix-to-prefix conversion.

type InfixAmb a tree = Either [(InfixHeight, a)] tree

Warning: This is only used in defined module.

Code tree or ambiguous infix.

type InfixTree p a = CodeTree p (InfixHeight, a)

Warning: This is only used in defined module.

Code tree with infix height.