Koshucode.Baala.Base.Code.Line

Description

Text lines delimited by carriage returns or line feeds.

Line

type LineNumber = Int

Line number.

type NumberedLine = (LineNumber, String)

Line with number.

linesCrlf :: String -> [String]

Split string into lines. The result strings do not contain carriage returns (\r) and line feeds (\n).

>>> linesCrlf "aaa\nbbb\r\nccc\n\nddd\n"
["aaa","bbb","ccc","","ddd"]

linesCrlfNumbered :: String -> [NumberedLine]

Line number and its content.

linesCrlfBz :: ToCode code => code -> [Bz]

Split lazy bytestring by newline character sequence. This function drops the BOM sequence.

linesCrlfBzNumbered :: ToCode code => code -> [NumberedLine]

Create numbered lines from lazy bytestring.

dropBom :: ToCode code => code -> Bz

Remove UTF-8 BOM (EF BB BF) from lazy bytestring.

CodeLine

data CodeLine a

Tokens in line.

Constructors

CodeLine

Fields

lineNumber :: LineNumber

Line number, from 1.

lineContent :: String

Line content without newline.

lineTokens :: [a]

Tokens in the line.

type IndentSize = Int

Type for indent size.

lineIndentPair :: (a -> IndentSize) -> CodeLine a -> (IndentSize, CodeLine a)

Calculate indent of line and pairing it.