Koshucode.Baala.Base.Code.Scan

Description

Code scanner.

Type

data CodeScan k t

Code scanner divides input text into output tokens.

Constructors

CodeScan

Fields

scanMapSaved :: [CodeScanMap k t]

Saved updater

scanMap :: CodeScanMap k t

Updater

scanCp :: TCodePos t

Code position

scanInput :: t

Input text

scanOutput :: [k t]

Output tokens

scanCache :: WordCache t

Cached words

type CodeScanMap k t = Map (CodeScan k t)

Update code scanner.

type WordCache t = Cache t t

Cached words.

emptyWordCache :: WordCache t

Empty word cache.

Function

isBol :: Test (CodeScan k t)

Test scanner at the beginning of line, i.e., no output collected.

codeScanSave :: CodeScanMap k t

Save current updater.

codeScanRestore :: CodeScanMap k t

Restore saved updater.

codeScanUp :: (ToCode code, Textual t) => CodeScanMap k t -> IxIOPoint -> code -> [CodeLine k t]

Split source text into CodeLine list.

Split source text into lines by line delimiters (carriage return \r or line feed \n).

Numbering lines from 1. Internally, this is represented as a list of pairs (LineNumber, Textual).

Tokenize each lines, and put tokens together in CodeLine.

codeUpdate :: t -> k t -> CodeScanMap k t

Update scanInput and push result element to scanOutput.

codeUpdateWords :: WordCache t -> t -> k t -> CodeScanMap k t

Update code scanner with word table.

codeUpdateList :: t -> [k t] -> CodeScanMap k t

Multi-element version of codeUpdate.

codeUpdateListWords :: WordCache t -> t -> [k t] -> CodeScanMap k t

Multi-element and cached version of codeUpdate.

codeChange :: Map (CodeScanMap k t)

Change mapper of code sc.