Koshucode.Baala.Base.Content.Literalize

Description

Literalizer: Make literal contents from token tree.

Library

Types

type LiteralizeFrom b a = b -> AbOr a

Make a from b.

type LiteralizeTrees a = LiteralizeFrom [TokenTree] a

Make a from list of token trees.

type LiteralizeTree a = LiteralizeFrom TokenTree a

Make a from a token tree.

type LiteralizeString a = LiteralizeFrom String a

Make a from a string.

Functions

litContent :: CContent v => LiteralizeTree v

Transform TokenTree into internal form of term content.

litList :: CContent v => LiteralizeTrees [v]

Construct list of term contents.

litJudge :: CContent v => Bool -> Relsign -> LiteralizeTrees (Judge v)

Construct judge from token trees. Judges itself are not content type. It can be only used in the top-level of sections.

Document

Types

litContent recognizes the following types.

Boolean

Boolean used for something is hold or unhold. Textual forms: #true, #fasle.

Nil

Nil means that there are no values. i.e., universal negation on the term holds. Textual form is the non-quoted parens: ().

String

Sequence of characters. Textual forms: abc, 'abc', "abc def", text abc def.

Number

Number. Textual forms: int 100, dec 99.50, hex AF.

Set

Set is an unordered collection of contents. Duplication among contents is not significant. Textual form is a sequence of contents with braces: { a b c }.

List

List is an ordered list of contents. Textual form is a sequence of contents with square brackets: [ abc def ].

Termset

Termset is a set of terms, i.e., a list of named contents. Textual form is a sequence of terms with bar-braces: {| /a 10 /b 20 |}.

Relation

Relation is a set of same-type tuples, Textual form is a sequence of tuples enclosed in bar-bracket. The first tuple is a heading of relation, and succeeding tuples are delimited by vertical bar: [| /a /b | 10 20 | 30 40 |].

Simple data

Prepere some definitions.

>>> :m +Koshucode.Baala.Vanilla.Value.Val
>>> let trees = tokenTrees . tokens
>>> let lit  = litList [] . trees :: String -> AbortOr [Val]

Boolean.

>>> lit "#true #false"
Right [VBool True, VBool False]

Words.

>>> lit "a 'b c'"
Right [VString "a", VString "b c"]

>>> lit "a 1"
Right [VString "a", VString "1"]

Integer.

>>> lit "(int 12)"
Right [VInt 12]

Nil as no value.

>>> lit "()"
Right [VNil]

Compound data

Set.

>>> lit "{ b a a c a }"
Right [VSet [VString "b",VString "a",VString "c"]]

List.

>>> lit "[ a 10 (int 20) ]"
Right [VList [VString "a",VString "10",VInt 20]]

Termset.

>>> lit "{| /a 0 /b [ a 1 ] |}"
Right [VTermset
  [ ("/a", VString "0")
  , ("/b", VList [VString "a", VString "1"]) ]]

Relation.

>>> lit "[| /a /x | A1 (int 20) | A3 (int 40) | A4 (int 60) |]"
Right [ VRel (Rel
  (Relhead [Term "/a", Term "/x"])
  [ [VString "A1", VInt 20]
  , [VString "A3", VInt 40]
  , [VString "A4", VInt 60] ])]