Koshucode.Baala.Base.Data.Judge

Description

Judgements: a symbolic representations of affirmed or denied statements.

Datatype

data Judge c

Judgement on type c.

Judgement (or judge for short) is divided into three parts: logical quality, sign of relation, and argument. Bool value of logical quality corresponds to affirmed or denied judge. String value of sign represents certain sentence pattern that gives intepretation of data. Sentence pattern has placeholders filled by (String, c) values of argument. -

Constructors

Judge Bool JudgePattern [Named c]

type JudgePattern = String

Pattern of judgement.

Logical quality

affirmJudge :: Map (Judge c)

Affirm judge, i.e., change logical quality to True.

denyJudge :: Map (Judge c)

Deny judge, i.e., change logical quality to False.

isAffirmed :: Judge c -> Bool

isDenied :: Judge c -> Bool

Writer

putJudges :: (Ord c, Pretty c) => [Judge c] -> IO ()

Print judges.

hPutJudges :: (Ord c, Pretty c) => Handle -> [Judge c] -> IO ()

abcJudge :: Ord c => Map (Judge c)

Sort terms in alphabetical order.