Koshucode.Baala.Base.Data.Rel

Description

Type for relations

Documentation

data Rel v

Relations on type v. Heading of relation and body of relation as a list of tuples.

Constructors

Rel Relhead (Relbody v)

type Relbody v = [[v]]

List of positional args

type RelmapFun v = Rel v -> Rel v

Relation-to-relation mapping

reldum :: Rel v

Relational constant that has no terms and no tuples

reldee :: Rel v

Relational constant that has no terms and the empty tuple

module Koshucode.Baala.Base.Data.Rel.Relhead