Koshucode.Baala.Base.Data.Type

Documentation

data Type

Type for types.

Constructors

TypeAny
Everything

TypeType
Type of types

TypeTerm
Term name

TypeInterp
Data interpreation

TypeEmpty
Empty

TypeBool
Boolean

TypeText
Text

TypeCode
Code

TypeDec
Decimal

TypeClock (Maybe String)
Clock

TypeTime (Maybe String)
Time

TypeBin
Binary data

TypeList Type
List

TypeSet Type
Set

TypeTag String Type
Tagged type

TypeAssn [NamedType]
Association

TypeRel [NamedType]
Relation

TypeTuple [Type]
Tuple (Product type)

TypeSum [Type]
Sum type

type NamedType = Named Type

typeExplain :: Type -> Doc

Print type as tree.

typeFlatRel :: [TermName] -> Type

typeConsRel :: TermName -> Map Type

typeConsNest :: TermName -> Type -> Map Type

typeAppendRel :: [TermName] -> Map Type

typeRelTermNames :: Type -> [TermName]

typeRelDegree :: Type -> Int

typeRelIndex :: Type -> TermPath -> [Int]

Term index

>>> typeRelIndex (typeFlatRel ["a", "b", "c"]) ["b"]
[1]

>>> typeRelIndex (typeFlatRel ["a", "b", "c"]) ["e"]
[-1]

>>> typeRelIndex (typeConsNest "r" (typeFlatRel ["a", "b"]) (typeFlatRel [])) ["r", "b"]
[0, 1]

typeRelIndexList :: Type -> [TermPath] -> [[Int]]

typeTermDoc :: Type -> Doc

typeTerms :: Type -> [NamedType]

isTypeRel :: Type -> Bool

typeRelMapTerms :: Map [NamedType] -> Map Type

typeRelMapTerm :: Map NamedType -> Map Type

typeRelMapName :: Map TermName -> Map Type

Empty

Empty means that there are no values. i.e., universal negation on the term holds. Textual form is the non-quoted parens: ().

Boolean

Boolean used for something is hold or unhold. Textual forms: <0> (false), <1> (true).

Text

Sequence of characters. Textual forms is chars with apostrophe or doubly-quoted line: 'abc, "abc def".

Decimal

Decimal number. Textual forms is sequence of digits: 100, 99.50, hex AF.

Set

Set is an unordered collection of contents. Duplication among contents is not significant. Textual form is a sequence of contents delimited by colon, enclosed in braces: { 'a : 'b : 'c }.

List

List is an ordered list of contents. Textual form is a sequence of contents delimited by colon, enclosed in square brackets: [ 'abc : 'def ].

Assn

Assn is an association of terms, i.e., a list of named contents. Textual form is a sequence of terms with bar-angles: << /a 10 /b 20 >>.

Relation

Relation is a set of same-type tuples, Textual form is a sequence of tuples enclosed in bar-braces. The first tuple is a heading of relation, and succeeding tuples are delimited by vertical bar: {| /a : /b | 'A1 : 20 | 'A3 : 40 |}.