Koshucode.Baala.Base.IO.MixText

Description

Mix text.

Type

data MixText

Text mixable string, text, and bytestring.

Constructor

Text

mixBs :: ByteString -> MixText

Create mix text from strict bytestring.

mixBz :: ByteString -> MixText

Create mix text from lazy bytestring.

mixTx :: Text -> MixText

Create mix text from strict text.

mixTz :: Text -> MixText

Create mix text from lazy text.

mixString :: String -> MixText

Create mix text from string.

>>> mixString "abc" <> mixString "def"
MixText "abcdef"

mixChar :: Char -> MixText

Create mix text from char.

Padding

mixLeft :: Char -> Int -> MixText -> MixText

Put mix text to left edge in given-length char sequence.

>>> mixLeft '.' 10 $ mixString "abc"
MixText "abc......."

mixRight :: Char -> Int -> MixText -> MixText

Put mix text to right edge in given-length char sequence.

>>> mixRight '.' 10 $ mixString "abc"
MixText ".......abc"

Space

mixSpace :: Int -> MixText

Create mix text of given-length spaces.

>>> mixString "|" <> mixSpace 4 <> mixString "|"
MixText "|    |"

mixSep :: MixText -> MixText -> MixText

Infix mix space.

mixSep2 :: MixText -> MixText -> MixText

mix0 :: MixText

Empty space.

mix1 :: MixText

One space.

mix2 :: MixText

Two-length spaces.

mix3 :: MixText

Three-length spaces.

mix4 :: MixText

Four-length spaces.

Number

mixOct :: (Integral n, Show n) => n -> MixText

Mix text of octal number.

>>> mixOct (12 :: Int)
MixText "14"

mixDec :: (Integral n, Show n) => n -> MixText

Mix text of decimal number.

>>> mixDec (12 :: Int)
MixText "12"

mixHex :: (Integral n, Show n) => n -> MixText

Mix text of hexadecimal number.

>>> mixHex (12 :: Int)
MixText "c"

mixSign :: (Num n, Ord n) => n -> MixText

mixDecZero :: (Integral n, Show n) => Int -> n -> MixText

Zero-padding decimal number.

>>> mixDecZero 6 (123 :: Int)
MixText "000123"

>>> mixDecZero 6 (-123 :: Int)
MixText "-00123"

mixNumZero :: (Integral n, Show n) => (n -> MixText) -> Int -> n -> MixText

Newline

mixLine :: MixText -> MixText

Append line break.

mixSoft :: MixText

Soft line break, i.e., suppress at the beginning of line.

mixHard :: MixText

Hard line break, i.e., suppress only after auto line break.

Other

mixShow :: Show a => a -> MixText

mixEmpty :: MixText

Empty mix text.

>>> mixEmpty
MixText ""

Convert

ByteString

mixToBuilder :: LineBreak -> MixText -> Builder

Convert mix text to lazy bytestring builder.

mixToBz :: LineBreak -> MixText -> ByteString

Convert mix text to lazy bytestring.

mixToString :: LineBreak -> MixText -> String

Convert mix text to string.

mixToFlatString :: MixText -> String

Convert mix text to string without line breaks.

Mix I/O

putMix :: LineBreak -> MixText -> IO ()

Print mix text to the standard output.

putMixLn :: LineBreak -> MixText -> IO ()

Print mix text and newline to the standard output.

hPutMix :: LineBreak -> Handle -> MixText -> IO ()

Print mix text to the given output handler.

hPutMixLn :: LineBreak -> Handle -> MixText -> IO ()

Print mix text and newline to the given output handler.

putMixLines :: LineBreak -> [MixText] -> IO ()

Print mix text lines to the standard output.

hPutMixLines :: LineBreak -> Handle -> [MixText] -> IO ()

Print mix text lines to the given output handler.