Koshucode.Baala.Base.List.List

List

tails :: [a] -> [[a]]

Tails of list.

>>> tails "abc"
["abc", "bc", "c"]

isSingleton :: [a] -> Bool

Test list has single element.

notNull :: [a] -> Bool

Test list is not empty.

maybeEmpty :: Maybe a -> (a -> [b]) -> [b]

right :: b -> Map (Either a b)

Take

takeFirst :: [a] -> [a]

Take first element; if empty list is given, returns empty list.

>>> takeFirst "abc"
"a"

>>> takeFirst ""
""

takeLast :: [a] -> [a]

Take last element; if empty list is given, returns empty list.

>>> takeLast "abc"
"c"

takeOdd :: [a] -> [a]

Take elements at odd positions, i.e., first, third, ...

>>> takeOdd "abcdeft"
"acet"

takeEven :: [a] -> [a]

Take elements at even positions. i.e., second, fourth, ...

>>> takeEven "abcdeft"
"bdf"

takeFill :: a -> Int -> [a] -> [a]

Take N elements with filler.

>>> takeFill 0 3 [0 .. 8]
  [0,1,2]

>>> takeFill 0 5 [6 .. 8]
  [6,7,8,0,0]

takeTailFill :: a -> Int -> [a] -> [a]

Take tail-side N elements with filler.

>>> takeTailFill 0 3 [0 .. 8]
  [6,7,8]

>>> takeTailFill 0 5 [6 .. 8]
  [0,0,6,7,8]

Uniqueness

duplicates :: Ord a => [a] -> [a]

Keep duplicate elements.

unique :: Ord a => [a] -> [a]

Remove duplicate elements.

unionUp :: Eq a => [a] -> [a] -> [a]

Union list to base list.

intersectionFilter :: Ord a => [a] -> [a] -> [a]

setList :: Ord a => [a] -> [a]

Convert to set-like list, in other words, remove duplicate elements and sort list.

setEq :: Ord a => [a] -> [a] -> Bool

Set-like equality, in other words, duplication and ordering are ignored.

Examples

>>> duplicates "banana"
"ana"

>>> unique "banana"
"ban"

>>> unionUp "cde" "abc"
"deabc"

>>> List.union "cde" "abc"
"cdeab"

Construct

li1 :: a -> [a]

Make singleton list.

li2 :: a -> a -> [a]

li3 :: a -> a -> a -> [a]

consIf :: Bool -> a -> [a] -> [a]

type RangeBy a = a -> a -> [a]

rangeBy :: Ord a => Map a -> RangeBy a

zipMaybe :: [Maybe a] -> [b] -> [(a, b)]

zipMaybe2 :: [Maybe a] -> [b] -> [b]

Map

map2 :: (Functor f, Functor g) => (a -> b) -> f (g a) -> f (g b)

mapAt :: Map a -> Int -> Map [a]

mapWithLast :: (a -> b) -> (a -> b) -> [a] -> [b]

omit :: (a -> Bool) -> Map [a]

filterFst :: (a -> Bool) -> Map [(a, b)]

squeeze :: (a -> Bool) -> Map [a]

squeezeEmptyLines :: Map [String]

reverseMap :: Map [a] -> Map [a]