Koshucode.Baala.Base.List.Set

Description

Set-like operation.

Documentation

duplicates :: Ord a => [a] -> [a]

Extract duplicate elements.

>>> duplicates "banana"
"ana"

>>> duplicates "grape"
""

duplicated :: Ord a => Test [a]

Test list has duplicated elements.

>>> duplicated "banana"
True

>>> duplicated "grape"
False

unique :: Ord a => [a] -> [a]

Remove duplicate elements.

>>> unique "banana"
"ban"

uniqueConcat :: Ord a => [[a]] -> [a]

Collect elements in occurence order.

>>> uniqueConcat ["apple", "banana", "cocoa"]
"aplebnco"

unionUp :: Eq a => [a] -> [a] -> [a]

Union list to base list.

>>> unionUp "cde" "abc"
"deabc"

>>> List.union "cde" "abc"
"cdeab"

keepMember :: Ord a => [a] -> [a] -> [a]

Deprecated: Use selectShare instead.

Filter by set membership. keepMember A B is same to the intersection of A and B, except for the ordering of B is preserved.

>>> keepMember "cba" "abcdefg"
"abc"

>>> keepMember "abcdefg" "cba"
"cba"

omitMember :: Ord a => [a] -> [a] -> [a]

Deprecated: Use selectRight instead.

Anti-filter by set membership. omitMember A B is same to B minus fA, except for the ordering of B is preserved.

>>> omitMember "cba" "abcdefg"
"defg"

>>> omitMember "abcdefg" "cba"
""

sublist :: Ord a => [a] -> [a] -> Bool

Test sublist.

>>> sublist "cab" "abcdefg"
True

>>> sublist "cab" "dec"
False

setList :: Ord a => [a] -> [a]

Convert to set-like list, in other words, remove duplicate elements and sort list.

>>> setList "abracadabra"
"abcdr"

setEq :: Ord a => [a] -> [a] -> Bool

Set-like equality, in other words, duplication and ordering are ignored.

>>> setEq "abc" "bca"
True

>>> setEq "abc" "ab"
False

disjoint :: Eq a => [a] -> [a] -> Bool

Test two list has no elements in common.

>>> disjoint "abc" "de"
True

>>> disjoint "abc" "cde"
False

overlap :: Eq a => [a] -> [a] -> Bool

Test two list has some common elements.

>>> overlap "abc" "de"
False

>>> overlap "abc" "cde"
True