Koshucode.Baala.Base.MixText.MixText

Description

Mix text.

Type

data MixText

Text mixable string, text, and bytestring.

Constructor

Text

mixBs :: ByteString -> MixText

Create mix text from strict bytestring.

mixBz :: ByteString -> MixText

Create mix text from lazy bytestring.

mixTx :: Text -> MixText

Create mix text from strict text.

mixTz :: Text -> MixText

Create mix text from lazy text.

mixString :: String -> MixText

Create mix text from string.

>>> mixString "abc" <> mixString "def"
MixText "abcdef"

mixChar :: Char -> MixText

Create mix text from char.

Padding

mixLeft :: Char -> Int -> MixText -> MixText

Put mix text to left edge in given-length char sequence.

>>> mixLeft '.' 10 $ mixString "abc"
MixText "abc......."

mixRight :: Char -> Int -> MixText -> MixText

Put mix text to right edge in given-length char sequence.

>>> mixRight '.' 10 $ mixString "abc"
MixText ".......abc"

Space

mixSpace :: Int -> MixText

Create mix text of given-length spaces.

>>> mixString "|" <> mixSpace 4 <> mixString "|"
MixText "|    |"

mixSep :: MixText -> MixText -> MixText

Infix mix space.

>>> mixString "a" `mixSep` mixString "b"
MixText "a b"

mixSep2 :: MixText -> MixText -> MixText

Infix mix space.

>>> mixString "a" `mixSep2` mixString "b"
MixText "a  b"

mix0 :: MixText

Empty space.

mix1 :: MixText

One space.

mix2 :: MixText

Two-length spaces.

mix3 :: MixText

Three-length spaces.

mix4 :: MixText

Four-length spaces.

Number

mixOct :: (Integral n, Show n) => n -> MixText

Mix text of octal number.

>>> mixOct (12 :: Int)
MixText "14"

mixDec :: (Integral n, Show n) => n -> MixText

Mix text of decimal number.

>>> mixDec (12 :: Int)
MixText "12"

mixHex :: (Integral n, Show n) => n -> MixText

Mix text of hexadecimal number.

>>> mixHex (12 :: Int)
MixText "c"

mixSign :: (Num n, Ord n) => n -> MixText

Sign of number or space.

>>> mixSign $ O.int 3
MixText "+"

>>> mixSign $ O.int (-3)
MixText "-"

>>> mixSign $ O.int 0
MixText " "

mixDecZero :: (Integral n, Show n) => Int -> n -> MixText

Zero-padding decimal number.

>>> mixDecZero 6 $ O.int 123
MixText "000123"

>>> mixDecZero 6 $ O.int (-123)
MixText "-00123"

mixNumZero :: (Integral n, Show n) => (n -> MixText) -> Int -> n -> MixText

Generalized version of mixDecZero.

>>> mixNumZero mixHex 6 $ O.int 123
MixText "00007b"

Newline

mixLine :: MixText -> MixText

Append line break.

mixLines :: [MixText] -> MixText

Append line break to each mix texts.

mixSoft :: MixText

Soft line break, i.e., suppress at the beginning of line.

mixHard :: MixText

Hard line break, i.e., suppress only after auto line break.

mixBreak :: LineBreak -> MixText -> MixText

Mix text with local line break setting.

Block

mixBlock :: [MixText] -> MixText

Block of multiple mix texts.

mixBlockMap :: (MixText -> MixText) -> MixText -> MixText

Map function to mix texts directly in mix block.

>>> let m1 = mixBlock [mixString "cd", mixString "ef" <> mixBlock [mixString "g"]]
>>> let m2 = mixBlock [mixString "ab", m1]

>>> m2
MixText "abcdegh"

>>> mixBlockMap mixLine m2
MixText "ab\r\ncd\r\nefg\r\n"

>>> mixBlockMap (\x -> mixString "[" <> x <> mixString "]") m2
MixText "[ab][cd][efg]"

Other

mixShow :: Show a => a -> MixText

Showed mix text.

>>> mixShow (Just 'a')
MixText "Just 'a'"

mixEmpty :: MixText

Empty mix text.

>>> mixEmpty
MixText ""

Convert

ByteString

mixToBuilder :: LineBreak -> MixText -> Builder

Convert mix text to lazy bytestring builder.

mixToBz :: LineBreak -> MixText -> ByteString

Convert mix text to lazy bytestring.

mixToString :: LineBreak -> MixText -> String

Convert mix text to string.

mixToFlatString :: MixText -> String

Convert mix text to string without line breaks.

Mix I/O

putMix :: LineBreak -> MixText -> IO ()

Print mix text to the standard output.

putMixLn :: LineBreak -> MixText -> IO ()

Print mix text and newline to the standard output.

hPutMix :: LineBreak -> Handle -> MixText -> IO ()

Print mix text to the given output handler.

hPutMixLn :: LineBreak -> Handle -> MixText -> IO ()

Print mix text and newline to the given output handler.

putMixLines :: LineBreak -> [MixText] -> IO ()

Print mix text lines to the standard output.

hPutMixLines :: LineBreak -> Handle -> [MixText] -> IO ()

Print mix text lines to the given output handler.

writeMix :: LineBreak -> FilePath -> MixText -> IO ()

Write mix text to a file.

outputMix :: LineBreak -> Maybe FilePath -> MixText -> IO ()

Output mix text to a file when the file path is given, or to the standard output.