Koshucode.Baala.Base.Prelude.Pair

Map

mapFst :: (a -> c) -> (a, b) -> (c, b)

mapSnd :: (b -> c) -> (a, b) -> (a, c)

consFst :: a -> Map ([a], b)

consSnd :: b -> Map (a, [b])

mapFstTo :: Functor m => (a -> c) -> m (a, b) -> m (c, b)

mapSndTo :: Functor m => (b -> c) -> m (a, b) -> m (a, c)

Either

right2 :: (Either a b, Either a c) -> Either a (b, c)

right3 :: (Either a b, Either a c, Either a d) -> Either a (b, c, d)

right4 :: (Either a b, Either a c, Either a d, Either a e) -> Either a (b, c, d, e)

Monad

sequenceFst :: Monad m => [(m a, b)] -> m [(a, b)]

sequenceSnd :: Monad m => [(a, m b)] -> m [(a, b)]