Koshucode.Baala.Base.Prelude.Pretty

Class

class Pretty a where

Type that has a pretty printer.

Methods

doc :: a -> Doc

Doc of a.

docv :: [a] -> Doc

Doc joined by vcat.

doch :: [a] -> Doc

Doc joined by hsep.

Function

docColon :: Pretty a => [a] -> Doc

Colon-seperated document.

>>> docBracket $ docColon [True, False]
[ #true : #false ]

docWrap :: Pretty a => String -> String -> a -> Doc

Wrap in open and close brackets.

>>> docWrap "(" ")" "abc"
(abc)

docWraps :: Pretty a => String -> String -> a -> Doc

Wrap in open and close brackets. Put spaces between content and brackets.

>>> docWraps "(" ")" "abc"
( abc )

Excerpt from PrettyPrint

data Doc

The abstract type of documents. A Doc represents a *set* of layouts. A Doc with no occurrences of Union or NoDoc represents just one layout.

(<>) :: Doc -> Doc -> Doc

Beside. <> is associative, with identity empty.

(<+>) :: Doc -> Doc -> Doc

Beside, separated by space, unless one of the arguments is empty. <+> is associative, with identity empty.

($$) :: Doc -> Doc -> Doc

Above, except that if the last line of the first argument stops at least one position before the first line of the second begins, these two lines are overlapped. For example:

    text "hi" $$ nest 5 (text "there")

lays out as

    hi   there

rather than

    hi
         there

$$ is associative, with identity empty, and also satisfies

(x $$ y) <> z = x $$ (y <> z), if y non-empty.

nest :: Int -> Doc -> Doc

Nest (or indent) a document by a given number of positions (which may also be negative). nest satisfies the laws:

nest 0 x = x

nest k (nest k' x) = nest (k+k') x

nest k (x <> y) = nest k z <> nest k y

nest k (x $$ y) = nest k x $$ nest k y

nest k empty = empty

x <> nest k y = x <> y, if x non-empty

The side condition on the last law is needed because empty is a left identity for <>.

docEmpty :: Doc

docHang :: Doc -> Int -> Doc -> Doc

docZero :: String -> Doc