Koshucode.Baala.Base.Section

Description

Runtime structure for calculations written in Koshucode. Section is bundle of calculation stuff.

Modules

module Koshucode.Baala.Base.Section.Section

module Koshucode.Baala.Base.Section.Clause

module Koshucode.Baala.Base.Section.Clausify

module Koshucode.Baala.Base.Section.Quoter

module Koshucode.Baala.Base.Section.SectionIO

module Koshucode.Baala.Base.Section.SectionUnion

module Koshucode.Baala.Base.Section.Utility

Module dependencies

Section module uses Relmap, Data, Abort, Syntax, and Prelude modules.

Relmap module uses Data, Abort, Syntax, and Prelude modules.

Abort module uses Syntax and Prelude modules.

Syntax module uses Prelude module.

Data module uses Prelude module.

Prelude module uses no modules in Koshucode.

Following picture shows reduced dependencies.

      Section
        |
      Relmap
     /      \
  Data     Abort
    \        |
     \     Syntax
      \     /
      Prelude

Data dependencies

Dependencies of significant data types.

Section uses Section, Assert, Relmap, Judge and Construct.

Clause uses Clause, HalfRelmap, and Token.