Koshucode.Baala.Base.Syntax.Binary

Description

Convert infixed-operator trees into prefixed-operator trees

Documentation

type BinaryHeight = Either Int Int

Direction and height for binary splitting.

Left H

Splits first the left-most operator, and operator height is H. For example, if the operator : is of Left 5, an expression (a : b : c) is splitted into (: a (b : c)) and then (: a (: b c)).

Right H

Right-most splitting and height H. For example, if the operator . is of Right 3, an expression (a . b . c) is splitted into (. (a . b) c) and then (. (. a b) c). An expression (a . b : c) is into (: (a . b) c), and then (: (. a b) c). Symbols that is not a binary operator like b are of height 0. For that reason, heights of expression (a . b : c) are (0 3 0 5 0)

binaryTree

Arguments

:: Show a

=> (a -> BinaryHeight)
Height function

-> Tree a
Infixed tree

-> Tree a
Prefixed tree

Split branches in a given tree at infixed binary operators

heightTable

Arguments

:: Ord a

=> [(BinaryHeight, [a])]
Height table

-> a -> BinaryHeight
Height function

Make the height function from a height table of operators.

heightTableUnbox

Arguments

:: Ord a

=> (b -> a)
Unbox function

-> [(BinaryHeight, [a])]
Height table

-> b -> BinaryHeight
Height function