Koshucode.Baala.Base.Syntax.Code

Description

This module provides a container for tokens. No tokens in Koshucode are in a extent of multiple lines. CodeLine includes whole tokens in a line. You can represent tokenized source code as a list of CodeLine.

CodeLine

data CodeLine a

Tokens in line.

Constructors

CodeLine

Fields

lineNumber :: LineNumber

Line number, from 1.

lineContent :: String

Line content without newline.

lineTokens :: [a]

Tokens in the line.

data CodeClause a

Tokens in clause.

Constructors

CodeClause

Fields

clauseLines :: [CodeLine a]

Source lines of clause

clauseTokens :: [a]

Source tokens of clause

lineIndentPair :: (a -> Int) -> CodeLine a -> (Int, CodeLine a)

splitClause :: Gather [(Int, a)] [a]

CodeRoll

data CodeRoll a

Constructors

CodeRoll

Fields

codeMap :: AbMap (CodeRoll a)

codeInputPt :: CodePt

codeInput :: String

codeOutput :: [a]

codeRollUp :: AbMap (CodeRoll a) -> Resource -> String -> Ab [CodeLine a]

Split source text into CodeLine list.

Split source text into lines by line delimiters (carriage return \r or line feed \n).

Numbering lines from 1. Internally, this is represented as a list of pairs (LineNumber, String).

Tokenize each lines, and put tokens together in CodeLine.

codeUpdate :: String -> a -> Map (CodeRoll a)

codeChange :: AbMap (CodeRoll a) -> Map (CodeRoll a)