Koshucode.Baala.Base.Syntax.CodeLine

Description

This module provides a container for tokens. No tokens in Koshucode are in a extent of multiple lines. CodeLine includes whole tokens in a line. You can represent tokenized source code as a list of CodeLine.

Documentation

data CodeLine a

Tokens per line.

Constructors

CodeLine

Fields

codeLineNumber :: LineNumber

Line number, from 1.

codeLineContent :: String

Line content without newline.

codeLineTokens :: [a]

Tokens in the line.

type NextToken a

Arguments

= TokenNumber
Token number, from 1

-> String
Source text

-> (a, String)
Token and rest of text

Type of function that splits a next token from string. Tokens can includes TokenNumber.

type TokenNumber = Int

Token number, from 1.

codeLines

Arguments

:: NextToken a
Token splitter

-> String
Source text

-> [CodeLine a]
Token list per lines

Split source text into CodeLine list.

Split source text into lines by line delimiters (carriage return \r or line feed \n).

Numbering lines from 1. Internally, this is represented as a list of pairs (LineNumber, String).

Tokenize each lines, and put tokens together in CodeLine.