Koshucode.Baala.Base.Syntax.Infix

Description

Convert infixed-operator trees into prefixed-operator trees.

Documentation

type InfixHeight = Either Int Int

Direction and height for binary splitting.

Left H

Splits first the left-most operator, and operator height is H. For example, if the operator : is of Left 5, an expression (a : b : c) is splitted into (: a (b : c)) and then (: a (: b c)).

Right H

Right-most splitting and height H. For example, if the operator . is of Right 3, an expression (a . b . c) is splitted into (. (a . b) c) and then (. (. a b) c). An expression (a . b : c) is into (: (a . b) c), and then (: (. a b) c). Symbols that is not a binary operator like b are of height 0. For that reason, heights of expression (a . b : c) are (0 3 0 5 0)

infixHeight :: Ord b => (a -> Maybe b) -> [(b, InfixHeight)] -> a -> InfixHeight

Make the height function from a height table of operators.

infixToPrefix :: forall p a. (Map a, Map a, Map a) -> (a -> InfixHeight) -> Collect (InfixTree p a) -> InfixMapper p a -> InfixAbMap a (CodeTree p a)

Split branches in a given tree at infixed binary operators.

type InfixAb a x = Either [(InfixHeight, a)] x

type InfixAbMap a x = x -> InfixAb a x

type InfixTree p a = CodeTree p (InfixHeight, a)

type InfixMapper p a = InfixAbMap a [InfixTree p a] -> InfixAbMap a (InfixTree p a)