Koshucode.Baala.Base.Syntax.Line

Description

Text lines delimited by carriage returns or line feeds.

Documentation

type LineNumber = Int

Line number.

type NumberedLine = (LineNumber, String)

linesCrlfNumbered :: String -> [NumberedLine]

Line number and its content.

linesCrlf :: String -> [String]

Split string into lines. The result strings do not contain carriage returns (\r) and line feeds (\n).