Koshucode.Baala.Base.Syntax.Scan

Description

Code scanner.

Type

data CodeScan i o

Code scanner divides input text into output tokens.

Constructors

CodeScan

Fields

codeMapSaved :: [CodeScanMap i o]

Saved updater

codeMap :: CodeScanMap i o

Updater

codeInputPt :: CodePt

Code point

codeInput :: i

Input text

codeOutput :: [o]

Output tokens

codeWords :: WordTable

Collected words

type CodeScanMap i o = Map (CodeScan i o)

Update code scanner.

type WordTable = Map String String

Collected words.

Function

isBol :: CodeScan i o -> Bool

Test scanner at the beginning of line, i.e., no output collected.

codeScanSave :: CodeScanMap i o

Save current updater.

codeScanRestore :: CodeScanMap i o

Restore saved updater.

codeScanUp :: CodeScanMap String o -> NIOPoint -> String -> [CodeLine o]

Split source text into CodeLine list.

Split source text into lines by line delimiters (carriage return \r or line feed \n).

Numbering lines from 1. Internally, this is represented as a list of pairs (LineNumber, String).

Tokenize each lines, and put tokens together in CodeLine.

codeScanUpBz :: CodeScanMap String o -> NIOPoint -> Bz -> [CodeLine o]

Lazy bytestring version of codeScanUp.

codeUpdate :: i -> o -> CodeScanMap i o

Update codeInput and push result element to codeOutput.

codeUpdateWords :: WordTable -> i -> o -> CodeScanMap i o

Update code scanner with word table.

codeChange :: CodeScanMap i o -> CodeScanMap i o

Change mapper of code sc.