Koshucode.Baala.Base.Syntax.Tokenizer

Description

Tokenizer of koshucode.

Library

Tokenizer

tokens :: String -> [Token]

Split string into list of tokens. Result token list does not contain newline characters.

untokens :: [Token] -> String

Convert back a token list to a source string.

untoken :: Token -> String

divideByToken :: String -> [Token] -> [[Token]]

Source lines

data SourceLine

Constructors

SourceLine

Fields

sourceLineNumber :: Int

Line number, from 1.

sourceLineContent :: String

Line content without newline.

sourceLineTokens :: [Token]

Tokens from line.

sourceLines

Arguments

:: String
Source text in Koshucode.

-> [SourceLine]
Token list per lines.

Split source text into SourceLine list.

Split source code into lines by line delimiters (carriage return \r or line feed \n)

Numbering line numbers. Internally, this is represented as a list of pairs (line#, content).

Tokenize each lines. This is represented as a list of SourceLine line# content tokens.

Document

Special characters

Asterisks

There are three uses of asterisks (*) in koshucode,

*

Single asterisk means multiplication, e.g., 3 * 4 (three times four).

**

Double asterisk leads line comment. Textx from ** to end of line are ignored.

****

Quadruple asterisk (fourfold asterisk) leads caluse comment. Texts from **** to end of clause are ignored. In other words, line staring with **** and following indented lines are ignored.

Triple asterisk *** is double and rest of text. Quintuple (fivefold) asterisk ***** is quadruple and rest of text.

Line comments like this:

 ** aaa bbbbb cc

Clause comments like this:

 **** aaa bbb
      ccccc dddddd
      ee fffff

You can type **** on top of a clause to hide it.

Syntactic content type

(Not implemented)

Tokenizer recognizes five types of content.

Word

Simple sequence of characters, e.g., abc, 123. Numbers are represented as words.

Code

Code are like tagged word, e.g., (color 'blue').

List

Orderd sequence of contents. Lists are enclosed in square brackets, e.g., [ ab cd 0 1 ]

Tuple

Set of pairs of term name and content. Tuples are enclosed in curely braces, e.g., { /a 10 /b 20 }.

Relation

Set of uniform-typed tuples. Relations are enclosed in curely-bar braces, and enveloped tuples are divided by vertical bar, e.g., {| /a /b | 10 20 | 10 30 |}.

Examples

Words and quotations.

>>> tokens "aa\n'bb'\n\"cc\""
[TWord 0 "aa", TWord 1 "bb", TWord 2 "cc"]

Example includes TWord, TTermN and TSpace tokens.

>>> tokens "|-- R  /a A0 /b 31"
[TWord 0 "|--", TSpace 1, TWord 0 "R", TSpace 2,
 TTermN ["/a"], TSpace 1, TWord 0 "A0", TSpace 1,
 TTermN ["/b"], TSpace 1, TWord 0 "31"]

Parens.

>>> tokens "aa (bb x y (z))"
[TWord 1 0 "aa", TSpace 2 1,
 TOpen 3 "(", TWord 4 0 "bb", TSpace 5 1,
   TWord 6 0 "x", TSpace 7 1, TWord 8 0 "y", TSpace 9 1,
   TOpen 10 "(", TWord 11 0 "z", TClose 12 ")", TClose 13 ")"]

A comment.

>>> tokens "abc ** this is a comment\ndef\n"
[TWord 1 0 "abc", TSpace 2 1, TComment 3 "** this is a comment",
 TWord 4 0 "def"]