Koshucode.Baala.Base.Syntax.Tree

Data type

data Bracket p

Constructors

BracketNone

BracketOpen p

BracketClose p

data CodeTree p a

Tree of leaf and branch.

Constructors

TreeL a
Leaf. Terminal of tree.

TreeB p (Maybe (a, a)) [CodeTree p a]
Branch. Bracket-type and subtrees.

Parsing

tree :: (Ord p, CodePtr a) => GetBracketType p a -> Bracket p -> p -> [a] -> Ab (CodeTree p a)

Convert a list of elements to a single tree.

trees :: forall a p. (Ord p, CodePtr a) => GetBracketType p a -> Bracket p -> [a] -> Ab [CodeTree p a]

Convert a list of elements to trees.

treeWrap :: p -> [CodeTree p a] -> CodeTree p a

untree :: CodeTree p a -> [a]

Convert tree to list of tokens.

untrees :: [CodeTree p a] -> [a]

Convert tree to list of tokens.

undouble :: Pred p -> Map (CodeTree p a)

Simplify tree by removing double brackets, like ((a)) to (a).

>>> undouble (== 0) $ TreeB 0 Nothing [TreeB 0 Nothing [TreeL "A", TreeL "B"]]
TreeB 0 Nothing [TreeL "A", TreeL "B"]

Bracket table

type GetBracketType p a = a -> Bracket p

Get a bracket type.

bracketTable

Arguments

:: Eq a

=> [(p, Pred a, Pred a)]
List of (type, open, close)

-> GetBracketType p a

Make GetBracketType functions from a type-open-close table.

Make bracket/type functions from () and [].

>>> let bracket n [a, b] = (n, (== a), (== b))
>>> let pt = bracketTable [ bracket 1 "()", bracket 2 "[]" ]

Get bracket types for each chars. Types of open brackets are positive integer, and closes are negative.

>>> map pt "ab(cd[ef])g"
[0, 0, 1, 0, 0, 2, 0, 0, -2, -1, 0]