Koshucode.Baala.Base.Token.AngleText

Description

Text enclosed in angle brackets.

Functions

angleQuote :: Map String

Convert string into double-quoted and angle-quoted form.

angleTexts :: [(String, String)]

Table of coresspondences of angle text and its replacement.

Examples

Simple word.

>>> putStrLn $ angleQuote "aaa"
"aaa"

Text contains line feed.

>>> putStrLn $ angleQuote "aaa\nbbb"
"aaa" #lf "bbb"

Lookup keyword table.

>>> lookup "lf" angleTexts
Just "\n"

Keyword table

Some text literals are written using the following angle text.

<cr>

Carriage return

<crlf>

Carriage return and line feed

<lf>

Line feed

<q>

Single quote (')

<qq>

Double quote (")

<sp>

Space

<tab>

Tab