Koshucode.Baala.Core.Assert.Dataset

Description

Dataset as a set of judges.

Dataset is like a bridge of judges and relations. We can get a relation from a dataset, that dataset is build from judges.

Documentation

data Dataset c

Dataset is a set of judges.

type RelSelect c = JudgePattern -> [String] -> Rel c

Relation selector

emptyDataset :: Dataset c

Dataset that has no judges

dataset :: [Judge c] -> Dataset c

Gather judges into a dataset

addJudges :: [Judge c] -> Dataset c -> Dataset c

Add judges to dataset.

selectRelation

Arguments

:: (Ord c, CNil c)

=> Dataset c
Dataset

-> RelSelect c
Relation selector

Select relation from dataset. If a giving term is not in judges, CNil sign is used.