Koshucode.Baala.Core.Content.Cox

Description

Term-content expression.

Data types

data Cox c

Term-content expressions.

Constructors

CoxLit [CodePt] c
Literal content

CoxTerm [CodePt] [TermName] [Int]
Term reference, its name and position

CoxCalc [CodePt] BlankName (CopCalc c)
Content calculator

CoxLocal [CodePt] String Int
Local blank, its name and De Bruijn index

CoxBlank [CodePt] BlankName
Blank in form

CoxFill [CodePt] (Cox c) [Cox c]
Fill args in a form

CoxForm1 [CodePt] CoxTag String (Cox c)
Form with single blank

CoxForm [CodePt] CoxTag [String] (Cox c)
Form with multiple blanks

CoxWith [CodePt] [NamedCox c] (Cox c)
Cox with outside arguments

type MaybeCox c = Maybe (Cox c)

type NamedCox c = Named (Cox c)

type CopCalc c = [Ab c] -> Ab c

Term-content calculator.

type CoxTag = Maybe String

Functions

coxLit :: c -> Cox c

coxSyntacticArity :: Cox c -> Int

coxMap :: Map (Map (Cox c))

coxCall :: Cox c -> Map (Cox c) -> Cox c

checkIrreducible :: AbMap (Cox c)

Process

Phase 1. Convert input string into token trees.

1. String -> [Token@\

@] Parse input string into a list of tokens.

2. [Token@\

-> [TTree]@] Analyze token list structure.

3. [TTree@\

-> @TTree] Wrap the list of token trees in a tree.

Phase 2. Transform token trees, and convert into abstract syntax trees. See Koshucode.Baala.Core.Content.Build module.

4. TTree -> TTree

Expand tree-level syntax operators.

5. TTree -> TTree

Transform binary operators from infix to prefix.

6. TTree -> Cox c

Convert token tree into an abstract syntax tree.

7. Cox c -> Cox c

Attach De Bruijn indecies for bound variables.

8. Cox c -> Cox c

Expand expression-level syntax operators.

Phase 3. Calculate content expressions. See Koshucode.Baala.Core.Content.Run module.

9. Relhead -> Cox c -> Cox c

Attach term positions using actural heading of relation.

10. CopSet c -> [NamedCox@ c\

-> Cox c -> Cox c@] Replace base and derived operators into its bodies.

11. Cox c -> Beta c

Reduce derived expressions into beta expressions.

12. @Beta c -> [c\

-> c@] Run beta expression for each association of relation.