Koshucode.Baala.Core.Content.Literal

Description

Make literal contents from token tree.

Functions

type CalcContent c = TokenTreeToAb c

literal :: forall c. CContent c => CalcContent c -> TokenTreeToAb c

Convert TokenTree into internal form of content.

getJudge :: CContent c => CalcContent c -> Char -> JudgePat -> TokenTreesToAb (Judge c)

Convert token trees into a judge. Judges itself are not content type. It can be only used in the top-level of sections.

getTermedTrees :: TokenTreesToAb [NamedTree]

Convert token trees into a list of named token trees.

Example

>>> getTermedTrees =<< B.tt "/a 'A3 /b 10"
Right [("/a", TreeB 1 [TreeL (TText 3 0 "'"), TreeL (TText 4 0 "A3")]),
       ("/b", TreeL (TText 8 0 "10"))]

Document

Types

literal recognizes the following types.

Boolean

Boolean used for something is hold or unhold. Textual forms: #true, #fasle.

Empty

Empty means that there are no values. i.e., universal negation on the term holds. Textual form is the non-quoted parens: ().

Text

Sequence of characters. Textual forms is chars with apostrophe or doubly-quoted line: 'abc, "abc def".

Decimal

Decimal number. Textual forms is sequence of digits: 100, 99.50, hex AF.

Set

Set is an unordered collection of contents. Duplication among contents is not significant. Textual form is a sequence of contents delimited by colon, enclosed in braces: { 'a : 'b : 'c }.

List

List is an ordered list of contents. Textual form is a sequence of contents delimited by colon, enclosed in square brackets: [ 'abc : 'def ].

Assn

Assn is an association of terms, i.e., a list of named contents. Textual form is a sequence of terms with bar-angles: <| /a 10 /b 20 |>.

Relation

Relation is a set of same-type tuples, Textual form is a sequence of tuples enclosed in bar-braces. The first tuple is a heading of relation, and succeeding tuples are delimited by vertical bar: {| /a : /b | 'A1 : 20 | 'A3 : 40 |}.

Simple data

Prepere some definitions.

>>> :m +Koshucode.Baala.Op.Vanilla.Type
>>> let trees = B.tokenTrees . B.tokens
>>> let lit  = literal [] :: B.TokenTree -> B.Ab VContent
>>> let lits = literals lit . trees

Boolean.

>>> lits "#true : #false"
Right [VBool True, VBool False]

Words.

>>> lits "'a : 'b #sp 'c"
Right [VText "a", VText "b c"]

Decimal.

>>> lits "12.0"
Right [VDec (Decimal (120, 10), 1, False)]

Empty as no ordinary value.

>>> lits "()"
Right [VEmpty]

Compound data

Set.

>>> lits "{ 'b : 'a : 'a : 'c : 'a }"
Right [VSet [VText "b", VText "a", VText "c"]]

List.

>>> lits "[ 'a : '10 : 20 ]"
Right [VList [VText "a", VText "10", VDec (Decimal (20, 1), 0, False)]]

Assn.

>>> lits "<| /a 'x  /b { 'y : 'z } |>"
Right [VAssn
  [ ("/a", VText "x")
  , ("/b", VSet [VText "y", VText "z"])]]

Relation.

>>> lits "{| /a : /x | 'A1 : 20 | 'A3 : 40 | 'A4 : 60 |}"
Right [VRel (Rel
  (Relhead [Term "/a", Term "/x"]),
  [ [VText "A1", VDec (Decimal (20,1), 0, False)]
  , [VText "A3", VDec (Decimal (40,1), 0, False)]
  , [VText "A4", VDec (Decimal (60,1), 0, False)] ])]