Koshucode.Baala.Core.Content.Literal.HashWord

Documentation

Text literals are written using the following hashed words.

#q

Single quote.

#qq

Double quote.

#tab

Tab (\t).

#sp

Space.

#cr

Carriage return (\r).

#lf

Line feed (\n).

#crlf

Carriage return and line feed (\r\n).

hashWordTable :: [(String, String)]

Table of coresspondences of hashed sequence and its text.

>>> lookup "lf" hashWordTable
Just "\n"  

hashWord :: Map String

Convert string into hashed form.

>>> putStrLn $ hashWord "aaa"
"aaa"

>>> putStrLn $ hashWord "aaa\nbbb"
"aaa" #lf "bbb"