Koshucode.Baala.Core.Content.Literalize

Description

Literalizer: Make literal contents from token tree.

Library

Types

type LitTrees a = AbMap2 [TokenTree] a

Make a from list of token trees.

type LitTree a = AbMap2 TokenTree a

Make a from a token tree.

type LitOperators c = [Named (LitTree c -> LitTrees c)]

Functions

litContentBy :: CContent c => LitOperators c -> LitTree c

Transform TokenTree into internal form of term content.

litTermset :: CContent c => LitTree c -> LitTrees [Named c]

Collect term name and content.

Document

Types

litContentBy recognizes the following types.

Boolean

Boolean used for something is hold or unhold. Textual forms: #true, #fasle.

Nil

Nil means that there are no values. i.e., universal negation on the term holds. Textual form is the non-quoted parens: ().

Text

Sequence of characters. Textual forms is chars with apostrophe or doubly-quoted line: 'abc, ''abc def.

Decimal

Decimal number. Textual forms is sequence of digits: 100, 99.50, hex AF.

Set

Set is an unordered collection of contents. Duplication among contents is not significant. Textual form is a sequence of contents delimited by colon, enclosed in braces: { 'a : 'b : 'c }.

List

List is an ordered list of contents. Textual form is a sequence of contents delimited by colon, enclosed in square brackets: [ 'abc : 'def ].

Termset

Termset is a set of terms, i.e., a list of named contents. Textual form is a sequence of terms with bar-angles: <| /a 10 /b 20 |>.

Relation

Relation is a set of same-type tuples, Textual form is a sequence of tuples enclosed in bar-braces. The first tuple is a heading of relation, and succeeding tuples are delimited by vertical bar: {| /a /b | 'A1 : 20 | 'A3 : 40 |}.

Simple data

Prepere some definitions.

>>> :m +Koshucode.Baala.Vanilla.Value.Val
>>> let trees = tokenTrees . tokens
>>> let lit  = litList [] . trees :: String -> AbortOr [Val]

Boolean.

>>> lit "#true #false"
Right [VBool True, VBool False]

Words.

>>> lit "a 'b c'"
Right [VText "a", VText "b c"]

>>> lit "a 1"
Right [VText "a", VText "1"]

Integer.

>>> lit "(int 12)"
Right [VInt 12]

Nil as no value.

>>> lit "()"
Right [VNil]

Compound data

Set.

>>> lit "{ b a a c a }"
Right [VSet [VText "b", VText "a", VText "c"]]

List.

>>> lit "[ a 10 (int 20) ]"
Right [VList [VText "a", VText "10", VInt 20]]

Termset.

>>> lit "{| /a 0 /b [ a 1 ] |}"
Right [VTermset
  [ ("/a", VText "0")
  , ("/b", VList [VText "a", VText "1"]) ]]

Relation.

>>> lit "[| /a /x | A1 (int 20) | A3 (int 40) | A4 (int 60) |]"
Right [ VRel (Rel
  (Relhead [Term "/a", Term "/x"])
  [ [VText "A1", VInt 20]
  , [VText "A3", VInt 40]
  , [VText "A4", VInt 60] ])]