Koshucode.Baala.Core.Content.Tree

Description

Utilities for token trees.

Documentation

treesToTokens :: TTreesToAb [Token]

treesToTexts :: Bool -> TTreesToAb [String]

treeToText :: Bool -> TTreeToAb String

treesToDigits :: TTreesToAb String

tokenClock :: Token -> Ab Clock

treesToTime :: TTreesToAb Time

treeToFlatTerm :: TTreeToAb TermName

Get flat term name from token tree. If the token tree contains nested term name, this function failed.

treesToTerms :: TTreesToAb [NamedTrees]

Convert token trees into a list of named token trees.

treesToTerms1 :: TTreesToAb [NamedTree]

treesToInterp :: TTreesToAb Interp

Example

>>> treesToTerms =<< B.tt "/a 'A3 /b 10"
Right [("/a", [TreeL ...]),
       ("/b", [TreeL ...])]