Koshucode.Baala.Core.Lexmap.Attribute

Description

Relmap operator attributes.

Data type

type Roa = [NamedTrees]

Relmap operation attributes as association list.

type Roal a = ((String, Roa), a)

Association of operator use and something. Operator use is represented as pair of operator name and attributes.

type RoaSorter = [TokenTree] -> Ab Roa

Sorter for attribute of relmap operator. Sorters docompose attribute trees, and give a name to subattribute.

type RoaSpec = (AbMap Roa, [String], [String])

Attribute sorter for relmap operator. It consists of trunk sorter, trunk names, and branch names.

Branch sorter

roa :: [TokenTree] -> Roa

Split attribute into named group. Non quoted words beginning with hyphen, e.g., -x, are name of group.

>>> roa $ B.tt "a b -x /c 'd -y e"
[ ("@trunk", [TreeL (TText 1 0 "a"), TreeL (TText 3 0 "b")])
, ("-x", [TreeL (TTerm 7 ["/c"]), TreeL (TText 9 1 "d")])
, ("-y", [TreeL (TText 14 0 "e")]) ]

roaBranch :: RoaSorter

roaSorter :: RoaSpec -> RoaSorter

Trunk sorter

roaNone :: [String] -> RoaSpec

Attribute sorter for no-element trunk.

roaEnum :: [String] -> [String] -> RoaSpec

Attribute sorter for enumerating trunk.

roaList :: String -> [String] -> RoaSpec

Attribute sorter for multiple-element trunk.

roaOne :: String -> [String] -> RoaSpec

Attribute sorter for one-element trunk.

roaTwo :: String -> String -> [String] -> RoaSpec

Attribute sorter for two-element trunk.

roaThree :: String -> String -> String -> [String] -> RoaSpec

roaFour :: String -> String -> String -> String -> [String] -> RoaSpec

roaOneList :: String -> String -> [String] -> RoaSpec

Attribute sorter for one-and-multiple-element trunk.

roaOneOpt :: String -> String -> [String] -> RoaSpec

Examples

One required attribute and no options.

 roaOne "-term" []

Two required attributes and no options.

 roaTwo "-name" "-relmap" []

Any number of attributes and no options.

 roaList "-term" []

One and any number of attributes.

 roaOneList "-pattern" "-term" []