Koshucode.Baala.Core.Lexmap.Lexmap

Description

Lexical relmap.

Data type

data Lexmap

Intermediate data that represents use of relmap operator. Lexmap is constructed from a list of TTree, and generic relmap is constructed from a lexmap.

Constructors

Lexmap

Fields

lexType :: LexmapType

Type of lexmap

lexRopToken :: Token

Token of operator

lexAttr :: [AttrTree]

Attribute of relmap operation

lexSubmap :: [Lexmap]

Submaps in the attribute

lexMessage :: [String]

Messages on lexmap

data LexmapType

Constructors

LexmapBase
Built-in relmap

LexmapDerived
User-defined relmap

LexmapNest
Nested relation reference

lexRopName :: Lexmap -> RopName

Name of relmap operator

lexAddMessage :: String -> Map Lexmap

lexMessageList :: Lexmap -> [String]

Constructor

consLexmap :: (RopName -> Maybe AttrSort) -> ConsLexmap

First step of constructing relmap, construct lexmap from token trees. The function returns lexmap and related lexmap links.

nestTerms :: [TTree] -> Ab [Terminal RopName]

Parse nested relation attribute.

type ConsLexmap = [GlobalSlot] -> FindRelmap -> ConsLexmapBody

type ConsLexmapBody = SecNo -> [TTree] -> Ab (Lexmap, LexmapLinkTable)

type SecNo = Int

Section number.

type FindRelmap = SecNo -> RopName -> [RelmapSource]

type LexmapLinkTable = [(Lexmap, Lexmap)]

type RelmapSource = NNamed ([TTree], Attrmap)

Source of relmap: its name, replacement, and attribute editor.

type NName = (SecNo, String)

Numbered name.

type NNamed a = (NName, a)

Pair which key is a numbered name.