Koshucode.Baala.Core.Relmap.Operand

Description

Sorter for operand of relmap operator.

Full sorter

type RopFullSorter = [TokenTree] -> Ab RopOperandAssoc

Sorter for operand of relmap operator. Sorters docompose operand trees, and give a name to suboperand.

ropFullSorter :: RopOperandSorter -> RopFullSorter

Branch sorter

type RopOperandAssoc = [Named [TokenTree]]

Relmap operand as association list.

ropOperandAssoc :: [TokenTree] -> RopOperandAssoc

Split operand into named group. Non quoted words beginning with hyphen, e.g., -x, are name of group.

>>> ropOperandAssoc $ B.tt "a b -x /c 'd -y e"
[ ("",   [TreeL (TWord 1 0 "a"), TreeL (TWord 3 0 "b")])
, ("-x", [TreeL (TTerm 7 ["/c"]), TreeL (TWord 9 1 "d")])
, ("-y", [TreeL (TWord 14 0 "e")]) ]

Trunk sorter

type RopOperandSorter = (RopTrunkSorter, [String], [String])

Operand sorter for relmap operator. It consists of trunk sorter, trunk names, and branch names. Trunk part is unnamed operand. Branch part is named operand.

type RopTrunkSorter = AbMap RopOperandAssoc

Mapping from basically sorted operand to fully sorted operand.

sortNone :: [String] -> RopOperandSorter

Operand sorter for no-element trunk.

sortEnum :: [String] -> [String] -> RopOperandSorter

Operand sorter for enumerating trunk.

sortList :: String -> [String] -> RopOperandSorter

Operand sorter for multiple-element trunk.

sortOne :: String -> [String] -> RopOperandSorter

Operand sorter for one-element trunk.

sortTwo :: String -> String -> [String] -> RopOperandSorter

Operand sorter for two-element trunk.

sortThree :: String -> String -> String -> [String] -> RopOperandSorter

sortFour :: String -> String -> String -> String -> [String] -> RopOperandSorter

sortOneList :: String -> String -> [String] -> RopOperandSorter

Operand sorter for one-and-multiple-element trunk.