Koshucode.Baala.Core.Relmap.Relmap

Description

Operations on Relmap.

Select from relmap

relmapSourceList :: Relmap c -> [Relmap c]

List of RelmapSource

relmapNameList :: Relmap c -> [String]

List of name in RelmapLink

Construct relmap

relmapSource :: RopUse c -> JudgePattern -> [Termname] -> Relmap c

Retrieve relation from dataset.

relmapConst :: RopUse c -> Rel c -> Relmap c

Make a constant relmap.

relmapFlow :: RopUse c -> RelkitCalc c -> Relmap c

Make a flow relmap. Flow relmaps take no subrelmaps.

relmapGlobal :: RopUse c -> RelkitGlobal c -> Relmap c

Make a global relmap. Global relmaps are flow relmaps with globals.

relmapBinary :: RopUse c -> RelkitBinary c -> Relmap c -> Relmap c

Make a binary relmap. Binary relmaps take one subrelmap.

relmapConfl :: RopUse c -> RelkitConfl c -> [Relmap c] -> Relmap c

Make a confluent relmap. Confluent relmaps take multiple subrelmaps.

Specialize relmap to relkit

relmapSpecialize :: forall c. Global c -> [RelmapDef c] -> [RelkitDef c] -> Maybe Relhead -> Relmap c -> Ab ([RelkitDef c], Relkit c)

Append relmaps

This picture represents calculation of mapping input relation to output relation.

 input -[ relmap ]- output

Relmap A maps relation R1 to relation R2. Another relmap B maps R2 to R3.

 R1 -[ A ]- R2
 R2 -[ B ]- R3

Two relmaps A and B are jointed with intermidiate relation R2.

 R1 -[ A ]- R2 -[ B ]- R3

Or, we can draw a directly jointed picture. This whole structure is also RelmapAppend A B.

 R1 -[ A ]--[ B ]- R3