Koshucode.Baala.Core.Relmap.Specialize

Description

Specialize generic relmaps. Generic relmaps are represented by Relmap, specialized relmaps by Relkit.

Documentation

relmapSpecialize

Arguments

:: h c
Hook.

-> RelmapLinkTable' h c
Lexmap-to-relmap search table.

-> RelkitTable c
List of known relkits, i.e., search table from relkit key to actual relkit.

-> Maybe Head
Input heading feeding into generic relmap.

-> Relmap' h c
Generic relmap to specialize.

-> Ab (RelkitTable c, Relkit c)
Extended table and specialized relmap.

Specializes generic relmap to specialized relmap. Specialized relmaps have fixed input/output headings. In constrast, generic relmaps do not. Specialized relmaps are also called relkits. This function returns relkit and extended table of relkits.

type RelmapLinkTable' h c = [(Lexmap, Relmap' h c)]

Table for referencing relmap of RelmapLink.