Koshucode.Baala.Core.Resource.Resource

Description

Internal representation of relational data resource. Data resource bundles data themselves and calculation stuffs. In Koshucode, data is represented as set of judgements, and calculation is as relation-to-relation mapping. This internal representation is usually constructed from computer files or web resources addressed by URI. Result of calculation on some resource becomes also data resource.

Data type

data Resource c

Relational data resource

Constructors

Resource

Fields

resGlobal :: Global c

Global parameter

resOption :: Option c

Options

resImport :: [Resource c]

Importing resources

resExport :: [String]

Exporting names

resSlot :: [NamedTrees]

Global slots

resLexmap :: [LexmapClause]

Source of relmaps

resAssert :: [ShortAssert c]

Assertions of relmaps

resJudge :: [Judge c]

Affirmative or denial judgements

resInputQueue :: InputQueue

Input points

resOutput :: IOPoint

Output point

resEcho :: [[TokenLine]]

Echo text

resLicense :: [(SecNo, String)]

License text

resMessage :: [String]

Collection of messages

resLastSecNo :: SecNo

Last section number

resSelect :: RelSelect c

Relation selector

type AbResource c = Ab (Resource c)

Abort or resource.

resIncluded :: Resource c -> [NIOPoint]

Included sources, i.e., done-part of input queue.

resInput :: Resource c -> [IOPoint]

resInputPoint :: Resource c -> [InputPoint]

All input points.

resClass :: Resource c -> [JudgeClass]

List of all judgement classes.

addMessage :: String -> Map (Resource c)

Add single message.

addMessages :: [String] -> Map (Resource c)

Add messages.

Input queue

type InputQueue = (Queue InputPoint, [NIOPoint])

Queue for input code: todo, ready and done.

resQueueTodo :: InputPoint -> Map (Resource c)

Add input to todo-part of input queue.

resQueueDone :: NIOPoint -> Map (Resource c)

Add input to done-part of input queue.

Hook

type Assert c = Assert' Resource c

type ConsRelmap c = ConsRelmap' Resource c

type Global c = Global' Resource c

type RelkitHook c = RelkitHook' Resource c

type Relmap c = Relmap' Resource c

type RelmapLinkTable c = RelmapLinkTable' Resource c

type Rop c = Rop' Resource c

type RopCons c = RopCons' Resource c

type Intmed c = Intmed' Resource c

type ShortAssert c = ShortAssert' Resource c

type ShortAsserts c = ShortAsserts' Resource c

global :: CContent c => Global c