Koshucode.Baala.Data.Class.Edge

Description

Edge content types.

Basis

class (Eq c, Ord c, MixEncode c) => Basis c where

Basis of contents.

Minimal complete definition

typeOf

Methods

typeOf :: c -> Type

Get type of content.

isA :: Test2 c Type

Test content is a some type.

sameType :: Test2 c c

Test two contents are same type.

type GetContent b c = Ab c -> Ab b

Get content in calculation.

getContent :: Basis c => Test c -> (c -> b) -> GetContent b c

Get content which may be aborted.

Empty

class Basis c => CEmpty c where

Empty: the minimum content.

Methods

isEmpty :: c -> Bool

empty :: c

maybeEmpty :: CEmpty c => (a -> c) -> Maybe a -> c

Create empty or non-empty content.

>>> maybeEmpty pText $ Just "foo" :: Content
ContentText "foo"

>>> maybeEmpty pText $ Nothing :: Content
ContentEmpty

maxContents :: CEmpty c => [c] -> c

Maximum content of contents list. If argument is the empty list, returns the empty content.

empties :: CEmpty c => Int -> [c]

List of empties.

>>> empties 4 :: [Content]
[ContentEmpty, ContentEmpty, ContentEmpty, ContentEmpty]

cutEmpty :: CEmpty c => Map [Term c]

Cut empty terms.

End

class Basis c => CEnd c where

End of everything: the maximum content.

Methods

isEnd :: c -> Bool

end :: c

minContents :: CEnd c => [c] -> c

Minimum content of contents list. If argument is the empty list, returns the end content.