Koshucode.Baala.Data.Decode.Dataset

Description

Dataset as a set of judges.

Dataset is like a bridge of judges and relations. We can get a relation from a dataset, that dataset is build from judges.

Documentation

data Dataset c

Dataset is a set of judges.

datasetClasses :: Dataset c -> Map JudgeClass [TermName]

Retrieve judgement class and its term names.

dataset :: [Judge c] -> Dataset c

Gather judges into a dataset.

datasetAdd :: [Judge c] -> Map (Dataset c)

Add judges to dataset.

readDataset :: CContent c => FilePath -> IOAb (Dataset c)

Read file and convert to dataset.

>>> readDataset "foo.k" :: Koshucode.Baala.Base.IOAb DatasetC
Right Dataset { 3 * A /x /y | 2 * B /y /z | 1 * C /z }