Koshucode.Baala.Data.Decode.Term

Description

Decode term name.

Term name

type CacheT = CacheS TermName

Term name cache.

cacheT :: CacheT

Empty term name cache.

treeFlatName :: TTree -> Ab TermName

Read flat term name from token tree. If the token tree contains nested term name, this function failed.

>>> S.tt1 "/a" >>= treeFlatName
Right (TermName "a")

>>> S.tt1 "+/a" >>= treeFlatName
Right (TermName "a")

>>> S.tt1 "/a/x" >>= treeFlatName
Left ...

treeFlatNameCached :: CacheT -> TTree -> Ab (CacheT, TermName)

Cached version of treeFlatName.

treesFlatNames :: [TTree] -> Ab [TermName]

Read flat term names.

>>> S.tt "/a /b" >>= treesFlatNames
Right ["a","b"]

treeSignedName :: TTree -> Ab TermName

Read signed term name.

>>> S.tt1 "+/a" >>= treeSignedName
Right (GT, "a")

treesTerms :: [TTree] -> Ab [Term [TTree]]

Read list of named token trees from token trees.

>>> S.tt "/a 'A3 /b 10" >>= treesTerms
Right [("a", [TreeL ...]),
       ("b", [TreeL ...])]

treesTermsCached :: CacheT -> [TTree] -> Ab (CacheT, [Term [TTree]])

Cached version of treesTerms.

treesTerms1 :: [TTree] -> Ab [Term TTree]

Read list of named token trees from token trees. This function wraps long branches into group.

treesFlatNamePairs :: [TTree] -> Ab [TermName2]

Decode a list of name-and-name pairs.

>>> S.tt "/a /x /b /y" >>= treesFlatNamePairs
Right [("/a", "/x"), ("/b", "/y")]

treesNamesByColon :: [TTree] -> Ab [[TermName]]

Decode term names grouped by colons.

>>> S.tt "/a /b : /p /q : /x /y" >>= treesNamesByColon
Right [["a","b"], ["p","q"], ["x","y"]]

treesFlatNamesCo :: [TTree] -> Ab (Bool, [TermName])

Decode term names with optional complement symbol.

>>> S.tt "~ /a /b" >>= treesFlatNamesCo
Right (True, ["a","b"])

>>> S.tt "/a /b" >>= treesFlatNamesCo
Right (False, ["a","b"])