Koshucode.Baala.Data.Type.Decimal.Fraction

Description

Decimal functions.

Part of decimals

Integer and fraction

decimalIntPart :: Map Decimal

Integer part of decimal.

decimalFracPart :: Map Decimal

Fractional part of decimal.

decimalIntFrac :: Decimal -> (Decimal, Decimal)

Integer and fractional parts of decimal.

>>> decimalIntFrac $ D.realDecimal 2 (4.5 :: Double)
(Decimal (2) 4, Decimal (2) 0 + 1 % 2)

Numerator and denominator

decimalNum :: Decimal -> DecimalInteger

Numerator part of decimal number.

decimalDenom :: Decimal -> DecimalInteger

Denominator part of decimal number.

Conver to integer

Round

decimalRound :: Map Decimal

Round decimal per self fractional length.

decimalRoundAt :: Bin Decimal

Round decimal per fractional length.

decimalRoundPer :: Bin Decimal

Round decimal per unit decimal.

Round to even

decimalRoundEven :: Map Decimal

Round decimal to even per self fractional length.

decimalRoundEvenAt :: Bin Decimal

Round decimal to even per fractional length.

decimalRoundEvenPer :: Bin Decimal

Round decimal to even per unit decimal.

Truncate

decimalTrunc :: Map Decimal

Truncate decimal per self fractional length.

decimalTruncAt :: Bin Decimal

Truncate decimal per fractional length.

decimalTruncPer :: Bin Decimal

Truncate decimal per unit decimal.

decimalTruncError :: Map Decimal

Truncation error of decimal.

Round out

decimalRoundOut :: Map Decimal

Round out (toward infinity) decimal to even per self fractional length.

decimalRoundOutAt :: Bin Decimal

Round out (toward infinity) decimal to even per fractional length.

decimalRoundOutPer :: Bin Decimal

Round out (toward infinity) decimal to even per unit decimal.

Floor

decimalFloor :: Map Decimal

Floor decimal per self fractional length.

decimalFloorAt :: Bin Decimal

Floor decimal per fractional length.

decimalFloorPer :: Bin Decimal

Floor decimal per unit decimal.

Ceil

decimalCeil :: Map Decimal

Ceiling decimal per self fractional length.

decimalCeilAt :: Bin Decimal

Ceiling decimal per fractional length.

decimalCeilPer :: Bin Decimal

Ceiling decimal per unit decimal.

Other

chopDigitsTrancate :: (Integral d, Integral n) => d -> n -> n

chopDigitsTrancate d n returns a number which does not have the tailing d digits. If d is zero or negative, it returns just n.

chopDigitsRound :: (Integral d, Integral n) => d -> n -> n

chopDigitsRound is similar to chopDigitsTrancate, but rounds chopped digit.

roundLastDigit :: Integral n => n -> n

Round the last (least significant) digit.