Koshucode.Baala.Data.Type.Judge.Judge

Description

Judgements: a symbolic representations of affirmed or denied statements.

Datatype

data Judge c

Judgement on type c.

Judgement (or judge for short) is divided into three parts: logical quality, name of class, and argument. Boolean values True or False of logical quality corresponds to affirmed or denied judge. A name of judgement class represents certain sentence class that gives intepretation of data. Sentence class has placeholders filled by Named c in argument.

Constructors

JudgeAffirm JudgeClass [Term c]
|-- P /x 10 /y 20

JudgeDeny JudgeClass [Term c]
|-x P /x 10 /y 20

JudgeMultiDeny JudgeClass [Term c]
|-xx P /x 10 /y 20

JudgeChange JudgeClass [Term c] [Term c]
|-c P /x 10 +/y 20

JudgeMultiChange JudgeClass [Term c] [Term c]
|-cc P /x 10 +/y 20

JudgeViolate JudgeClass [Term c]
|-v P /x 10 /y 20

judgeTermsMap :: ([Term a] -> [Term b]) -> Judge a -> Judge b

Map function to terms.

judgeAdd :: Term c -> Map (Judge c)

Add a term into judgement.

Logical quality

type JudgeOf c = JudgeClass -> [Term c] -> Judge c

Construct judgement from its class and terms.

affirm :: JudgeOf c

Construct affirmative judgement.

>>> affirm "P" [("x",10), ("y",20)] :: Judge Int
JudgeAffirm "P" [("x",10), ("y",20)]

deny :: JudgeOf c

Construct denial judgement.

affirmJudge :: Map (Judge c)

Affirm judgement, i.e., change logical quality to affirmative.

denyJudge :: Map (Judge c)

Deny judgement, i.e., change logical quality to denial.

>>> denyJudge $ affirm "P" [("x",10), ("y",20)] :: Judge Int
JudgeDeny "P" [("x",10), ("y",20)]

isAffirmative :: Test (Judge c)

Test which judgement is affirmed.

isDenial :: Test (Judge c)

Test which judgement is denied.

isViolative :: Test (Judge c)

Test which judgement is for violation.

data AssertType

Type of assertions.

Constructors

AssertAffirm
|== C : R generates affirmative judges.

AssertDeny
|=x C : R generates denial judges.

AssertMultiDeny
|=xx C : R generates multiple-denial judges.

AssertChange
|=c C : R generates changement judges.

AssertMultiChange
|=cc C : R generates multiple-changement judges.

AssertViolate
|=v C : R generates violation judges.

assertAs :: AssertType -> JudgeOf c

Create judgement corresponding to assertion type.

assertSymbol :: AssertType -> String

Frege's stroke and various assertion lines.

>>> assertSymbol AssertAffirm
"|=="

Encode

type EncodeJudge c = Shorten -> Judge c -> MixText

Encode judgement with short sign converter.

judgeBreak :: LineBreak

Conventional line-break setting for judges: 4-spaces indent and 120-columns line.

judgeMix :: MixShortEncode c => MixText -> EncodeJudge c

Encode judgement with term separator.

judgeMix2 :: MixShortEncode c => EncodeJudge c

Encode judgement with two-spaces term separator.

judgeMixTab :: MixShortEncode c => EncodeJudge c

Encode judgement with tab term separator.

judgeMix2Tab :: MixShortEncode c => Bool -> EncodeJudge c

Conditional judgement encoder. If the first argument is true, judgeMix2 is used, otherwise judgeMixTab.

termNameToMix :: TermName -> MixText

Encode term name.

>>> termNameToMix "foo"
MixText "/foo"

termsToMix1 :: MixShortEncode c => Shorten -> [Term c] -> MixText

Encode term list with one-space separator.

termsToMix2 :: MixShortEncode c => Shorten -> [Term c] -> MixText

Encode term list with two-spaces separator.