Koshucode.Baala.Data.Type.Rel.TermPicker

Description

Term picker is a data for picking target terms based on input heading terms. For example, when picking /a /b from /a /b /c /d /e, use termPicker like:

>>> let pk = termPicker ["a", "b"] ["a", "b", "c", "d", "e"]
>>> pickTerms pk ["1", "2", "3", "4", "5"]
["1", "2"]

Construct

data TermPicker c

Data for picking terms.

Constructors

TermPicker

Fields

ssLShareIndex :: [Int]

Indicies of right-shared part

ssRShareIndex :: [Int]

Indicies of left-shared part

ssDisjoint :: Bool

Whether shared part is empty

ssLSideNames :: [TermName]

Left-side term names

ssLShareNames :: [TermName]

Left-shared term names

ssRShareNames :: [TermName]

Right-shared term names

ssRSideNames :: [TermName]

Right-side term names

ssLSide :: [c] -> [c]

Pick left-side part from left contents

ssLShare :: [c] -> [c]

Pick left-shared part from left contents

ssRShare :: [c] -> [c]

Pick right-shared part from right contents

ssRSide :: [c] -> [c]

Pick right-side part from right contents

ssRSplit :: [c] -> ([c], [c])

Pick right-shared and right-side part

ssRAssoc :: [c] -> ([c], [c])

Pick right-shared part and right contents

ssRForward :: [c] -> [c]

Move shared terms forward.

ssRBackward :: [c] -> [c]

Move shared terms backward.

type TermPick c = TermPicker c -> [c] -> [c]

Type for picking terms.

type TermPick2 a b = (TermPick a, TermPick b)

Double picker.

termPicker :: (GetTermNames l, GetTermNames r) => l -> r -> TermPicker c

Create term picker from left and right term names.

>>> let ss = termPicker ["a", "b", "c"] ["b", "c", "d", "e"]

>>> ssDisjoint ss
False

Left terms.

>>> ssLShareIndex ss
[1, 2]

>>> ssLShareNames ss
["b", "c"]

>>> ssLSide ss "ABC"
"A"

>>> ssLShare ss "ABC"
"BC"

Right terms.

>>> ssRShareIndex ss
[0, 1]

>>> ssRShareNames ss
["b", "c"]

>>> ssRSideNames ss
["d", "e"]

>>> ssRShare ss "BCDE"
"BC"

>>> ssRSide ss "BCDE"
"DE"

>>> ssRSplit ss "BCDE"
("BC", "DE")

>>> ssRAssoc ss "BCDE"
("BC", "BCDE")

>>> ssRForward ss "BCDE"
"BCDE"

>>> ssRBackward ss "BCDE"
"DEBC"

Pre & new terms

preTerms :: TermPicker c -> [TermName]

List of present terms.

>>> preTerms $ termPicker (words "a b c") (words "a b d e")
["a","b"]

newTerms :: TermPicker c -> [TermName]

List of new terms.

>>> newTerms $ termPicker (words "a b c") (words "a b d e")
["c"]

preTermsExist :: Test (TermPicker c)

Test present terms exist.

>>> preTermsExist $ termPicker (words "d e") (words "a b c")
False

>>> preTermsExist $ termPicker (words "c d e") (words "a b c")
True

newTermsExist :: Test (TermPicker c)

Test new terms exist.

>>> newTermsExist $ termPicker (words "a b c") (words "a b d e")
True

>>> newTermsExist $ termPicker (words "a b") (words "a b d e")
False

Mapping

picker :: (GetTermNames t1, GetTermNames t2) => t1 -> t2 -> Map [c]

Pick contents.

>>> picker (words "a c") (words "a b c") (words "1 2 3")
["1", "3"]

pickTermsIndex :: TermPicker c -> [Int]

Extract indices of terms.

>>> pickTermsIndex $ termPicker (words "b d") (words "a b d e")
[1,2]

pickTerms :: TermPicker c -> Map [c]

Pick terms according to term picker.

cutTerms :: TermPicker c -> Map [c]

Cut terms according to term picker.

forwardTerms :: TermPicker c -> Map [c]

Move terms forward.

backwardTerms :: TermPicker c -> Map [c]

Move terms backward.

towardTerms :: Bool -> TermPicker c -> Map [c]

Move terms forward (True) or backward (False).

>>> let pk = termPicker (words "c b") (words "a b c d")
>>> towardTerms True pk (words "A B C D")
["C","B","A","D"]
>>> towardTerms False pk (words "A B C D")
["A","D","C","B"]