Koshucode.Baala.Data.Type.Time.Time

Description

Time

Data type

data Time

Time is a small duration on timeline.

Constructors

TimeYmdcz Date Clock Sec
Date and clock with time zone

TimeYmdc Date Clock
Date and clock

TimeYmd Date
Year, month, and day

TimeYw MJDay
Year and week

TimeYm MJDay
Year and month

timeMjd :: Time -> DayCount

Get integer content of the Modified Julian Day of time.

timePrecision :: Time -> String

Get the name of time precision.

timeYmdTuple :: Time -> YmdTuple

Construction

timeYmd :: MJDay -> Time

Create monthly date from the Modified Julian Day.

timeYmdc :: MJDay -> Clock -> Time

Create monthly time from the Modified Julian Day and clock.

timeFromMjd :: DayCount -> Time

Create time data form modified Julian date.

>>> timeFromMjd 55555
2010-12-25

timeFromYmAb :: Year -> Month -> Ab Time

Create time data from year and month.

timeFromYwAb :: Year -> Week -> Ab Time

Create time data from year and week.

timeFromDczAb :: Date -> Clock -> Maybe Sec -> Ab Time

Create time data from date, clock, and time zone.

timeFromYmd :: Year -> Month -> Day -> Time

timeFromYmdTuple :: YmdTuple -> Time

Construction with I/O

today :: IO Time

Get today.

>>> today
2013-04-18

now :: IO Time

Get current time without time zone.

>>> now
2013-04-18 08:37:36

nowZoned :: IO Time

Get current time with time zone.

>>> nowZoned
2013-04-18 08:37:36 +09:00

Conversion

timeOmitClock :: Map Time

Omit clock part.

timeOmitZone :: Map Time

Omit timezone.

timeMapDate :: Map Date -> Map Time

Map day part of time.

timeMapMjd :: Map DayCount -> Map Time

Map integer content of the Modified Julian Day.

timeFromZonedTime :: ZonedTime -> Time

Convert local time to Koshu time content.