Koshucode.Baala.Data.Type.Type.Genus

Description

Genus of contents.

Constructor

data Genus c

Set of various contents. A genus C - X + Y has (1) C is a set of types as representative contents, (2) X is exclusive contents from the set, and (3) Y is inclusive contents to the set.

Contents in X must be in C, contents in Y must not be in C. Internal function genusFrom trims unnecessary contents according to this conditions.

Diagram of genus G = C - X + Y:

G -------$$$-$$$$$$------$-$$$-$$$$---------$--$----
            :             :   :             :  :
C    T1 |$$$$$$$$$$| T2 |$$$$$$$$$$|        :  :
X           x             x   x             :  :
Y                                           y  y

Constructors

Genus

Fields

genusTypes :: Set c

Representative contents.

genusExclusive :: Set c

Particular exclusive contents.

genusInclusive :: Set c

Particular inclusive contents.

genusVoid :: Genus c

Completely empty genus.

genusType :: (Ord c, CTypeOf c) => c -> Genus c

Create genus of single type.

genusOf :: (Ord c, CTypeOf c) => c -> Genus c

Create genus of single inclusives content.

genusMember :: (Ord c, CTypeOf c) => c -> Genus c -> Bool

Test membership of content to genus.

Operator

genusMeet :: (Ord c, CTypeOf c) => Bin (Genus c)

Meet of two genera.

Meet of A - X1 + Y1

-------$$-$$$$$$$------$-$$$$$-$$--------$---$-----------
         :              :     :          :   :
A  T1 |$$$$$$$$$$| T2 |$$$$$$$$$$|       :   :
X1       x1             x1    x1         :   :
Y1                                       y1  y1

and B - X2 + Y2

-------------$---------$$$$-$$$$$------$$-$$$$$$$---$----
             :             :             :          :
B            :     T2 |$$$$$$$$$$| T3 |$$$$$$$$$$|  :
X2           :             x2            x2         :
Y2           y2                                     y2

is:

-------------$---------$-$$-$$-$$------------$-----------
             :          :  :  :              :
C            :     T2 |$$$$$$$$$$|           :
X3      (x1) :          x1 x2 x1        (x2) :
Y3           y2                              y1
        <A * Y2 - X1>                   <B * Y1 - X2>

genusDiff :: (Ord c, CTypeOf c) => Bin (Genus c)

Difference of two genera.

Difference of A - X1 + Y1

-------$$$-$$$$$$------$-$$$$$-$$--------$--$-----------$----
          :             :     :          :  :           :
A  T1 |$$$$$$$$$$| T2 |$$$$$$$$$$|       :  :           :
X1        x1            x1    x1         :  :           :
Y1                                       y1 y1          y1

and B - X2 + Y2

-------------$---------$$$$-$$$$$------$$-$$$$$$$---$--------
             :             :             :          :
B            :     T2 |$$$$$$$$$$| T3 |$$$$$$$$$$|  :
X2           :             x2            x2         :
Y2           y2                                     y2

is:

-------$$$-$$-$$$----------$-------------$--------------$----
          :  :             :             :              :
C  T1 |$$$$$$$$$$|         :             :              :
X3        x1 y2       (x1) : (x1)        :         (y2) :
Y3                         x2            y1             y1

genusJoin :: (Ord c, CTypeOf c) => Bin (Genus c)

Join of two genera.

Join of A - X1 + Y1

-------$$$-$$$$$$------$-$$$$-$$$--------$---------------
          :             :    :           :
A  T1 |$$$$$$$$$$| T2 |$$$$$$$$$$|       :
X1        x1            x1   x1          :
Y1                                       y1

and B - X2 + Y2

-------------$---------$-$$$$$$$$------$$-$$$$$$$---$----
             :          :                :          :
B            :     T2 |$$$$$$$$$$| T3 |$$$$$$$$$$|  :
X2           :          x2               x2         :
Y2           y2                                     y2

is:

-------$$$-$$$$$$------$-$$$$$$$$------$$$$$$$$$$---$----
          :             :                           :
C  T1 |$$$$$$$$$$| T2 |$$$$$$$$$$| T3 |$$$$$$$$$$|  :
X3        x1            x1 = x2                     :
      <X1 - B - Y2>    <X1 * X2>      <X2 - A - Y1> :
Y3          (y2)                                    y2

genusSplit :: (Ord c, CTypeOf c) => Genus c -> Genus c -> (Genus c, Genus c)

Split acceptable genus by actual genus.

genusInclude :: (Ord c, CTypeOf c) => Genus c -> [c] -> Genus c

Include specific contents.

genusExclude :: (Ord c, CTypeOf c) => Genus c -> [c] -> Genus c

Exclude specific contents.