Koshucode.Baala.Minimal

Description

Minimal relational operators

Relatoinal operators

module Koshucode.Baala.Minimal.Operator

module Koshucode.Baala.Minimal.Origin

module Koshucode.Baala.Minimal.Restrict

module Koshucode.Baala.Minimal.Term

module Koshucode.Baala.Minimal.Tropashko

Naming conventions

Functions for relational mapping are named under some conventions.

xxx

Name of relational operator.

ropConsXxx

Functions of RopCons. These functions construct Relmap from operator usages, but constructions are failed if operators are misused.

relmapXxx

Functions from Relmap to Relmap.

relfyXxx

Functions that make Relfy.

likeXxx

Operand parser.

LikeXxx

Operand pattern.