Koshucode.Baala.Minimal.OpKit

Description

Kit for implementing relational operators

Monoid

mappend :: Monoid a => a -> a -> a

An associative operation

mconcat :: Monoid a => [a] -> a

Fold a list using the monoid. For most types, the default definition for mconcat will be used, but the function is included in the class definition so that an optimized version can be provided for specific types.

Abort

type Abort = AbortType AbortReason

Abort reason and source information.

type AbortOr b = AbortOrType AbortReason b

Either of (1) right result, or (2) abort reason with source.

(<!!>) :: [Named a] -> String -> AbortOr a

Lookup association list. This function may abort on AbortLookup.

bug :: a

Stop on error ''bug in koshucode''

Content

class PrimContent c => CInt c where

Methods

putInt :: Int -> c

class PrimContent c => CNil c where

Types that can be nil

Methods

nil :: c

class PrimContent c => CRel c where

Methods

putRel :: Rel c -> c

class PrimContent c => CText c where

Methods

putText :: String -> c

nil :: CNil c => c

putBool :: CBool c => Bool -> c

Put Boolean value into content c.

putInt :: CInt c => Int -> c

putList :: CList c => [c] -> c

putRel :: CRel c => Rel c -> c

putText :: CText c => String -> c

Data

data Rel v

Relations on type v. Heading of relation and body of relation as a list of tuples.

Constructors

Rel Relhead (Relbody v)

data Relhead

Heading of relation as a list of terms

Constructors

Relhead

Fields

headTerms :: [Relterm]

data Relterm

Term in heading of relation

Constructors

Term String
For non-relation

Nest String [Relterm]
For relation

headChange :: Map [String] -> Map Relhead

Reconstruct head.

headFrom :: [String] -> Relhead

Make head from term names.

>>> headFrom ["/a", "/b"]
Relhead [Term "/a", Term "/b"]

posOf :: Relhead -> [[String]] -> [TermPos]

Positions of given names in a head

posFrom :: Relhead -> Relhead -> [TermPos]

Positions of given (sub)head in a head

posPoss :: [TermPos] -> [Int]

Indicies

csPick :: [TermPos] -> Map [c]

Pick contents by positions.

csCut :: [TermPos] -> Map [c]

Cut contents by positions.

reldee :: Rel v

Relational constant that has no terms and the empty tuple.

reldum :: Rel v

Relational constant that has no terms and no tuples.

termsInner :: [TermPos] -> [[String]]

Pick an inner part.

termsOuter :: [TermPos] -> [[String]]

Pick an outer part.

Prelude

type Map a = a -> a

type Listmap a = forall a. Map [a]

gatherToMap :: Ord k => [(k, v)] -> Map k [v]

Gather (key,value) to a Map key [value]

indexCut :: [Int] -> [a] -> [a]

Cut indexed elements

>>> indexCut [1,3] "abcdefg"
"acefg"

indexPick :: [Int] -> [a] -> [a]

Pick indexed elements.

>>> indexPick [1,3] "abcdefg"
"bd"

lookupMap :: Ord k => k -> Map k a -> Maybe a

unique :: Ord a => [a] -> [a]

Remove duplicate elements.

Relmap

type Relop v = OpUse v -> AbortOr (Relmap v)

Constructor of relational operator Relmap. Relmap is constructed from HalfRelmap and subrelmaps in it.

data OpImplement v

Implementation of relmap operator. It consists of (1) operator name, (2) operand parser, (3) constructor of operator, and (4) usage of operator.

Constructors

OpImplement String OpParser (Relop v) [String]

type OpParser

Arguments

= [TokenTree]
Unparsed operand

-> [Named [TokenTree]]
Parsed operand

Parser for operand of relational operator. This parsers docompose operand trees, and give a name to suboperand.

type OpParser'

Arguments

= [Named [TokenTree]]
Partially parsed operand

-> [Named [TokenTree]]
Full parsed operand

data OpUse v

Use of operator

Constructors

OpUse

Fields

opHalf :: HalfRelmap

Syntactic data of operator use

opSubmap :: [Relmap v]

Subrelmaps

data Relmap v

Relation-to-relation mapping. A Relmap is correspond to a use of relational operator.

Constructors

RelmapSource HalfRelmap String [String]
Retrieve a relation from a dataset

RelmapConst HalfRelmap String (Rel v)
Constant relation

RelmapAlias HalfRelmap (Relmap v)
Equavalent relmap

RelmapCalc HalfRelmap String (RelmapSub v) [Relmap v]
Relmap that maps relations to a relation

RelmapAppend (Relmap v) (Relmap v)
Connect two relmaps

RelmapName HalfRelmap String
Relmap reference

data RelmapCons v

Half and full relmap constructor

Constructors

RelmapCons

Fields

consHalf :: RelmapHalfCons

consFull :: RelmapFullCons v

type RelmapFullCons v

Arguments

= HalfRelmap
Half relmap from RelmapHalfCons

-> AbortOr (Relmap v)
Result relmap

Second step of constructing relmap, make Relmap from contents of HalfRelmap.

type RelmapHalfCons

Arguments

= [SourceLine]
Source information

-> [TokenTree]
Operand as source trees

-> HalfRelmap
Result half relmap

First step of constructing relmap, make HalfRelmap from use of relational operator.

halfOperand :: HalfRelmap -> [Named [TokenTree]]

Operand of relmap operation

relmapAlias

Arguments

:: OpUse v
Use of operator

-> Relmap v

-> Relmap v
Result relmap

Alias relmap.

relmapCalc

Arguments

:: OpUse v
Use of operator

-> String
Operator name

-> RelmapSub v
Calculation of operation

-> Relmap v
Result relmap

Make a non-confluent relmap.

relmapConfl

Arguments

:: OpUse v
Use of operator

-> String
Operator name

-> RelmapSub v
Calculation of operation

-> [Relmap v]
Subrelmaps

-> Relmap v
Result relmap

Make a confluent relmap.

relmapCons

Arguments

:: [OpImplement v]
Implementations of relational operators

-> RelmapCons v
Relmap constructors

Make half and full relmap constructors.

relmapConst

Arguments

:: OpUse v
Use of operator

-> String
Operator name

-> Rel v
Constant relation

-> Relmap v
Result relmap

Constant relmap.

relmapSource

Arguments

:: OpUse v
Use of operator

-> String
Operator name

-> [String]
List of term names

-> Relmap v
Result relmap

Retrieve relation from dataset.

Section

data QuasiQuoter

koshuQuoter

Arguments

:: RelmapHalfCons
Relmap half constructor

-> ExpQ
Quotation expression of RelmapFullCons

-> QuasiQuoter
Quoter that outputs Section or Relmap

Make quasiquoter for [koshu| ... |].

operandGroup :: [TokenTree] -> [Named [TokenTree]]

Split operand into named group. Non quoted words beginning with hyphen, e.g., -x, are name of group.

>>> operandGroup $ tokenTrees $ tokens "a b -x c 'd' -y e"
[("",   [TreeL (Word 0 "a"), TreeL (Word 0 "b")]),
 ("-x", [TreeL (Word 0 "c"), TreeL (Word 1 "d")]),
 ("-y", [TreeL (Word 0 "e")])]

Syntax

data Token

Constructors

TWord TNumber Int String
Word. Int represents quotation level, e.g., 0 for non-quoted, 1 for single-quoted, 2 for double-quoted.

TTerm TNumber [String]
Term name

TOpen TNumber String
Open paren

TClose TNumber String
Close paren

TSpace TNumber Int
N space characters

TComment TNumber String
Comment text

type TokenTree = Tree Token

Tree of tokens.

data Tree a

Tree of leaf and branch.

Constructors

TreeL a
Leaf. Terminal of tree.

TreeB Int [Tree a]
Branch. Paren-type and subtrees.

binaryTree

Arguments

:: Show a

=> (a -> BinaryHeight)
Height function

-> Tree a
Infixed tree

-> Tree a
Prefixed tree

Split branches in a given tree at infixed binary operators

heightTable

Arguments

:: Ord a

=> [(BinaryHeight, [a])]
Height table

-> a
Height function

-> BinaryHeight

Make the height function from a height table of operators.

heightTableUnbox

Arguments

:: Ord a

=> (b -> a)
Unbox function

-> [(BinaryHeight, [a])]
Height table

-> b
Height function

-> BinaryHeight

OpKit

module Koshucode.Baala.Minimal.OpKit.Get

module Koshucode.Baala.Minimal.OpKit.Pattern

module Koshucode.Baala.Minimal.OpKit.Term