Koshucode.Baala.Minimal.Relmap.Tropashko

Description

Fundamental operators in relational algebra.

Tropashko's relational lattice is a kind of relational algebra. Relational algebra is an algebraic formulation for relational model. In constrast to Codd's original relational algebra, Tropashko lattice is in more conventional and strict ways. The lattice has fundamental operators from which other operators are derived.

Fundamental operators

In lattice theory, there are two operators meet and join.

R1 ^ R2

Meet of R1 and R2. This operator is one of generalized intersection. ''Meet'' is same as ''natural join'' in SQL. R1 tuple is meetable to R2 tuple if and only if shared terms of R1 tuple and R2 tuple are same contents.

R1 v R2

Join of R1 and R2. This operator is one of generalized union. Toropashko calls it ''inner union''. Join of R1 and R2 is ordinary union of shared-term-projection tuples in R1 and R2.

Meet (Natural join)

relopMeet :: Ord v => Relop v

relmapMeet

Arguments

:: Ord v

=> OpUse v
Source infomation

-> Relmap v
Subrelmap of meet operator

-> Relmap v
Relmap of meet operator

Meet two relations.

relMeet

Arguments

:: Ord c

=> Rel c
Input relation R1

-> Rel c
Input relation R2

-> AbOr (Rel c)
Meet of R1 and R2

Meet two relations.

Summary of calculating meet of relations like (/a /b /c) meet (/b /c /d) = (/d ++ /a /b /c). In this case, shared terms are /b and /c, left-sided term is /a, and right-sided term is /d.

Input

Relations (Rel h1 b1) and (Rel h2 b2)

Output

Relation (Rel h3 b3)

Let s1 be shared-term information of h1.

Let s2 be shared-term information of h2.

Let cs1 :: [c] be element of b1.

Let cs2 :: [c] be element of b2.

Split each cs2, by s2, into shared and sided terms :: ([c], [c]).

For each same shared terms, collect the sided terms. Let it be (share2, sides2) :: ([c], [[c]]).

Pick shared terms share1 :: [c], by s1, from each cs1.

Concat cs1 and each of sides2 where share1 equals share2,

Join (Inner union)

relopJoin :: Ord v => Relop v

relmapJoin

Arguments

:: Ord v

=> OpUse v
Source infomation

-> Relmap v
Subrelmap of join operator

-> Relmap v
Relmap of join operator

Join two relations.

relJoin

Arguments

:: Ord v

=> Rel v
Input relation R1

-> Rel v
Input relation R2

-> AbOr (Rel v)
Join of R1 and R2

Join two relations.