Koshucode.Baala.Op.Cox.GeoDatumJp

Description

Conversion geographic point from the plane rectangular coordinates /n /x /y to the geographic coordinates /latitude /longtitude. This program is based on the method described in Kazushige KAWASE, 2011, A More Concise Method of Calculation for the Coordinate Conversion between Geographic and Plane Rectangular Coordinates on the Gauss-Krüger Projection, Bulletin of the Geospatial Information Authority of Japan, 121.

Documentation

type GeoPoint = (Double, Double)

convDegree :: Int -> GeoPoint -> GeoPoint

convRadian :: Int -> GeoPoint -> GeoPoint