Koshucode.Baala.Overture.Shorthand

Description

Shorthand types and functions.

Derived types

type Pair a b = (a, b)

Pair of a and b.

type Twin a = (a, a)

Pair of a and a.

Infix operators

(&) :: a -> b -> (a, b) infixr 0

Pairing operator.

>>> "a" & "b"
("a", "b")

>>> "a" & "b" & "c"
("a", ("b", "c"))

(++) :: Monoid a => a -> a -> a infixr 6

Associative operator of monoid, same as mappend or infix <>.

>>> Just "abc" ++ Nothing ++ Just "def"
Just "abcdef"

(<++>) :: (a -> [b]) -> [a] -> [b]

Same as concatMap.

>>> words <++> ["foo bar", "baz quux"]
["foo","bar","baz","quux"]

(<$$>) :: (Functor f, Functor g) => (a -> b) -> f (g a) -> f (g b) infixl 4

Double fmap.

>>> length <$$> [("a", "apple"), ("b", "banana"), ("c", "cocoa")]
[("a",5), ("b",6), ("c",5)]

(<#>) :: (Monad m, Traversable t) => (a -> m b) -> t a -> m (t b)

Monadic mapping, same as mapM.

(<#!>) :: (Monad m, Traversable t) => (a -> m b) -> t a -> m ()

Execution by monadic mapping, same as mapM_.

Functions

ordEq :: Ord a => a -> a -> Bool

Test equality using Ord method. This function can be used for implementing Eq instance.

instance Eq X where
  (==) = ordEq

compareOn :: Ord b => (a -> b) -> a -> a -> Ordering

compareOn f x y comapres f x and f y instead of x and y.

>>> compareOn length "foo" "bar"
EQ

int :: Map Int

Int shorthand.

>>> int 12
12

This is same as:

>>> 12 :: Int
12

integer :: Map Integer

Integer shorthand.

>>> integer 12
12

This is same as:

>>> 12 :: Integer
12

nothing :: a -> Maybe b

Function which always returns Nothing.

>>> nothing True :: Maybe String
Nothing