Koshucode.Baala.Overture.Text.Textual

Description

Textual class.

Documentation

class (Show t, Eq t, Ord t, Monoid t, IsString t) => Textual t where

Text-like value.

Minimal complete definition

tAdd, tCut, tTake, tDrop, tString, stringT

Methods

tEmpty :: t

Empty text.

>>> tEmpty :: String
""

tIsEmpty :: t -> Bool

Test text is empty.

>>> tIsEmpty ""
True

>>> tIsEmpty "foo"
False

tJoin :: t -> t -> t

Append two texts.

>>> tJoin "foo" "bar"
"foobar"

tJoinAll :: [t] -> t

Append multiple texts.

>>> tJoinAll ["foo", "bar", "baz"]
"foobarbaz"

tJoinWith :: t -> [t] -> t

Append multiple texts with separator.

>>> tJoinWith " | " ["foo", "bar", "baz"]
"foo | bar | baz"

tUnwords :: [t] -> t

Append multiple texts with space.

>>> tUnwords ["foo", "bar", "baz"]
"foo bar baz"

tAdd :: Char -> t -> t

Preppend character to text.

>>> tAdd 'g' "foo"
"gfoo"

tAdd2 :: Char -> Char -> t -> t

Preppend two characters to text.

>>> tAdd2 'b' 'a' "foo"
"bafoo"

tAdd3 :: Char -> Char -> Char -> t -> t

Preppend two characters to text.

>>> tAdd3 'b' 'a' 'r' "foo"
"barfoo"

tCut :: t -> Maybe (Char, t)

Split text into first character and rest of text.

>>> tCut "foo"
Just ('f',"oo")

>>> tCut ""
Nothing

tCut2 :: t -> Maybe (Char, Maybe (Char, t))

Split one or two characters.

>>> tCut2 "bar"
Just ('b', Just ('a', "r"))

tCut3 :: t -> Maybe (Char, Maybe (Char, Maybe (Char, t)))

Split one, two, or three characters.

>>> tCut3 "bar"
Just ('b', Just ('a', Just ('r', "")))

>>> tCut3 "b"
Just ('b', Nothing)

tTake :: Int -> t -> t

Take N-#1 characters from text-#2.

>>> tTake 3 "foobar"
"foo"

tDrop :: Int -> t -> t

Drop N-#1 characters from text-#2.

>>> tDrop 3 "foobar"
"bar"

tDropPrefix :: Textual a => a -> t -> Maybe t

Drop prefix-#1 from text-#2.

>>> tDropPrefix "foo" "foobar"
Just "bar"

>>> tDropPrefix "bar" "foobar"
Nothing

tMap :: (Char -> Char) -> t -> t

Map function-#1 to characters of text-value-#2.

tList :: (Char -> a) -> t -> [a]

Map function-#1 to characters of text-value-#2.

>>> tList fromEnum "bar"
[98,97,114]

tAll :: (Char -> Bool) -> t -> Bool

Test all characters of text value.

>>> tAll (`elem` "0123456789.") "120.5"
True

>>> tAll (`elem` "0123456789.") "120,000"
False

tLength :: Integral n => t -> n

Length of textual-value-#1.

>>> tLength "foo" :: Int
3

tWhile :: (Char -> Bool) -> t -> (t, t)

Take subtext from textual-value-#2 while tester-#1 holds, and pairing the subtext with after.

>>> tWhile (/= ' ') "foo bar baz"
("foo", " bar baz")

tWhileNot :: (Char -> Bool) -> t -> (t, t)

Take subtext from textual-value-#2 while tester-#1 do not holds.

tWhileNot f t = tWhile (not . f) t

tDivide :: (Char -> Bool) -> t -> [t]

Divide textual-value-#2 to textual list by delimiter-tester-#1.

>>> tDivide (== '/') "foo/bar/baz"
["foo","bar","baz"]

tWords :: t -> [t]

Divide text into words.

>>> tWords "a bb  ccc "
["a", "bb", "ccc"]

tWordsBy :: (Char -> Bool) -> t -> [t]

Divide text into words by character tester.

>>> tWordsBy (== '|') "a|bb||ccc|"
["a", "bb", "ccc"]

charT :: Char -> t

Create text from single character.

>>> charT 'b' :: String
"b"

char2T :: Char -> Char -> t

Create text from two characters.

>>> char2T 'b' 'a' :: String
"ba"

char3T :: Char -> Char -> Char -> t

Create text from three characters.

>>> char3T 'b' 'a' 'r' :: String
"bar"

charsT :: Int -> Char -> t

N-length textual value.

>>> charsT 4 '+' :: String
"++++"

tString :: t -> String

Convert textual value to string.

>>> tString "foo"
"foo"

stringT :: String -> t

Convert string to textual value.

>>> stringT "foo" :: String
"foo"

showT :: Show a => a -> t

Convert show instance to textual value.

>>> showT (120 :: Int) :: String
"120"

bsT :: Bs -> t

Convert strict bytestring to textual value.

bzT :: Bz -> t

Convert lazy bytestring to textual value.

stringify :: (Functor f, Textual t) => f t -> f String

Convert textual something to string something.