Koshucode.Baala.Overture.Text.Utility

Description

Text utility.

Reexport from Data.Char

isSpace :: Char -> Bool

Returns True for any Unicode space character, and the control characters \t, \n, \r, \f, \v.

Trim

trimBegin :: Textual t => t -> t

Remove space and space-like characters from the beginning of string.

>>> trimBegin "  abc  "
"abc  "

trimEnd :: Textual t => t -> t

Remove space and space-like characters from the end of string.

>>> trimEnd "  abc  "
"  abc"

trimBoth :: Textual t => t -> t

Remove space and space-like characters from the beginning and end of string.

>>> trimBoth "  abc  "
"abc"

Padding

padBegin :: Int -> StringMap

Add spaces to the left.

>>> padBegin 10 "abc"
"       abc"

padEnd :: Int -> StringMap

Add spaces to the right.

>>> padEnd 10 "abc"
"abc       "

padBeginWith :: Char -> Int -> StringMap

Add given character to the left.

>>> padBeginWith '.' 10 "abc"
".......abc"

padEndWith :: Char -> Int -> StringMap

Add given character to the right.

stringWidth :: String -> Int

Calculate width of string.

addSpace :: Textual t => t -> t

Add space character if first character is non-space.

>>> addSpace "aaa"
" aaa"

>>> addSpace " bbb"
" bbb"

>>> addSpace ""
""

Code point

isAsciiCode :: Integral n => n -> Bool

Test ASCII code point (0 .. 127).

isControlCode :: Integral n => n -> Bool

Test ASCII control code point (0 .. 31, 127).

isFormatCode :: Integral n => n -> Bool

Test format code point (9 .. 13).

isNewlineCode :: Integral n => n -> Bool

Test newline code point (10, 13).

integralHT :: Integral n => n

Generic code point (9) of horizontal tab character ('\t').

integralLF :: Integral n => n

Generic code point (10) of line feed character ('\n').

integralVT :: Integral n => n

Generic code point (11) of vertical tab character ('\v').

integralFF :: Integral n => n

Generic code point (12) of form feed character ('\f').

integralCR :: Integral n => n

Generic code point (13) of carriage return character ('\r').

integralSpace :: Integral n => n

Generic code point (32) of space character (' ').

Put

putLn :: IO ()

Print newline.

hPutLn :: Handle -> IO ()

Print newline.

putShow :: Show a => a -> IO ()

Print showing value.

putShowLn :: Show a => a -> IO ()

Print showing value with newline.

putLines :: [String] -> IO ()

Print multiple lines.

hPutLines :: Handle -> [String] -> IO ()

Print multiple lines.

prompt :: IO String

Print prompt (>>) and read user input. It returns trimmed input string. Reread when input is empty.

promptWith :: String -> IO String

Print prompt with give string and read user input.