Koshucode.Baala.Rop.Cox.Calc

Description

Relmap operator with content calculation.

Documentation

ropsCoxCalc :: CContent c => [Rop c]

Implementation of relational operators.

add

consAdd :: CContent c => RopCons c

add /N E ... -let { D | ... }

Calculate expression E of term /N, and add terms /N ... to input relation.

relmapAdd :: CContent c => Intmed c -> (CopSet c, [TermCox c]) -> Relmap c

Create add relmap.

alt

consAlt :: CContent c => RopCons c

alt /P E ... -let { D | ... }

Calculate expression E of term /P, and alter the contents of present terms /P ....

relmapAlt :: CContent c => Intmed c -> (CopSet c, [TermCox c]) -> Relmap c

Create alt relmap.

fill

consFill :: CContent c => RopCons c

fill /P ... -with E

Fill contents of present terms /P ... with the result of E.

relmapFill :: CContent c => Intmed c -> ([TermName], CopSet c, Cox c) -> Relmap c

Create fill relmap.

map

consReplace :: CContent c => RopCons c

map /P ... -by F

replace-all

consReplaceAll :: CContent c => RopCons c

replace-all -from E -to E

relmapReplaceAll :: CContent c => Intmed c -> (CopSet c, Cox c, Cox c) -> Relmap c

Create replace-all relmap.

split

consSplit :: CContent c => RopCons c

split /N E ... -let { D | ... }

Split input relation by the boolean functions E ..., and add terms /N ... which has the splitted relations.

RELMAP
  split /a ( /x = 1 )
        /b ( /y = 10 )

INPUT              OUTPUT
   /x    /y          /a          /b
   ----- -----       ----------- -----------
   1     10          /x    /y    /x    /y
   2     20          ----- ----- ----- -----
   3     10          1     10    3     10
   1     20          1     20
   2     30

relmapSplit :: CContent c => Intmed c -> (CopSet c, [TermCox c]) -> Relmap c

Create split relmap.

relkitSplit :: forall c. CContent c => (CopSet c, [TermCox c]) -> RelkitFlow c

Create split relkit.

unary

consUnary :: CContent c => RopCons c

unary /N E : ...

relmapUnary :: CContent c => Intmed c -> (TermName, [c]) -> Relmap c

Create unary relmap.

dump-cox

consDumpCox :: CContent c => RopCons c

dump-cox E