Koshucode.Baala.Rop.Cox.Empty

Description

Relational compositions whose contents are maybe the empty.

Documentation

ropsCoxEmpty :: CContent c => [Rop c]

Relmap operators that handles empties.

both

consBoth :: CContent c => RopCons c

both R -share /P ... -fill E

relmapBoth :: (Ord c, CRel c) => Intmed c -> SharedTerms -> c -> Map (Relmap c)

Create both relmap.

maybe

consMaybe :: CContent c => RopCons c

maybe R -share /P ... -fill E

Meet input and given relation R. It keeps input tuples of which counterparts are totally negated.

maybe R -only E -order /P ... -top N -mid N -bot N

Maybe with restriction by E and limit to N tuples.

maybe R -only E -order /P ... -part N -of N

Maybe with restriction by E and take N'th part.

relmapMaybe :: (Ord c, CRel c) => Intmed c -> SharedTerms -> c -> Map (Relmap c)

Create maybe relmap.

relkitMaybe :: forall c. (Ord c, CRel c) => SharedTerms -> c -> RelkitBinary c

Create maybe relkit.

maybe with restriction

type MeetOnly c = (SharedTerms, CopSet c, Cox c, RelPortion)

Parameter for meet -only.

relmapMaybeOnly :: CContent c => Intmed c -> (MeetOnly c, c) -> Map (Relmap c)

Create maybe-with-restriction relmap.

compose-maybe

consComposeMaybe :: CContent c => RopCons c

compose-maybe R -share /P ... -fill E

Construct relmap for relational composition.

relmapComposeMaybe :: (Ord c, CRel c) => Intmed c -> SharedTerms -> c -> Map (Relmap c)

Create compose-maybe relmap.