Koshucode.Baala.Rop.Cox.Gadget

Description

Gadgets with content calculation.

Documentation

ropsCoxGadget :: CContent c => [Rop c]

Implementation of relational operators.

const

consConst :: CContent c => RopCons c

const E

Output the constant relation E. Especially, const {= [] =} is equivalent to dee, const {= =} is equivalent to dum.

relmapConst :: Intmed c -> Rel c -> Relmap c

Create const relmap.

relkitConst :: Rel c -> RelkitFlow c

Create const relkit.

number

consNumber :: (Ord c, CContent c) => RopCons c

number /N -from I -order /P ...

relmapNumber :: (CDec c, Ord c) => Intmed c -> (TermName, [TermName], Int) -> Relmap c

Create number relmap.

relkitNumber :: (Ord c, CDec c) => (TermName, [TermName], Int) -> RelkitFlow c

Create number relkit.

rank

consRank :: (Ord c, CContent c) => RopCons c

rank /N -from I -order /P ...

Calculate standard competition ranking (like 1224).

rank /N -dense -from I -order /P...

Calculate dense ranking (like 1223).

relmapGapRank :: (CDec c, Ord c) => Intmed c -> (TermName, [TermName], Int) -> Relmap c

Create rank relmap.

relkitGapRank :: (Ord c, CDec c) => (TermName, [TermName], Int) -> RelkitFlow c

Create rank relkit.

relmapDenseRank :: (CDec c, Ord c) => Intmed c -> (TermName, [TermName], Int) -> Relmap c

Create rank -dense relmap.

relkitDenseRank :: (Ord c, CDec c) => (TermName, [TermName], Int) -> RelkitFlow c

Create rank -dense relkit.

repeat

consRepeat :: (Ord c, CContent c) => RopCons c

repeat I R

array

consArray :: CContent c => RopCons c

array /P /P ... -to /N E ...

Convert discrete tuples of /P /P ... to term array /N .... The first /P of /P /P is a key term of which content is equal to E, and the second is a value of the key. When E is omitted, term name content is used, e.g., -to /1 /2 means -to /1 '/1 /2 '/2. The length of /P /P ... must be even. Relmap like array /k /v /ka /va -to /1 /1a /2 /2a ... works for multiple key-value pairs.

RELMAP
  array /key /val  -to /1 1  /2 2  /3 3  /4 4
                          :     :     :     :
                          :.....:.....:.....:.... /key

INPUT                    OUTPUT
  /group  /key  /val       /group  /1    /2    /3    /4
  ======= ===== -----      ======= ----- ----- ----- -----
  'a      1     10         'a      10    20    30    ()
  'a      2     20         'b      40    50    ()    ()
  'a      3     30                 .                  .
  'b      1     40                 :..................:
  'b      2     50                 term array

relmapArray :: CContent c => Intmed c -> ([TermName2], Int, [[Term c]]) -> Relmap c

Create array relmap.

relkitArray :: forall c. CContent c => ([TermName2], Int, [[Term c]]) -> RelkitFlow c

Create array relkit.

unarray

consUnarray :: CContent c => RopCons c

unarray /N /N ... -from /P E ...

Convert term array /P ... to discrete tuples of /N /N ....

RELMAP
  unarray /key /val  -from /1 1  /2 2  /3 3  /4 4
                              :     :     :     :
                              :.....:.....:.....:.... /key

INPUT                                 OUTPUT
  /group  /1    /2    /3    /4          /group  /key  /val
  ======= ----- ----- ----- -----       ======= ===== -----
  'a      10    20    30    ()          'a      1     10
  'b      40    50    ()    ()          'a      2     20
          .                  .          'a      3     30
          :..................:          'a      4     ()
          term array                    'b      1     40
                                        'b      2     50
                                        'b      3     ()
                                        'b      4     ()

relmapUnarray :: CContent c => Intmed c -> ([TermName2], Int, [[Term c]]) -> Relmap c

Create unarray relmap.

relkitUnarray :: forall c. CContent c => ([TermName2], Int, [[Term c]]) -> RelkitFlow c

Create unarray relkit.