Koshucode.Baala.Rop.Cox.GeoDatumJp

Description

Conversion geographic point from the plane rectangular coordinates /n /x /y to the geographic coordinates /latitude /longtitude. This program is based on the method described in Kazushige KAWASE, 2011, A More Concise Method of Calculation for the Coordinate Conversion between Geographic and Plane Rectangular Coordinates on the Gauss-Krüger Projection, Bulletin of the Geospatial Information Authority of Japan, 121.

Documentation

type GeoPoint = (Double, Double)

Point in geographic coordinate system.

convDegree :: Int -> GeoPoint -> GeoPoint

Convert point in plane rectangular coordinates to degrees of latitude and longitude.

>>> convDegree 8 (0.0, 0.0)
(35.99999999999999, 138.5)

convRadian :: Int -> GeoPoint -> GeoPoint

Convert point in plane rectangular coordinates to geographic latitude and longitude.

>>> convRadian 8 (0.0, 0.0)
(0.6283185307179585, 498600.0)