Koshucode.Baala.Rop.Cox.Type.Clock

Description

Relmap operators for clock type.

Documentation

ropsTypeClock :: CContent c => [Rop c]

Implementation of relational operators.

add-clock

consAddClock :: CContent c => RopCons c

add-clock /c -day D

Add term /c as a clock of D days, i.e., |D'|.

add-clock /c -hour H

Add term /c as a clock of H hours, i.e., |H|.

add-clock /c -min M

Add term /c as a clock of M minutes, i.e., |00:M|.

add-clock /c -sec S

Add term /c as a clock of S seconds, i.e., |00:00:S|.

relmapAddClock :: CContent c => Intmed c -> (CopSet c, TermName, (Cox c, Cox c, (MaybeCox c, MaybeCox c, MaybeCox c))) -> Relmap c

Create clock relmap.

relkitAddClock :: CContent c => (CopSet c, TermName, (Cox c, Cox c, (MaybeCox c, MaybeCox c, MaybeCox c))) -> RelkitFlow c

Create clock relkit.

of-clock

consOfClock :: CContent c => RopCons c

of-clock E -day /N

Add term /N as decimal days of clock.

of-clock E -hour /N

Add term /N as decimal hours of clock.

of-clock E -min /N

Add term /N as decimal minutes of clock.

of-clock E -sec /N

Add term /N as decimal seconds of clock.

of-clock E -sign /N

Add term /N as decimal sign of clock, i.e., -1, 0, or 1.

relmapOfClock :: CContent c => Intmed c -> (CopSet c, Cox c, [Maybe TermName]) -> Relmap c

Create of-clockt relmap.

relkitOfClock :: CContent c => (CopSet c, Cox c, [Maybe TermName]) -> RelkitFlow c

Create of-clock relkit.

alt-clock

consAltClock :: CContent c => RopCons c

alt-clock /P ... -day D

Alter days of clock to D in term /P ....

alt-clock /P ... -hour H

Alter hours of clock to H in term /P ....

alt-clock /P ... -min M

Alter minutes of clock to M in term /P ....

alt-clock /P ... -sec S

Alter seconds of clock to S in term /P ....

relmapAltClock :: CContent c => Intmed c -> (CopSet c, [TermName], (MaybeCox c, MaybeCox c, MaybeCox c, MaybeCox c)) -> Relmap c

Create alt-clock relmap.

relkitAltClock :: CContent c => (CopSet c, [TermName], (MaybeCox c, MaybeCox c, MaybeCox c, MaybeCox c)) -> RelkitFlow c

Create alt-clock relkit.