Koshucode.Baala.Rop.Flat.Lattice.Tropashko

Description

Fundamental operators in relational algebra.

Tropashko's relational lattice is a kind of relational algebra. Relational algebra is an algebraic formulation for relational model. In constrast to Codd's original relational algebra, Tropashko lattice is in more conventional and strict ways. The lattice has fundamental operators from which other operators are derived.

Fundamental operators

In lattice theory, there are two operators meet and join.

R1 ^ R2

Meet of R1 and R2. This operator is one of generalized intersection, as it were, a more-speaking-and-less-mentioning operation. 'Meet' is same as ''natural join'' in SQL. R1 tuple is meetable to R2 tuple if and only if shared terms of R1 tuple and R2 tuple are same contents.

R1 v R2

Join of R1 and R2. This operator is one of generalized union, as it were, a less-speaking-and-more-mentioning operation. Toropashko calls it ''inner union''. Join of R1 and R2 is ordinary union of shared-term-projection tuples in R1 and R2.

meet (natural join)

consMeet :: Ord c => RopCons c

Construct relmap to meet relation.

relmapMeet

Arguments

:: Ord c

=> Intmed c
Source infomation

-> SharedTerms
Shared terms

-> Relmap c
Subrelmap of meet operator

-> Relmap c
Relmap of meet operator

Meet two relations.

relkitMeet :: forall c. Ord c => SharedTerms -> RelkitBinary c

Meet two relations.

Summary of calculating meet of relations like (/a /b /c) meet (/b /c /d) = (/d ++ /a /b /c). In this case, shared terms are /b and /c, left-sided term is /a, and right-sided term is /d.

Input

Relations (Rel he1 bo1) and (Rel he2 bo2)

Output

Relation (Rel he3 b3)

Let s1 be shared-term information of he1.

Let s2 be shared-term information of he2.

Let cs1 :: [c] be element of bo1.

Let cs2 :: [c] be element of bo2.

Split each cs2, by s2, into shared and sided terms :: ([c], [c]).

For each same shared terms, collect the sided terms. Let it be (share2, sides2) :: ([c], [[c]]).

Pick shared terms share1 :: [c], by s1, from each cs1.

Concat cs1 and each of sides2 where share1 equals share2,

join (inner union)

consJoin :: Ord c => RopCons c

Construct relmap to join relation.

relmapJoin

Arguments

:: Ord c

=> Intmed c
Source infomation

-> SharedTerms
Shared terms

-> Relmap c
Subrelmap of join operator

-> Relmap c
Relmap of join operator

Join two relations.

relmapJoinList :: Ord c => Intmed c -> [Relmap c] -> Relmap c

Join multiple relations.

relkitJoin :: SharedTerms -> RelkitBinary c

Join two relations.

Check shared terms

type SharedTerms = Maybe [TermName]

Shared terms for composing check.

unmatchShare :: SharedTerms -> ShareSide c -> Maybe ([TermName], [TermName])

Calculate unmatch shared terms.