Koshucode.Baala.Rop.Flat.Peripheral

Description

Type-specific operators.

Documentation

ropsPeripheral :: CContent c => [Rop c]

Implementation of relational operators.

RDF

consRdf :: RopCons c

rdf P /S /O

tie

consTie :: CTie c => RopCons c

tie /P ... -to /N

relmapTie :: CTie c => Intmed c -> ([TermName], TermName) -> Relmap c

Create tie relmap.

relkitTie :: CTie c => ([TermName], TermName) -> RelkitFlow c

Create tie relkit.

untie

consUntie :: CTie c => RopCons c

untie /P -to /N ...

relmapUntie :: CTie c => Intmed c -> (TermName, [TermName]) -> Relmap c

Create untie relmap.

relkitUntie :: CTie c => (TermName, [TermName]) -> RelkitFlow c

Create untie relkit.

term-name

consTermName :: CTerm c => RopCons c

term-name /N

relmapTermName :: CTerm c => Intmed c -> TermName -> Relmap c

Create term-name relmap.

relkitTermName :: CTerm c => TermName -> RelkitFlow c

Create term-name relkit.

time

consToday :: CTime c => RopCons c

today /N

Get today's time at term /N.

consNow :: CTime c => RopCons c

now /N

Get current local time without time zone at term /N.

now /N -zoned

Get current local time with time zone at term /N.

now /N -utc

Get current UTC.

add term

consAdd1 :: Term c -> RopCons c

Add single constant term to each tuples.

relmapAdd1 :: Term c -> Intmed c -> Relmap c

Create term-adding relmap.

relkitAdd1 :: Term c -> Maybe Head -> Ab (Relkit c)

Create term-adding relkit.