Koshucode.Baala.Rop.Flat.Term

Description

Relmap operators for manipulating term names.

Documentation

ropsTerm :: Ord c => [Rop c]

Relmap operators for manipulating term names.

pick

consPick :: RopCons c

pick /P ...

Project relation to specified terms.

relmapPick :: Intmed c -> [TermName] -> Relmap c

Create pick relmap.

relkitPick :: [TermName] -> RelkitFlow c

Create pick relkit.

cut

consCut :: RopCons c

cut /P ...

Project relation to unspecified terms. This operator is an inversion of pick.

relmapCut :: Intmed c -> [TermName] -> Relmap c

Create cut relmap.

relkitCut :: [TermName] -> RelkitFlow c

Create cut relkit.

pick-term

consPickTerm :: RopCons c

pick-term R

Construct pick-term relmap operator. This operator is similar to pick instead of given terms but relation terms.

relmapPickTerm :: Intmed c -> Relmap c -> Relmap c

Create pick-term relmap.

relkitPickTerm :: RelkitBinary c

Create pick-term relkit.

cut-term

consCutTerm :: RopCons c

cut-term R

Construct cut-term relmap operator. This operator is similar to cut instead of given terms but relation terms.

relmapCutTerm :: Intmed c -> Relmap c -> Relmap c

Create cut-term relmap.

relkitCutTerm :: RelkitBinary c

Create cut-term relkit.

move

consMove :: RopCons c

move /P ... -to /N ...

Rename term names. Terms /P ... of input relation are renamed corresponding terms /N .... The move operator provides same functionality as rename. Operator move is suitable for renaming few terms, operator rename for many terms. The following relmaps have same effect:

rename /N /P /N /P ...
move /P ... -to /N ...

relmapMove :: Intmed c -> ([TermName], [TermName]) -> Relmap c

Create move relmap.

relkitMove :: ([TermName], [TermName]) -> RelkitFlow c

Create move relkit.

rename

consRename :: RopCons c

rename /N /P /N /P ...

Rename term names. Term /P is renamed to term /N for each pair of terms /N /P. The rename operator provides same functionality as move.

relmapRename :: Intmed c -> [TermName2] -> Relmap c

Create rename relmap.