Koshucode.Baala.Rop.Nest.Confl

Description

Confluent operators for nested relations.

copy

consCopy :: RopCons c

copy N R

Copy input relation and referenced by ^N in R.

For example, down /r can be rewrite to copy r ( dee | slice /r ^r ).

for

consFor :: CRel c => RopCons c

for /P R

Map nested relation at /P by relmap R.

The relamp for /r ( cut /c ) removes term /c from a nested relation at the term /r.

relmapFor :: CRel c => Intmed c -> TermName -> Map (Relmap c)

Create for relmap.

relkitFor :: forall c. CRel c => TermName -> RelkitBinary c

Create for relkit.

group

consGroup :: (Ord c, CRel c) => RopCons c

group R -to /N

Group relation from relmap R by input relation. Grouped relations are added as content of term /N to each input tuples.

relmapGroup :: (Ord c, CRel c) => Intmed c -> SharedTerms -> TermName -> Map (Relmap c)

Create group relmap.

relkitGroup :: forall c. (Ord c, CRel c) => SharedTerms -> TermName -> RelkitBinary c

Create group relkit.

slice

consSlice :: CRel c => RopCons c

slice /N [R]

Slice input relation. Each slices have single tuple of input relation, and added as content of term /N to each input tuples. If R is given, slices are mapped by R. If input relation has a nested relation at term /P, the relation can be referenced by ^/P in R.

relmapSlice :: CRel c => Intmed c -> TermName -> Map (Relmap c)

Create slice relmap.

relkitSlice :: CRel c => TermName -> RelkitBinary c

Create slice relkit.

slice-up

consSliceUp :: CRel c => RopCons c

slice-up R

relmapSliceUp :: CRel c => Intmed c -> Map (Relmap c)

Create slice-up relmap.

relkitSliceUp :: CRel c => RelkitBinary c

Create slice-up relkit.