Koshucode.Baala.Subtext.Operator.Char

Description

Operators for text match.

Type

type CharExpr = Expr Char

Subtext expression for character input.

Character

char :: [Char] -> CharExpr

Match some character in a list of characters.

word :: String -> CharExpr

Match some word in a list of words from space-separated string.

space :: CharExpr

Match space character.

spaces :: CharExpr

Match zero-or-more space characters.

spaces1 :: CharExpr

Match one-or-more space characters.

digit :: CharExpr

Match digit characters.

letter :: CharExpr

Match letter characters.

ascii :: CharExpr

Match first 128 characters of Unicode character set.

latin1 :: CharExpr

Match first 256 characters of Unicode character set.

General category

categorySet :: Set GeneralCategory -> CharExpr

Test unicode general category with category set.

categoryList :: [GeneralCategory] -> CharExpr

Test unicode general category with category list.

category :: String -> Either String CharExpr

Test unicode general category with short names. The argument string is space-separated names. If unknown name is given, returns the name in Left data.

Conversion

asLower :: CharExpr -> CharExpr

Convert matched text into lower case.

asUpper :: CharExpr -> CharExpr

Convert matched text into upper case.