Koshucode.Baala.Syntax.Attr.ParaSpec

Description

Parameter specification.

Specification

data ParaSpec n

Parameter specification.

Constructors

ParaSpec

Fields

paraSpecPos :: ParaSpecPos

Positional parameter

paraSpecReq :: [n]

Required parameter

paraSpecOpt :: [n]

Optional parameter

paraSpecMult :: [n]

Multiple-occurence parameter

data ParaSpecPos

Positional parameter specification.

Constructors

ParaPosMin Int
Lower bound of parameter length

ParaPosRange Int Int
Lower and upper bound of parameter length

Construction

type ParaSpecMap n = Map (ParaSpec n)

paraSpec :: (Show n, Ord n) => ParaSpecMap n -> ParaSpec n

Positional parameter

paraMin :: Int -> ParaSpecMap n

Lower bound of length of positional parameter.

paraMax :: Int -> ParaSpecMap n

Upper bound of length of positional parameter.

paraJust :: Int -> ParaSpecMap n

Fixed-length positional parameter.

paraRange :: Int -> Int -> ParaSpecMap n

Lower and upper bound of length of positional parameter.

Named parameter

paraReq :: [n] -> ParaSpecMap n

Required named parameter.

paraOpt :: [n] -> ParaSpecMap n

Optional named parameter.

paraMult :: [n] -> ParaSpecMap n

Multiple-occurence parameter.

Unmatch reason

data ParaUnmatch n

Unmatch reason of real parameter and its specifition.

Constructors

ParaOutOfRange Int ParaSpecPos
Positional parameter is unmatched.

ParaUnknown [n]
Unknown parameter is specified.

ParaMissing [n]
Required parameter is missing.

ParaMultiple [n]
Parameter occurs more than once.

paraUnmatch :: forall n a. Eq n => Para n a -> ParaSpec n -> Maybe (ParaUnmatch n)

Create unmatch reason when parameter does not satisfies specification.

paraMatch :: Eq n => Para n a -> ParaSpec n -> Bool

Test parameter satisfies specification.

type ParaTo n a b = Para n a -> b

Map parameter to some value.

paraSelect :: Eq n => b -> [(ParaSpec n, ParaTo n a b)] -> ParaTo n a b

Select matched specification and apply ParaTo function.