Koshucode.Baala.Syntax.Subtree.Decode

Description

Decode subtree.

Decode

readSubtreeClauses :: TextualTermName t => FilePath -> IOAb [TokenClause t]

Read subtree clauses from file.

>>> mapM_ (S.printToks . B.clauseTokens) O.# B.abortLeft O.# readSubtreeClauses "subtree.txt"
********** > "a1"
TText /0.10.0/ TextRaw ">"
TText /0.10.2/ TextQQ "a1"
TOpen /0.0.0/ "("
TText /0.11.2/ TextRaw "-"
TText /0.11.4/ TextQQ "bb"
TText /0.0.0/ TextRaw "|"
TText /0.12.2/ TextRaw "-"
TText /0.12.4/ TextQQ "cc"
TClose /0.0.0/ ")"

subtreeClause :: Textual t => TokenClause t -> Ab (TokenClause t)

Insert open/sep/close to subtree clause.

decodeSubtreePattern :: (TextualTermName t, ToSivExpr t) => [TTree t] -> Ab [SubtreePattern]

Decode token trees to subtree patterns.

>>> S.withTrees decodeSubtreePattern "- \"Z1\""
Right [SubtreeL SubtreeNone (SivmapEq "Z1")]

>>> S.withTrees decodeSubtreePattern "> \"Y1\" ( - \"Z1\" )"
Right [SubtreeB SubtreeNone (SivmapEq "Y1") [SubtreeL SubtreeNone (SivmapEq "Z1")]]

>>> S.withTrees decodeSubtreePattern "> \"Y1\" ( - \"Z1\" || - \"Z2\" )"
Right [SubtreeB SubtreeNone (SivmapEq "Y1")
         [SubtreeL SubtreeNone (SivmapEq "Z1"),
          SubtreeL SubtreeNone (SivmapEq "Z2")]]

>>> S.withTrees decodeSubtreePattern ">> \"Y1\" ( - \"Z1\" )"
Right [SubtreeR (SivmapEq "Y1") [SubtreeL SubtreeNone (SivmapEq "Z1")]]

>>> S.withTrees decodeSubtreePattern ">> \"Y1\" ( - A B /z \"Z1\" )"
Right [SubtreeR (SivmapEq "Y1") [SubtreeL (SubtreeText ["A","B"] (TermName EQ "z")) (SivmapEq "Z1")]]

Syntax of subtree

Pattern

- Filter

Leaf which satisfies Filter.

> Filter { Pattern | ... }

Branch which satisfies Filter, and inner trees match Pattern.

>> Filter { Pattern | ... }

Recursive branch pattern include current branch, and inner trees match Pattern.

>>> Filter { Pattern | ... }

Recursive branch pattern only inner branches, and inner trees match Pattern.

Filter

Element | ...

Chain of filter Elements.

Element

"Text"

Node name is equal to Text.

? "Like"

Node name satisfies Like pattern.

! "Like"

Node name does not satisfies Like pattern.

id "Text"

Used for XML tree: Value of id attribute is equal to Text.

class "Text"

Used for XML tree: Space-delimited values of class attribute contains Text.

"Text" = "Text"

Used for XML tree: Attribute value is equal to Text.