Koshucode.Baala.Syntax.Symbol.AngleText

Description

Text enclosed in angle brackets.

Functions

angleQuote :: Textual t => t -> t

Convert string into double-quoted and angle-quoted form.

>>> putStrLn $ angleQuote "aaa"
"aaa"

>>> putStrLn $ angleQuote "\t"
<tab>

>>> putStrLn $ angleQuote "aaa\nbbb\r\nccc\r\n\r\nddd"
"aaa" <lf> "bbb" <crlf> "ccc" <crlf> <crlf> "ddd"

angleTexts :: Textual t => [(t, t)]

Table of coresspondences of angle text and its replacement.

>>> lookup "lf" angleTexts
Just "\n"

Keyword table

Some text literals are written using the following angle text.

<cr>

Carriage return

<crlf>

Carriage return and line feed

<lf>

Line feed

<q>

Single quote (')

<qq>

Double quote (")

<sp>

Space

<tab>

Tab