Koshucode.Baala.Syntax.Symbol.Next

Description

Next character sequence.

Next

type InputText = String

Input data type.

type Next a = InputText -> (InputText, a)

Split next character sequence from input text.

type AbNext a = InputText -> Ab (InputText, a)

Split next character sequence from input text.

nextSpace :: Next Int

Get next spaces. Space character is decided by isSpace.

>>> nextSpace " abc"
("abc", 1)

>>> nextSpace " \tabc"
("abc", 2)

nextQQ :: AbNext String

Get next double quoted text. A single double quote ends text. A double double quote does not end text, but it is converted to double quote.

>>> nextQQ "abc\" def"
Right (" def","abc")

>>> nextQQ "abc\"\"def\" ghi"
Right (" ghi","abc\"def")

Symbol

Data type

data Symbol

Symbol class.

Relation between character classes and symbol classes. The char class number' means Unicode number except for 0-9. G for general symbol, P for plain, N for numeric.

  Char class   Symbol class
  ------------ -----------------
  "0-9"        ( G ) ( P ) ( N )
  "-"          ( G ) ( P ) ( N )
  letter       ( G )   P
  mark         ( G )   P
  number'      ( G )   P
  "_?"         ( G )   P
  "+"          ( G )         N
  "*=<>~"        G
  ".#"                       N

Partial order of symbol classes.

                    ( GPN ) Common
                   /      |
          Plain ( GP )  ( GN ) Numeric
               /  |    /  |
  Short ( P.P )   |   /   |
           :      |  /    |
        General ( G )     |
           :      :     ( N ) Numeric
           :      :       :
           ................ Unknown

Constructors

SymbolCommon String
General-plain-numeric symbol

SymbolGeneral String
General symbol

SymbolPlain String
Plain symbol

SymbolNumeric String
Numeric symbol

SymbolShort String String
Short symbol (Plain "." Plain)

SymbolUnknown String
Unknown symbol

Symbol test

isGeneralSymbol :: Test Symbol

Test symbol is general, in other words, SymbolCommon, SymbolPlain or SymbolGeneral.

isPlainSymbol :: Test Symbol

Test symbol is plain, in other words, SymbolCommon or SymbolPlain.

isNumericSymbol :: Test Symbol

Test symbol is numeric, in other words, SymbolCommon or SymbolNumeric.

isShortSymbol :: Test Symbol

Test symbol is SymbolShort.

Char test

isSymbolChar :: Test Char

Test character is a symbol component.

isGeneralChar :: Test Char

Test character is a general-symbol component.

isPlainChar :: Test Char

Test character is a plain-symbol component.

isNumericChar :: Test Char

Test character is a numeric-symbol component.

Next symbol

nextSymbol :: Next Symbol

Get next symbol.

nextSymbolPlain :: AbNext String

Get next plain symbol.