Koshucode.Baala.Syntax.TTree.Parse

Description

Parser for token tree.

Parser

ttrees :: [Token] -> Ab [TTree]

Parse tokens with brackets into trees. Blank tokens and comments are excluded.

ttreeGroup :: [TTree] -> TTree

Wrap trees in group.

Abbreviation

tt :: String -> Ab [TTree]

Convert string to token trees.

tt1 :: String -> Ab TTree

Parse string and group it.

ttPrint :: String -> IO ()

Parse string and print it.

ttDoc :: [TTree] -> Doc

Get Doc value of token trees for pretty printing.

Split and divide

splitTokensBy :: Test String -> SplitList3e Token

Split token list by unquoted word. If token list contains the word, pair of before-list, the-word and after-list is returned. If does not contains the word, original token list is returned.

>>> let Right t = S.toks "b c"
>>> splitTokensBy (== "|") t
Left [ TText  CodePt {..} TextRaw "b"
     , TSpace CodePt {..} 1
     , TText  CodePt {..} TextRaw "c" ]

>>> let Right t = S.toks "a | b | c"
>>> splitTokensBy (== "|") t
Right ( [ TText CodePt {..} TextRaw "a", TSpace CodePt {..} 1]
      , TText CodePt {..} TextRaw "|"
      , [ TText  CodePt {..} TextRaw "b"
        , TSpace CodePt {..} 1
        , TText  CodePt {..} TextRaw "|"
        , TText  CodePt {..} TextRaw "c" ] )

splitTreesBy :: Test String -> SplitList3e TTree

Split token trees by quoteless token of given string.

divideTreesBy :: Test String -> [TTree] -> [[TTree]]

Divide token trees by quoteless token of given string.

divideTreesByBar :: [TTree] -> [[TTree]]

Divide token trees by vertical bar "|".

divideTreesByColon :: [TTree] -> [[TTree]]

Divide token trees by colon ":".

divideTreesByEqual :: [TTree] -> [[TTree]]

Divide token trees by equal sign "=".