Koshucode.Baala.Syntax.Token.Clip

Description

Clip token text.

Char test

isSymbol :: Test Char

Test character is symbolic.

isSpace :: Test Char

Test character is space.

isQQ :: Test Char

Test character is double-quote.

isTerm :: Test Char

Test character is term slash.

isJudge :: Test Char

Test character is content line, i.e., '-' or '='.

isClock :: Test Char

Test character is component of clock text.

Type

type TokenScan = CodeScan String Token

Code scanner for token list.

type TokenScanMap = Map TokenScan

Read single token.

type ClipResult = (InputText, Token)

Clip single-token result.

type ClipResultC = (WordCache, InputText, Token)

Clip single-token result with word cache.

type ClipResultCL = (WordCache, InputText, [Token])

Clip multiple-tokens result with word cache.

type ClipToken = CodePos -> InputText -> ClipResult

Clip a next token.

type ClipTokenC = CodePos -> WordCache -> InputText -> ClipResultC

Clip a next token with word cache.

type ClipTokenCL = CodePos -> WordCache -> InputText -> ClipResultCL

Clip a next token with word cache.

Clip

Updater

clipUpdate :: TokenScan -> ClipResult -> TokenScan

Update token scanner by clip result.

clipUpdateC :: TokenScan -> ClipResultC -> TokenScan

Update token scanner by clip result with word cache.

clipUpdateCL :: TokenScan -> ClipResultCL -> TokenScan

Update token scanner by clip result with word cache.

Textual

clipSpace :: ClipToken

Clip space token.

>>> clipSpace B.def "  foo bar baz"
("foo bar baz", TSpace /0.0.0/ 3)

clipQ :: ClipTokenC

Clip single-quoted text.

>>> clipQ B.def O.emptyWordCache "foo bar baz"
(Cache 8 ["foo"] [], " bar baz", TText /0.0.0/ TextQ "foo")

clipQQ :: ClipTokenC

Clip double-quoted text.

>>> clipQQ B.def O.emptyWordCache "foo\" bar baz"
(Cache 8 ["foo"] [], " bar baz", TText /0.0.0/ TextQQ "foo")

Identifier

clipSymbol :: ClipTokenC

Clip symbolic token.

>>> clipSymbol B.def O.emptyWordCache "foo bar baz"
(Cache 8 ["foo"] [], " bar baz", TText /0.0.0/ TextRaw "foo")

clipSlot :: Int -> ClipToken

Clip slot name, like @foo.

>>> clipSlot 1 B.def "foo bar baz"
(" bar baz", TSlot /0.0.0/ 1 "foo")

Term name

clipTermSign :: String -> ClipTokenCL

Clip signed term name.

>>> clipTermSign "+/" B.def Map.empty "foo bar baz"
(fromList [("foo","foo")], " bar baz", [TTerm <I0-L0-C0> "+/foo"])

clipTermName :: ClipTokenCL

Clip term name.

>>> clipTermName B.def Map.empty "foo bar baz"
(fromList [("foo","foo")], " bar baz", [TTerm <I0-L0-C0> "/foo"])

clipTermQ :: ClipTokenCL

Clip quoted term.

>>> clipTermQ B.def Map.empty "foo bar baz"
(fromList [], " bar baz", [TText <I0-L0-C0> TextTerm "foo"])

Symbol

clipBar :: CodePos -> String -> ClipResult

Clip token beginning with "|".

>>> clipBar B.def "-- C"
(" C", TText /0.0.0/ TextBar "|--")

>>> clipBar B.def "| ..."
("...", TText /0.0.0/ TextRaw "||")

>>> clipBar B.def "12:00| ..."
(" ...", TText /0.0.0/ TextBar "|12:00|")