Koshucode.Baala.Syntax.Token.TokenTree

Description

Tree of tokens

Token tree

Tree type

type TTree = CodeTree BracketType Token

Tree of tokens.

type NamedTree = Named TTree

Pair of token tree and its name.

type NamedTrees = Named [TTree]

Pair of token trees and its name.

Conversion

type TTreeTo a = TTree -> a

Convert token tree to something.

type TTreesTo a = [TTree] -> a

Convert list of token tree to sometning.

type TTreeToAb a = TTree -> Ab a

Convert token tree to something, abortable.

type TTreesToAb a = [TTree] -> Ab a

Convert list of token tree to sometning, abortable.

Parser

ttrees :: [Token] -> Ab [TTree]

Parse tokens with brackets into trees. Blank tokens and comments are excluded.

ttreeGroup :: TTreesTo TTree

Wrap trees in group.

Pattern

Term

pattern TermLeaf :: CodePt -> TermType -> TermPath -> CodeTree t Token

pattern TermLeafName :: CodePt -> Ordering -> TermName -> CodeTree t Token

pattern TermLeafPath :: CodePt -> TermPath -> CodeTree t Token

pattern TermLeafLocal :: CodePt -> Local String -> Int -> [Token] -> CodeTree t Token

Text

pattern TextLeaf :: TextForm -> CodePt -> String -> CodeTree t Token

Text leaf.

pattern TextLeafRaw :: CodePt -> String -> CodeTree t Token

Text leaf of TextRaw.

pattern TextLeafAttr :: CodePt -> [Char] -> CodeTree t Token

Text leaf beginning with single hyphen.

pattern TextLeafAttr2 :: CodePt -> [Char] -> CodeTree t Token

Text leaf beginning with double hyphens.

pattern TextLeafQ :: CodePt -> String -> CodeTree t Token

Text leaf of TextQ.

pattern TextLeafQQ :: CodePt -> String -> CodeTree t Token

Text leaf of TextQQ.

pattern TextLeafKey :: CodePt -> String -> CodeTree t Token

Text leaf of TextKey.

Divide trees

splitTokensBy

Arguments

:: Pred String
Predicate

-> [Token]
Tokens

-> Either [Token] ([Token], Token, [Token])
Original-tokens or (before-list, the-word, after-list)

Split token list by unquoted word. If token list contains the word, pair of before-list, the-word and after-list is returned. If does not contains the word, original token list is returned.

>>> splitTokensBy (== "|") . tokens $ "b c"
Left [ TText 1 0 "b", TSpace 2 1, TText 3 0 "c" ]

>>> splitTokensBy (== "|") . tokens $ "a | b | c"
Right ( [ TText 1 0 "a", TSpace 2 1 ]
        , TText 3 0 "|"
        , [ TSpace 4 1, TText 5 0 "b", TSpace 6 1
          , TText 7 0 "|", TSpace 8 1, TText 9 0 "c" ] )

divideTreesBy :: String -> TTreesTo [[TTree]]

Divide token trees by quoteless token of given string.

divideTreesByBar :: TTreesTo [[TTree]]

Divide token trees by vertical bar "|".

divideTreesByColon :: TTreesTo [[TTree]]

Divide token trees by colon ":".

divideTreesByEqual :: TTreesTo [[TTree]]

Divide token trees by equal sign "=".

Abbreviation

tt :: String -> Ab [TTree]

Convert text to token trees.

tt1 :: String -> Ab TTree

Parse string and group it.

ttPrint :: String -> IO ()

Parse string and print it.

ttDoc :: TTreesTo Doc

Get Doc value of token trees for pretty printing.