Koshucode.Baala.Syntax.Tree.Split

Description

Split token trees.

Split

splitTokensBy :: Test t -> SplitList3e (TToken t)

Split token list by unquoted word. If token list contains the word, pair of before-list, the-word and after-list is returned. If does not contains the word, original token list is returned.

>>> let Right t = S.toks "b c"
>>> splitTokensBy (== "|") t
Left [ TText  CodePos {..} TextRaw "b"
     , TSpace CodePos {..} 1
     , TText  CodePos {..} TextRaw "c" ]

>>> let Right t = S.toks "a | b | c"
>>> splitTokensBy (== "|") t
Right ( [ TText CodePos {..} TextRaw "a", TSpace CodePos {..} 1]
      , TText CodePos {..} TextRaw "|"
      , [ TText  CodePos {..} TextRaw "b"
        , TSpace CodePos {..} 1
        , TText  CodePos {..} TextRaw "|"
        , TText  CodePos {..} TextRaw "c" ] )

splitTreesBy :: Test t -> SplitList3e (TTree t)

Split token trees by quoteless token of given string.

Divide

divideTreesBy :: Test t -> [TTree t] -> [[TTree t]]

Divide token trees by raw text of given string.

divideTreesByBar :: Textual t => [TTree t] -> [[TTree t]]

Divide token trees by vertical bar "|".

>>> S.withTrees (Right . divideTreesByBar) "a | b x | c y z"
Right [ [ TreeL (TText /0.1.0/ TextRaw "a") ]
      , [ TreeL (TText /0.1.4/ TextRaw "b")
        , TreeL (TText /0.1.6/ TextRaw "x") ]
      , [ TreeL (TText /0.1.10/ TextRaw "c")
        , TreeL (TText /0.1.12/ TextRaw "y")
        , TreeL (TText /0.1.14/ TextRaw "z") ]]

divideTreesByBar2 :: Textual t => [TTree t] -> [[TTree t]]

Divide token trees by double vertical bar "||".

divideTreesByColon :: Textual t => [TTree t] -> [[TTree t]]

Divide token trees by colon ":".

divideTreesByEqual :: Textual t => [TTree t] -> [[TTree t]]

Divide token trees by equal sign "=".