Koshucode.Baala.Toolkit.Library.Change

Description

Functions for changes between two datasets.

Minus

minusInput

Arguments

:: Input
B: Source of base section

-> Input
A: Source of altered section

-> IO ()

Calculate and write judges B minus A.

minusInputJudge :: Input -> Input -> IO [Judge VContent]

minusJudge :: Ord c => [Judge c] -> [Judge c] -> [Judge c]

Update

updateInput

Arguments

:: Input
B: Source of base section

-> Input
C: Source of change section

-> IO ()

Calculate and write judges B update by C.

updateJudge :: Ord c => [Judge c] -> [Judge c] -> [Judge c]

Changeset

Rules for calculating changeset C = A - B. Altering dataset B into A, changeset C is calculated by the following rules.

Affirmed judge is in C if and only if the judge is not in A, and is in B.

Denied judge is in C if and only if the judge is not in B, and is in A.

Judge is not in C otherwise.