Koshucode.Baala.Toolkit.Main.KoshuChange

Description

Koshucode syntactic tool.

Documentation

koshuChangeMain :: IO ()

The main function for koshu-change command.

koshu-change.hs

koshu-change command is implemented using koshuChangeMain.

 import Koshucode.Baala.Toolkit.Main.KoshuChange
 
 main :: IO ()
 main = koshuChangeMain

Changeset

Rules for calculating changeset C = A - B. Altering dataset B into A, changeset C is calculated by the following rules.

Affirmed judge is in C if and only if the judge is not in A, and is in B.

Denied judge is in C if and only if the judge is not in B, and is in A.

Judge is not in C otherwise.