Koshucode.Baala.Type.Rel.Rel

Description

Relation type

Data type

data Rel c

Relations on type c. Relation is consist of heading and body. Body is thoretically a set of tuples, but implemented using list of list.

Constructors

Rel

Fields

relHead :: Head

Heading of relation

relBody :: Body c

Body of relation

type Body c = [[c]]

Body of relation, i.e., a list of tuples. Tuple is a list of contents.

data JudgeRel c

Relation with judgement class.

Constructors

JudgeRel JudgeClass (Rel c)

relSort :: Ord c => Map (Rel c)

Sort head and body of relation.

Constant

reldum :: Rel c

The nullary empty relation. In other words, relational constant that has no terms and no tuples.

>>> B.doc (reldum :: Rel Bool)
{= =}

reldee :: Rel c

The nullary full relation. In other words, relational constant that has no terms and the empty tuple.

>>> B.doc (reldee :: Rel Bool)
{= [] =}

reldau :: Rel c

The empty relation with all terms. This is fake.

Converter

class SelectRel r where

Type for having relations.

Minimal complete definition

selectRel

Methods

selectRel :: (Ord c, Default c) => r c -> RelSelect c

Convert judges to relation.

selectJudgeRel :: (Ord c, Default c) => r c -> JudgeRelSelect c

Convert judges to judgemental relation.

type RelSelect c = JudgeClass -> [TermName] -> Rel c

Select relation.

type JudgeRelSelect c = JudgeClass -> [TermName] -> JudgeRel c

Select judgemental relation.

judgesFromRel :: JudgeOf c -> JudgeClass -> Rel c -> [Judge c]

Convert relation to list of judges.