Koshucode.Baala.Type.Rel.TermPicker

Description

Term picker is a data for picking target terms based on input heading terms.

Construct

type TermPicker c = Picker TermName c

Data for picking shared and proper terms.

type TermPick c = TermPicker c -> [c] -> [c]

Type for picking terms.

type TermPick2 a b = (TermPick a, TermPick b)

Double picker.

termPicker :: (GetTermNames target, GetTermNames input) => target -> input -> TermPicker c

Create term picker from left and right term names

Presend & new terms

preTerms :: TermPicker c -> [TermName]

List of present terms. The first argument of termPicker is treated as target terms, and the second is terms of input heading. Because firstly target terms are given to relmap operator, secondly input heading is fixed.

>>> preTerms $ termPicker "/a /b /c" "/a /b /d /e"
[TermName EQ "a",TermName EQ "b"]

newTerms :: TermPicker c -> [TermName]

List of new terms.

>>> newTerms $ termPicker "/a /b /c" "/a /b /d /e"
[TermName EQ "c"]

preTermsExist :: Test (TermPicker c)

Test present terms exist.

>>> preTermsExist $ termPicker "/d /e" "/a /b /c"
False

>>> preTermsExist $ termPicker "/c /d /e" "/a /b /c"
True

newTermsExist :: Test (TermPicker c)

Test new terms exist.

>>> newTermsExist $ termPicker "/a /b /c" "/a /b /d /e"
True

>>> newTermsExist $ termPicker "/a /b" "/a /b /d /e"
False

Mapping

termsIndex :: TermPicker c -> [Int]

Indices of target terms on input terms.

>>> termsIndex $ termPicker "/b /d" "/a /b /c /d"
[1,3]

pickDirect :: (GetTermNames target, GetTermNames input) => target -> input -> Map [c]

Pick contents of target terms.

>>> pickDirect "/b /d" "/a /b /c /d" "ABCD"
"BD"

pickTerms :: TermPick c

Pick target terms from input terms.

>>> pickTerms (termPicker "/b /d" "/a /b /c /d") "ABCD"
"BD"

cutTerms :: TermPick c

Cut target terms from input terms.

>>> cutTerms (termPicker "/b /d" "/a /b /c /d") "ABCD"
"AC"

forwardTerms :: TermPick c

Move target terms forward.

>>> forwardTerms (termPicker "/b /d" "/a /b /c /d") "ABCD"
"BDAC"

backwardTerms :: TermPick c

Move target terms backward.

>>> backwardTerms (termPicker "/b /d" "/a /b /c /d") "ABCD"
"ACBD"

towardTerms :: Bool -> TermPick c

Move target terms forward if True or backward if False.

Double mapping

pickTerms2 :: TermPick2 a b

Double pickTerms.

cutTerms2 :: TermPick2 a b

Double cutTerms.

forwardTerms2 :: TermPick2 a b

Double forwardTerms.

backwardTerms2 :: TermPick2 a b

Double backwardTerms.