Koshucode.Baala.Type.Time.Time

Description

Time

Data type

data Time

Time is a small duration on timeline.

Constructors

TimeYmdcz Date Clock Zone
Date and clock with time zone

TimeYmdc Date Clock
Date and clock

TimeYmd Date
Year, month, and day

TimeYw Mjd
Year and week

TimeYm Mjd
Year and month

type Zone = Sec

Time zone as offset from UTC.

timeYmdTuple :: Time -> Ymd

Convert time to year/month/day tuple.

>>> timeYmdTuple $ timeFromMjd 55555
(2010,12,25)

timePrecision :: Time -> String

Get the name of time precision.

Creation

monthlyTime :: (ToMjdClip day, ToClock clock) => day -> Maybe clock -> Ab Time

Create monthly time.

>>> monthlyTime (D.DateYmd 2013 4 18) (Just $ D.ClockPartsMin 0 12 05)
Right 2013-04-18 12:05

>>> monthlyTime (D.DateYmd 2013 4 18) (Nothing :: Maybe D.ClockParts)
Right 2013-04-18

>>> monthlyTime (D.DateYmd 2013 4 31) (Nothing :: Maybe D.ClockParts)
Left AbortReason "Not monthly date" ...

monthlyClippedTime :: (ToMjdClip day, ToClock clock) => day -> Maybe clock -> Time

Create monthly clipped time.

monthlyTimeDate :: ToMjdClip day => day -> Ab Time

Create monthly day-precision time.

monthlyTimeClock :: (ToMjdClip day, ToClock clock) => day -> clock -> Ab Time

Create monthly clock-precision time.

timeFromMjd :: ToMjd mjd => mjd -> Time

Create time data from the Modified Julian day.

>>> timeFromMjd (55555 :: Int)
2010-12-25

timeFromYmAb :: Year -> Month -> Ab Time

Create time data from year and month.

timeFromYwAb :: Year -> Week -> Ab Time

Create time data from year and week.

timeFromDczAb :: Date -> Clock -> Maybe Zone -> Ab Time

Create time data from date, clock, and time zone.

timeFromYmd :: Year -> Month -> Day -> Time

Create time from year, month, and day.

>>> timeFromYmd 2013 4 18
2013-04-18

>>> timeFromYmd 2013 4 0
2013-04-01

>>> timeFromYmd 2013 4 33
2013-04-30

timeFromYmdTuple :: Ymd -> Time

Craete time from tuple of year, month, and day.

>>> timeFromYmdTuple (2013, 4, 18)
2013-04-18

Construction with I/O

nowUtc :: IO Time

Get UTC.

>>> nowUtc
2013-04-18 08:37:36 UTC

nowZoned :: IO Time

Get current time with time zone.

>>> nowZoned
2013-04-18 17:37:36 +09:00

now :: IO Time

Get current local time without time zone.

>>> now
2013-04-18 17:37:36

today :: IO Time

Get today.

>>> today
2013-04-18

Conversion

timeCutClock :: Map Time

Cut clock part.

timeAltZone :: Map Zone -> Map Time

Alter time zone part.

timeCutZone :: Map Time

Cut timezone.

timeLocalize :: Map Time

Cut timezone and convert to local time, i.e., addition of UTC and time zone.

timeAltDate :: Map Date -> Map Time

Alter day part.

timeAltDays :: Map Days -> Map Time

Alter integer content of the Modified Julian Day.

timeFromZonedTime :: ZonedTime -> Time

Convert local time to Koshu time content.

timeDaysSec :: Time -> DaysSec

Days and seconds of time.

>>> timeDaysSec $ timeFromMjd 55555
(55555,0)