Koshucode.Baala.Type.Type

Description

Type for types.

Type

data Type

Type for types.

Constructors

TypeAny
Everything

TypeType
Type of types

TypeTerm
Term name

TypeInterp
Data interpreation

TypeEmpty
Empty

TypeEnd
End

TypeBool
Boolean

TypeText
Text

TypeCode
Code

TypeDec
Decimal

TypeClock (Maybe String)
Clock

TypeTime (Maybe String)
Time

TypeBin
Binary data

TypeList Type
List

TypeSet Type
Set

TypeTag String Type
Tagged type

TypeTie [Term Type]
Tie

TypeRel [Term Type]
Relation

TypeTuple [Type]
Tuple (Product type)

TypeSum [Type]
Sum type

type NamedType = Named Type

Term type, i.e., term name and type of its content.

typeExplain :: Type -> Doc

Print type as tree.

typeTermMix :: Type -> MixText

Encode term types.

Utility

Construct

typeFlatRel :: [TermName] -> Type

Create relation type from term names.

typeConsRel :: TermName -> Map Type

Add term name to relation type.

typeConsNest :: TermName -> Type -> Map Type

Add term name and type to relation type.

typeAppendRel :: [TermName] -> Map Type

Add term names to relation type.

Select

typeRelTermNames :: Type -> [TermName]

Get term names from relation type.

typeRelDegree :: Type -> Int

Get degree of relation type.

typeTerms :: Type -> [Term Type]

Get named type from relation type or tie type.

isTypeRel :: Type -> Bool

Test type is relational.

typeRelIndex :: Type -> TermPath -> [Int]

Calculate term index.

>>> typeRelIndex (typeFlatRel ["a", "b", "c"]) ["b"]
[1]

>>> typeRelIndex (typeFlatRel ["a", "b", "c"]) ["e"]
[-1]

>>> typeRelIndex (typeConsNest "r" (typeFlatRel ["a", "b"]) (typeFlatRel [])) ["r", "b"]
[0, 1]

Modify

typeRelMapTerms :: Map [Term Type] -> Map Type

Modify terms of relation type.

typeRelMapTerm :: Map (Term Type) -> Map Type

Modify term of relation type.

typeRelMapName :: Map TermName -> Map Type

Modify term name of relation type.

Empty

Empty means that there are no values. i.e., universal negation on the term holds. Textual form is the non-quoted parens: ().

Boolean

Boolean used for something is hold or unhold. Textual forms: <0> (false), <1> (true).

Text

Sequence of characters. Textual forms is chars with apostrophe or doubly-quoted line: 'abc, "abc def".

Decimal

Decimal number. Textual forms is sequence of digits: 100, 99.50, hex AF.

Set

Set is an unordered collection of contents. Duplication among contents is not significant. Textual form is a sequence of contents delimited by colon, enclosed in braces: { 'a | 'b | 'c }.

List

List is an ordered list of contents. Textual form is a sequence of contents delimited by colon, enclosed in square brackets: [ 'abc | 'def ].

Tie

Tie is an bundle of terms, i.e., a list of named contents. Textual form is a sequence of terms with hyphen-braces: {- /a 10 /b 20 -}.

Relation

Relation is a set of same-type tuples, Textual form is a sequence of tuples enclosed in equal-braces. The first tuple is a heading of relation, and succeeding tuples are delimited by vertical bar: {= /a /b [ 'A1 | 20 ][ 'A3 | 40 ] =}.