Koshucode.Baala.Base.Code.Scan

Description

Code scanner.

Type

data CodeScan i o

Code scanner divides input text into output tokens.

Constructors

CodeScan

Fields

codeMapSaved :: [CodeScanMap i o]

Saved updater

codeMap :: CodeScanMap i o

Updater

codeInputPt :: CodePos

Code position

codeInput :: i

Input text

codeOutput :: [o]

Output tokens

codeWords :: WordCache

Cached words

type CodeScanMap i o = Map (CodeScan i o)

Update code scanner.

type WordCache = CacheS String

Cached words.

emptyWordCache :: WordCache

Empty word cache.

Function

isBol :: Test (CodeScan i o)

Test scanner at the beginning of line, i.e., no output collected.

codeScanSave :: CodeScanMap i o

Save current updater.

codeScanRestore :: CodeScanMap i o

Restore saved updater.

codeScanUp :: CodeScanMap String o -> IxIOPoint -> String -> [CodeLine o]

Split source text into CodeLine list.

Split source text into lines by line delimiters (carriage return \r or line feed \n).

Numbering lines from 1. Internally, this is represented as a list of pairs (LineNumber, String).

Tokenize each lines, and put tokens together in CodeLine.

codeScanUpBz :: ToCode code => CodeScanMap String o -> IxIOPoint -> code -> [CodeLine o]

Lazy bytestring version of codeScanUp.

codeUpdate :: i -> o -> CodeScanMap i o

Update codeInput and push result element to codeOutput.

codeUpdateWords :: WordCache -> i -> o -> CodeScanMap i o

Update code scanner with word table.

codeUpdateList :: i -> [o] -> CodeScanMap i o

Multi-element version of codeUpdate.

codeUpdateListWords :: WordCache -> i -> [o] -> CodeScanMap i o

Multi-element and cached version of codeUpdate.

codeChange :: Map (CodeScanMap i o)

Change mapper of code sc.