Koshucode.Baala.Base.Data.Judge

Description

Judgements: a symbolic representations of affirmed or denied statements.

Datatype

data Judge v

Judgement on type v.

Judgement (or judge for short) is divided into three parts: logical quality, sign of relation, and argument. Bool value of logical quality corresponds to affirmed or denied judge. String value of sign represents certain sentence pattern that gives intepretation of data. Sentence pattern has placeholders filled by (String, v) values of argument.

Constructors

Judge Bool Relsign (Relarg v)

type Relsign = String

Sign of relation, or relation name.

type Relarg v = [Named v]

List of terms.

Writer

putJudges :: (Ord v, Pretty v) => [Judge v] -> IO ()

hPutJudges :: (Ord v, Pretty v) => Handle -> [Judge v] -> IO ()

abcJudge :: Ord v => Judge v -> Judge v

Sort terms in alphabetical order.

Logical quality

affirmJudge :: Map (Judge v)

Affirm judge, i.e., change logical quality to True.

denyJudge :: Map (Judge v)

Deny judge, i.e., change logical quality to False.

isAffirmed :: Judge t -> Bool

isDenied :: Judge t -> Bool