Koshucode.Baala.Base.Data.Judge

Description

Judgements: a symbolic representations of affirmed or denied statements.

Datatype

data Judge c

Judgement on type c.

Judgement (or judge for short) is divided into three parts: logical quality, name of pattern, and argument. Boolean values True or False of logical quality corresponds to affirmed or denied judge. A name of judgement pattern represents certain sentence pattern that gives intepretation of data. Sentence pattern has placeholders filled by Named c in argument.

Constructors

JudgeAffirm JudgePat [Term c]
|-- P /x 10 /y 20

JudgeDeny JudgePat [Term c]
|-x P /x 10 /y 20

JudgeMultiDeny JudgePat [Term c]
|-xx P /x 10 /y 20

JudgeChange JudgePat [Term c] [Term c]
|-c P /x 10 +/y 20

JudgeMultiChange JudgePat [Term c] [Term c]
|-cc P /x 10 +/y 20

JudgeViolate JudgePat [Term c]
|-v P /x 10 /y 20

type JudgePat = String

Name of judgement pattern, in other words, name of propositional function.

judgePat :: Judge c -> JudgePat

Return pattern of judgement.

judgeTerms :: Judge c -> [Term c]

Return term list of judgement.

judgeTermsMap :: ([Term a] -> [Term b]) -> Judge a -> Judge b

judgeCons :: Term c -> Map (Judge c)

Prepend a term into judgement.

sortJudgeTerms :: Ord c => Map (Judge c)

Sort terms in alphabetical order.

Logical quality

type JudgeOf c = JudgePat -> [Term c] -> Judge c

Construct judgement from its pattern and terms.

affirm :: JudgeOf c

Construct affirmative judgement.

deny :: JudgeOf c

Construct denial judgement.

affirmJudge :: Map (Judge c)

Affirm judgement, i.e., change logical quality to affirmative.

denyJudge :: Map (Judge c)

Deny judgement, i.e., change logical quality to denial.

isAffirmative :: Judge c -> Bool

Test which judgement is affirmed.

isDenial :: Judge c -> Bool

Test which judgement is denied.

isViolative :: Judge c -> Bool

Test which judgement is for violation.

Writer

writeDownJudge :: Write c => StringMap -> Judge c -> String

writeDownTerms :: Write c => StringMap -> [Term c] -> String

textualjudge :: Write c => StringMap -> Judge c -> Judge String

judgeText :: Judge String -> String

termText :: TermName -> String -> String

termsText :: [Term String] -> String

putJudge :: Write c => Judge c -> IO ()

hPutJudge :: Write c => Handle -> Judge c -> IO ()