Koshucode.Baala.Base.Data.Judge

Description

Judgements: a symbolic representations of affirmed or denied statements.

Documentation

data Judge v

Judgement on type v.

Judgement (or judge for short) is divided into three parts: logical quality, sign of relation, and argument. Bool value of logical quality corresponds to affirmed or denied judge. String value of sign represents certain sentence pattern that gives intepretation of data. Sentence pattern has placeholders filled by (String, v) values of argument.

Constructors

Judge Bool Relsign (Relarg v)

type Relsign = String

Sign of relation, or relation name.

type Relarg v = [(String, v)]

List of terms.

abcJudge :: Ord v => Judge v -> Judge v

Sort terms in alphabetical order.