Koshucode.Baala.Base.Data.Judge

Description

Judgements: a symbolic representations of affirmed or denied statements.

Datatype

data Judge c

Judgement on type c.

Judgement (or judge for short) is divided into three parts: logical quality, name of pattern, and argument. Boolean values True or False of logical quality corresponds to affirmed or denied judge. A name of judgement pattern represents certain sentence pattern that gives intepretation of data. Sentence pattern has placeholders filled by Named c in argument.

Constructors

Judge Bool JudgePattern [Named c]

type JudgePattern = String

Name of judgement pattern.

abcJudge :: Ord c => Map (Judge c)

Sort terms in alphabetical order.

Logical quality

affirm :: JudgePattern -> [Named c] -> Judge c

Construct affirmed judgement.

deny :: JudgePattern -> [Named c] -> Judge c

Construct denied judgement.

affirmJudge :: Map (Judge c)

Affirm judgement, i.e., change logical quality to True.

denyJudge :: Map (Judge c)

Deny judgement, i.e., change logical quality to False.

isAffirmed :: Judge c -> Bool

Test that judgement is affirmd.

isDenied :: Judge c -> Bool

Test that judgement is denied.

>>> isDenied $ Judge True "A" []
False

Writer

type ShortJudge c = Short [Judge c]

putJudges :: (Ord c, Pretty c) => Int -> [Judge c] -> IO Int

Print judges to stdout.

hPutJudges :: (Ord c, Pretty c) => Handle -> ([ShortJudge c], [ShortJudge c]) -> IO Int

hPutJudgesFlat :: (Ord c, Pretty c) => Handle -> Int -> [Judge c] -> IO Int