Koshucode.Baala.Base.Data.Rel

Description

Relation type

Data type

data Rel c

Relations on type c. Relation is consist of heading and body. Body is thoretically a set of tuples, but implemented using list of list.

Constructors

Rel

Fields

relHead :: Head

Heading of relation

relBody :: Body c

Body of relation

type Body c = [[c]]

Body of relation, i.e., a list of tuples. Tuple is a list of contents.

relSort :: Ord c => Map (Rel c)

Constant

reldum :: Rel c

The nullary empty relation. In other words, relational constant that has no terms and no tuples.

reldee :: Rel c

The nullary full relation. In other words, relational constant that has no terms and the empty tuple.

reldau :: Rel c

The empty relation with all terms. This is fake.

Examples

>>> B.doc (reldum :: Rel Bool)
{| |}

>>> B.doc (reldee :: Rel Bool)
{| | |}