Koshucode.Baala.Base.Prelude.Import

Control.Monad

(<=<) :: Monad m => (b -> m c) -> (a -> m b) -> a -> m c infixr 1

Right-to-left Kleisli composition of monads. (>=>), with the arguments flipped

(>=>) :: Monad m => (a -> m b) -> (b -> m c) -> a -> m c infixr 1

Left-to-right Kleisli composition of monads.

sequence :: Monad m => [m a] -> m [a]

Evaluate each action in the sequence from left to right, and collect the results.

filterM :: Monad m => (a -> m Bool) -> [a] -> m [a]

This generalizes the list-based filter function.

foldM :: Monad m => (a -> b -> m a) -> a -> [b] -> m a

The foldM function is analogous to foldl, except that its result is encapsulated in a monad. Note that foldM works from left-to-right over the list arguments. This could be an issue where (>>) and the `folded function' are not commutative.

      foldM f a1 [x1, x2, ..., xm]

==

      do
        a2 <- f a1 x1
        a3 <- f a2 x2
        ...
        f am xm

If right-to-left evaluation is required, the input list should be reversed.

when :: Monad m => Bool -> m () -> m ()

Conditional execution of monadic expressions. For example,

      when debug (putStr "Debugging\n")

will output the string Debugging\n if the Boolean value debug is True, and otherwise do nothing.

unless :: Monad m => Bool -> m () -> m ()

The reverse of when.

guard :: MonadPlus m => Bool -> m ()

guard b is return () if b is True, and mzero if b is False.

concatMapM :: Monad m => (a -> m [b]) -> [a] -> m [b]

Data.List

sort :: Ord a => [a] -> [a]

The sort function implements a stable sorting algorithm. It is a special case of sortBy, which allows the programmer to supply their own comparison function.

transpose :: [[a]] -> [[a]]

The transpose function transposes the rows and columns of its argument. For example,

transpose [[1,2,3],[4,5,6]] == [[1,4],[2,5],[3,6]]

intercalate :: [a] -> [[a]] -> [a]

intercalate xs xss is equivalent to (concat (intersperse xs xss)). It inserts the list xs in between the lists in xss and concatenates the result.

Data.Map

lookupMap :: Ord k => k -> Map k a -> Maybe a

Same as lookup in Data.Map module.

Data.Maybe

catMaybes :: [Maybe a] -> [a]

The catMaybes function takes a list of Maybes and returns a list of all the Just values.

mapMaybe :: (a -> Maybe b) -> [a] -> [b]

The mapMaybe function is a version of map which can throw out elements. In particular, the functional argument returns something of type Maybe b. If this is Nothing, no element is added on to the result list. If it just Just b, then b is included in the result list.

fromJust :: Maybe a -> a

The fromJust function extracts the element out of a Just and throws an error if its argument is Nothing.

fromMaybe :: a -> Maybe a -> a

The fromMaybe function takes a default value and and Maybe value. If the Maybe is Nothing, it returns the default values; otherwise, it returns the value contained in the Maybe.

Data.Monoid

class Monoid a where

The class of monoids (types with an associative binary operation that has an identity). Instances should satisfy the following laws:

mappend mempty x = x

mappend x mempty = x

mappend x (mappend y z) = mappend (mappend x y) z

mconcat = foldr mappend mempty

The method names refer to the monoid of lists under concatenation, but there are many other instances.

Minimal complete definition: mempty and mappend.

Some types can be viewed as a monoid in more than one way, e.g. both addition and multiplication on numbers. In such cases we often define newtypes and make those instances of Monoid, e.g. Sum and Product.

Minimal complete definition

mempty, mappend

Methods

mempty :: a

Identity of mappend

mappend :: a -> a -> a

An associative operation

mconcat :: [a] -> a

Fold a list using the monoid. For most types, the default definition for mconcat will be used, but the function is included in the class definition so that an optimized version can be provided for specific types.

mappend :: Monoid a => a -> a -> a

An associative operation

mconcat :: Monoid a => [a] -> a

Fold a list using the monoid. For most types, the default definition for mconcat will be used, but the function is included in the class definition so that an optimized version can be provided for specific types.

mempty :: Monoid a => a

Identity of mappend

Data.Tuple

swap :: (a, b) -> (b, a)

Swap the components of a pair.

Text.PrettyPrint

data Doc :: *

The abstract type of documents. A Doc represents a *set* of layouts. A Doc with no occurrences of Union or NoDoc represents just one layout.

(<>) :: Doc -> Doc -> Doc infixl 6

Beside. <> is associative, with identity empty.

(<+>) :: Doc -> Doc -> Doc infixl 6

Beside, separated by space, unless one of the arguments is empty. <+> is associative, with identity empty.

($$) :: Doc -> Doc -> Doc infixl 5

Above, except that if the last line of the first argument stops at least one position before the first line of the second begins, these two lines are overlapped. For example:

   text "hi" $$ nest 5 (text "there")

lays out as

   hi   there

rather than

   hi
        there

$$ is associative, with identity empty, and also satisfies

(x $$ y) <> z = x $$ (y <> z), if y non-empty.

nest :: Int -> Doc -> Doc

Nest (or indent) a document by a given number of positions (which may also be negative). nest satisfies the laws:

nest 0 x = x

nest k (nest k' x) = nest (k+k') x

nest k (x <> y) = nest k z <> nest k y

nest k (x $$ y) = nest k x $$ nest k y

nest k empty = empty

x <> nest k y = x <> y, if x non-empty

The side condition on the last law is needed because empty is a left identity for <>.

docEmpty :: Doc

docHang :: Doc -> Int -> Doc -> Doc

docZero :: String -> Doc