Koshucode.Baala.Base.Prelude.Pair

Description

Additional functions for pairs.

Map

mapFst :: (a -> c) -> (a, b) -> (c, b)

Apply function to fst element.

mapSnd :: (b -> c) -> (a, b) -> (a, c)

Apply function to snd element.

consFst :: a -> Map ([a], b)

Cons up an element to fst list.

consSnd :: b -> Map (a, [b])

Cons up an element to snd list.

mapFstTo :: Functor m => (a -> c) -> m (a, b) -> m (c, b)

Apply function to fst element in functor data.

>>> mapFstTo length $ Just ("aa", "bbbb")
Just (2, "bbbb")

Either

right2 :: (Either a b, Either a c) -> Either a (b, c)

Create right pair from either pair.

>>> right2 (Right 'a', Right 'b') :: Either Char (Char, Char)
Right ('a','b')

>>> right2 (Left 'a', Right 'b') :: Either Char (Char, Char)
Left 'a'

right3 :: (Either a b, Either a c, Either a d) -> Either a (b, c, d)

Create right triple from either triple.

right4 :: (Either a b, Either a c, Either a d, Either a e) -> Either a (b, c, d, e)

Create right quadruple from either quadruple.

Monad

sequenceFst :: Monad m => [(m a, b)] -> m [(a, b)]

Monadic sequencing of fst elements.

sequenceSnd :: Monad m => [(a, m b)] -> m [(a, b)]

Monadic sequencing of snd elements.

>>> sequenceSnd [("a", Just "x"), ("b", Just "y")]
Just [("a","x"), ("b","y")]

>>> sequenceSnd [("a", Just "x"), ("b", Nothing)]
Nothing