Koshucode.Baala.Base.Syntax.Line

Description

Text lines delimited by carriage returns or line feeds.

Line

type LineNumber = Int

Line number.

type NumberedLine = (LineNumber, String)

Line with number.

linesCrlfNumbered :: String -> [NumberedLine]

Line number and its content.

linesCrlf :: String -> [String]

Split string into lines. The result strings do not contain carriage returns (\r) and line feeds (\n).

linesFrom :: Show a => a -> [String]

CodeLine

data CodeLine a

Tokens in line.

Constructors

CodeLine

Fields

lineNumber :: LineNumber

Line number, from 1.

lineContent :: String

Line content without newline.

lineTokens :: [a]

Tokens in the line.

type IndentSize = Int

lineIndentPair :: (a -> IndentSize) -> CodeLine a -> (IndentSize, CodeLine a)

CodeRoll

data CodeRoll a

Constructors

CodeRoll

Fields

codeMap :: AbMap (CodeRoll a)

codeInputPt :: CodePt

codeInput :: String

codeOutput :: [a]

codeWords :: WordTable

type WordTable = Map String String

codeRollUp :: AbMap (CodeRoll a) -> CodePiece -> String -> Ab [CodeLine a]

Split source text into CodeLine list.

Split source text into lines by line delimiters (carriage return \r or line feed \n).

Numbering lines from 1. Internally, this is represented as a list of pairs (LineNumber, String).

Tokenize each lines, and put tokens together in CodeLine.

codeUpdate :: String -> a -> Map (CodeRoll a)

Update codeInput and push result element to codeOutput.

codeUpdateWords :: WordTable -> String -> a -> Map (CodeRoll a)

codeChange :: AbMap (CodeRoll a) -> Map (CodeRoll a)

Change mapper of code roll.