Koshucode.Baala.Base.Syntax.Tokenizer

Description

Tokenizer of koshucode.

Source lines

data SourceLine

Constructors

SourceLine

Fields

sourceLineNumber :: Int

Line number, from 1.

sourceLineContent :: String

Line content without newline.

sourceLineTokens :: [Token]

Tokens from line.

sourceLines

Arguments

:: String
Source text in Koshucode.

-> [SourceLine]
Token list per lines.

Split source text into SourceLine list.

Split source code into lines by line delimiters (carriage return \r or line feed \n)

Numbering line numbers. Internally, this is represented as a list of pairs (line#, content).

Tokenize each lines. This is represented as a list of SourceLine line# content tokens.

Tokenizer

tokens

Arguments

:: String
Input text in koshucode

-> [Token]
Token list from the input

Split string into list of tokens.

Result token list does not contain newline characters. Insead of newlines, Line tokens are there at the start of each lines.

>>> tokens "aa\n'bb'\n\"cc\""
[Line (SourceLine 1 "aa"     [Word 0 "aa"]), Word 0 "aa",
 Line (SourceLine 2 "'bb'"   [Word 1 "bb"]), Word 1 "bb",
 Line (SourceLine 3 "\"cc\"" [Word 2 "cc"]), Word 2 "cc"]

Example includes Line, Word, TermN and Space tokens.

>>> tokens "|-- R  /a A0 /b 31"
[Line (SourceLine 1 "|-- R  /a A0 /b 31", [Word 0 "|--", ...]),
 Word 0 "|--", Space 1, Word 0 "R", Space 2,
 TermN ["/a"], Space 1, Word 0 "A0", Space 1,
 TermN ["/b"], Space 1, Word 0 "31"]

untokens :: [Token] -> String

Convert back a token list to a source string.

untoken :: Token -> String

Asterisks

There are three uses of asterisks (*) in koshucode,

*

Single asterisk means multiplication, e.g., 3 * 4 (three times four).

**

Double asterisk leads line comment. Textx from ** to end of line are ignored.

****

Quadruple asterisk (fourfold asterisk) leads caluse comment. Texts from **** to end of clause are ignored. In other words, line staring with **** and following indented lines are ignored.

Triple asterisk *** is double and rest of text. Quintuple (fivefold) asterisk ***** is quadruple and rest of text.

Line comments like this:

 ** aaa bbbbb cc

Clause comments like this:

 **** aaa bbb
      ccccc dddddd
      ee fffff

You can type **** on top of a clause to hide it.