Koshucode.Baala.Base.Token.TokenTree

Description

Tree of tokens

Token tree

type TTree = CodeTree BracketType Token

Tree of tokens.

type NamedTree = Named TTree

Pair of token tree and its name.

type NamedTrees = Named [TTree]

Pair of token trees and its name.

type TTreeTo a = TTree -> a

Convert token tree to something.

type TTreesTo a = [TTree] -> a

Convert list of token tree to sometning.

type TTreeToAb a = TTree -> Ab a

Convert token tree to something, abortable.

type TTreesToAb a = [TTree] -> Ab a

Convert list of token tree to sometning, abortable.

tokenTrees :: [Token] -> Ab [TTree]

Parse tokens with brackets into trees. Blank tokens and comments are excluded.

wrapTrees :: TTreesTo TTree

Wrap trees in group.

Pattern

pattern TermLeaf CodePt TermType TermPath :: CodeTree t Token

pattern TermLeafPath CodePt TermPath :: CodeTree t Token

pattern TermLeafLocal CodePt (Local String) Int [Token] :: CodeTree t Token

pattern TextLeaf TextForm CodePt String :: CodeTree t Token

pattern TextLeafRaw CodePt String :: CodeTree t Token

pattern TextLeafQ CodePt String :: CodeTree t Token

pattern TextLeafQQ CodePt String :: CodeTree t Token

pattern TextLeafKey CodePt String :: CodeTree t Token

Bracket type

data BracketType

There are six types of brackets

Constructors

BracketGroup
Round brackets for grouping: ( E ... )

BracketForm
Round-bar brackets for form with blanks: (| V ... | E ... |)

BracketList
Square brackets for lists: [ C : ... ]

BracketSet
Curely braces for sets: { C : .... }

BracketRel
Curely-bar braces for relations: {| /N : ... | C : ... | C : ... |}

BracketAssn
Double-angle brackets for associations etc.: << /N C .... >>

BracketInterp
Triple-angle brackets for data interpretation: <<< ... /N ... >>>

BracketType
Square-hyphen brackets for data type: [- ... -]

BracketUnknown
Unknown bracket

Divide trees

splitTokensBy

Arguments

:: Pred String
Predicate

-> [Token]
Tokens

-> Either [Token] ([Token], Token, [Token])
Original-tokens or (before-list, the-word, after-list)

Split token list by unquoted word. If token list contains the word, pair of before-list, the-word and after-list is returned. If does not contains the word, original token list is returned.

>>> splitTokensBy (== "|") . tokens $ "b c"
Left [ TText 1 0 "b", TSpace 2 1, TText 3 0 "c" ]

>>> splitTokensBy (== "|") . tokens $ "a | b | c"
Right ( [ TText 1 0 "a", TSpace 2 1 ]
        , TText 3 0 "|"
        , [ TSpace 4 1, TText 5 0 "b", TSpace 6 1
          , TText 7 0 "|", TSpace 8 1, TText 9 0 "c" ] )

divideTreesBy :: String -> TTreesTo [[TTree]]

Divide token trees by quoteless token of given string.

divideTreesByBar :: TTreesTo [[TTree]]

Divide token trees by vertical bar "|".

divideTreesByColon :: TTreesTo [[TTree]]

Divide token trees by colon ":".

divideTreesByEqual :: TTreesTo [[TTree]]

Divide token trees by equal sign "=".

Abortable

abortableTree :: String -> TTreeTo (Map (Ab b))

Same as abortable except for using TTree instead of list of Token.

abortableTrees :: String -> TTreesTo (Map (Ab b))

Same as abortable except for using list of TTree instead of list of Token.

Abbreviation

tt :: String -> Ab [TTree]

Convert text to token trees.

tt1 :: String -> Ab TTree

ttDoc :: TTreesTo Doc

Get Doc value of token trees for pretty printing.

ttPrint :: String -> IO ()