Koshucode.Baala.Core.Assert.Dataset

Description

Dataset as a set of judges.

Dataset is like a bridge of judges and relations. We can get a relation from a dataset, that dataset is build from judges.

Documentation

data Dataset c

Dataset is a set of judges.

dataset :: [Judge c] -> Dataset c

Gather judges into a dataset.

datasetAdd :: [Judge c] -> Dataset c -> Dataset c

Add judges to dataset.

datasetSelect

Arguments

:: (Ord c, CEmpty c)

=> Dataset c
Dataset

-> RelSelect c
Relation selector

Select relation from dataset. If a giving term is not in judges, CEmpty sign is used.