Koshucode.Baala.Core.Content.Literal

Description

Literalizer: Make literal contents from token tree.

Library

Types

type Literalize a = TokenTree -> Ab a

Transform TokenTree to something.

type LitTrees a = [TokenTree] -> Ab a

Transform list of TokenTree to something.

type LitOperators c = [Named (Literalize c -> LitTrees c)]

Functions

litContentBy :: forall c. CContent c => LitOperators c -> Literalize c

Transform TokenTree into internal form of content.

litTermset :: CContent c => Literalize c -> LitTrees [Named c]

Collect term name and content.

litNamedTrees :: LitTrees [Named TokenTree]

Read list of termname and its content.

>>> litNamedTrees . B.tokenTrees . B.tokens $ "/a 'A3 /b 10"
Right [("/a", TreeB 1 [TreeL (TWord 3 0 "'"), TreeL (TWord 4 0 "A3")]),
       ("/b", TreeL (TWord 8 0 "10"))]

Document

Types

litContentBy recognizes the following types.

Boolean

Boolean used for something is hold or unhold. Textual forms: #true, #fasle.

Nil

Nil means that there are no values. i.e., universal negation on the term holds. Textual form is the non-quoted parens: ().

Text

Sequence of characters. Textual forms is chars with apostrophe or doubly-quoted line: 'abc, "abc def".

Decimal

Decimal number. Textual forms is sequence of digits: 100, 99.50, hex AF.

Set

Set is an unordered collection of contents. Duplication among contents is not significant. Textual form is a sequence of contents delimited by colon, enclosed in braces: { 'a : 'b : 'c }.

List

List is an ordered list of contents. Textual form is a sequence of contents delimited by colon, enclosed in square brackets: [ 'abc : 'def ].

Termset

Termset is a set of terms, i.e., a list of named contents. Textual form is a sequence of terms with bar-angles: <| /a 10 /b 20 |>.

Relation

Relation is a set of same-type tuples, Textual form is a sequence of tuples enclosed in bar-braces. The first tuple is a heading of relation, and succeeding tuples are delimited by vertical bar: {| /a : /b | 'A1 : 20 | 'A3 : 40 |}.

Simple data

Prepere some definitions.

>>> :m +Koshucode.Baala.Op.Vanilla.Type
>>> let trees = B.tokenTrees . B.tokens
>>> let lit  = litContentBy [] :: B.TokenTree -> B.Ab VContent
>>> let lits = litList lit . trees

Boolean.

>>> lits "#true : #false"
Right [VBool True, VBool False]

Words.

>>> lits "'a : 'b #sp 'c"
Right [VText "a", VText "b c"]

Decimal.

>>> lits "12.0"
Right [VDec (Decimal (120, 10), 1, False)]

Nil as no ordinary value.

>>> lits "()"
Right [VNil]

Compound data

Set.

>>> lits "{ 'b : 'a : 'a : 'c : 'a }"
Right [VSet [VText "b", VText "a", VText "c"]]

List.

>>> lits "[ 'a : '10 : 20 ]"
Right [VList [VText "a", VText "10", VDec (Decimal (20, 1), 0, False)]]

Termset.

>>> lits "<| /a 'x  /b { 'y : 'z } |>"
Right [VTermset
  [ ("/a", VText "x")
  , ("/b", VSet [VText "y", VText "z"])]]

Relation.

>>> lits "{| /a : /x | 'A1 : 20 | 'A3 : 40 | 'A4 : 60 |}"
Right [VRel (Rel
  (Relhead [Term "/a", Term "/x"]),
  [ [VText "A1", VDec (Decimal (20,1), 0, False)]
  , [VText "A3", VDec (Decimal (40,1), 0, False)]
  , [VText "A4", VDec (Decimal (60,1), 0, False)] ])]