Koshucode.Baala.Core.Lexmap.Attribute

Description

Attributes of relmap operator.

Attribute name

data AttrName

Constructors

AttrTree String

AttrRelmap String

isAttrRelmap :: AttrName -> Bool

Test attribute name is for relmap.

attrNameText :: AttrName -> String

String part of attribute names.

attrNameAttr :: AttrName

Constant for attribute name @attr.

attrNameTrunk :: AttrName

Constant for attribute name @trunk.

Attribute trees

data AttrDefine

Definition of attribute sorter.

Constructors

AttrDefine

Fields

attrTrunkSorter :: AbMap AttrTrees

attrClassifier :: AbMap AttrTrees

attrTrunkNames :: [AttrName]

attrBranchNames :: [AttrName]

type AttrTrees = [(AttrName, [TokenTree])]

List of attribute name and its contents.

type AttrSort = [TokenTree] -> Ab AttrTrees

Sorter for attribute of relmap operator. Sorters docompose attribute trees, and give a name to subattribute.

type TreeSort = [TokenTree] -> [NamedTrees]

type RopName = String

Name of relmap operator.

type Roal a = ((RopName, AttrTrees), a)

Association of operator use and something. Operator use is represented as pair of operator name and attributes.

Attribute sorter

attrSort :: AttrDefine -> AttrSort

attrSortBranch :: AttrSort

hyphenAssc :: TreeSort

Split attribute into named group. Non quoted words beginning with hyphen, e.g., -x, are name of group.

>>> Right . hyphenAssc =<< B.tt "a b -x /c 'd -y e"
[ ("@trunk", [TreeL (TText 1 0 "a"), TreeL (TText 3 0 "b")])
, ("-x", [TreeL (TTerm 7 ["/c"]), TreeL (TText 9 1 "d")])
, ("-y", [TreeL (TText 14 0 "e")]) ]