Koshucode.Baala.Core.Lexmap.Lexmap

Description

Lexical relmap.

Data type

data Lexmap

Intermediate data that represents use of relmap operator. Lexmap is constructed from a list of TokenTree, and generic relmap is constructed from a lexmap.

Constructors

Lexmap

Fields

lexType :: LexmapType

Type of lexmap

lexOpToken :: Token

Token of operator

lexAttr :: AttrTrees

Attribute of relmap operation

lexSubmap :: [Lexmap]

Submaps in the attribute

lexMessage :: [String]

Messages on lexmap

data LexmapType

Constructors

LexmapBase
Built-in relmap

LexmapDerived
User-defined relmap

LexmapWith
-with variable

lexOpName :: Lexmap -> RopName

Name of relmap operator

lexAddMessage :: String -> Map Lexmap

lexMessageList :: Lexmap -> [String]

Constructor

type ConsLexmap = [GlobalSlot] -> [RelmapSource] -> ConsLexmapBody

First step of constructing relmap, make lexmap from source of relmap operator.

type ConsLexmapBody = [TokenTree] -> Ab (Lexmap, [Roal Lexmap])

Construct lexmap and its submaps from source of lexmap

type RelmapSource = Named ([TokenTree], Roamap)

Source of relmap: its name, replacement, and attribute editor.

consLexmap :: [Named AttrSort] -> ConsLexmap

withTerms :: [TokenTree] -> Ab [Terminal String]

Parse -with attribute.