Koshucode.Baala.Core.Relmap.Construct

Description

Relmap construction.

Documentation

relmapCons :: Global c -> RelmapCons c

Make lex and full relmap constructors.

data RelmapCons c

Lex and full relmap constructor

Constructors

RelmapCons RelmapConsLex (RelmapConsFull c)

type RelmapConsLex

Arguments

= [TokenTree]
Source of relmap operator

-> Ab Lexmap
Result lex relmap

First step of constructing relmap, make Lexmap from use of relmap operator.

type RelmapConsFull c

Arguments

= Lexmap
Lexical relmap

-> Ab (Relmap c)
Result full relmap

Second step of constructing relmap, make Relmap from contents of Lexmap.

Construction process

Construction process of lex relmaps from source trees.

[TokenTree] -> [[TokenTree]]

Dicide list of TokenTree by vertical bar (|).

[[TokenTree]] -> [Lexmap]

Construct each Lexmap from lists of TokenTree. When there are subrelmaps in token trees, constructs Lexmap recursively.

[Lexmap] -> Lexmap

Wrap list of Lexmap into one Lexmap that has these relmaps in lexSubrelmap.