Koshucode.Baala.Core.Relmap.Construct

Description

Relmap construction.

Documentation

relmapCons

Arguments

:: [Rop c]
Implementations of relational operators

-> RelmapCons c
Relmap constructors

Make half and full relmap constructors.

data RelmapCons c

Half and full relmap constructor

Constructors

RelmapCons

Fields

consHalf :: RelmapHalfCons

consFull :: RelmapFullCons c

type RelmapHalfCons

Arguments

= [TokenLine]
Source information

-> [TokenTree]
Operand as source trees

-> Ab HalfRelmap
Result half relmap

First step of constructing relmap, make HalfRelmap from use of relational operator.

type RelmapFullCons c

Arguments

= HalfRelmap
Half relmap from RelmapHalfCons

-> AbortTokens (Relmap c)
Result relmap

Second step of constructing relmap, make Relmap from contents of HalfRelmap.

Construction process

Construction process of half relmaps from source trees.

[TokenTree] -> [[TokenTree]]

Dicide list of TokenTree by vertical bar (|).

[[TokenTree]] -> [HalfRelmap]

Construct each HalfRelmap from lists of TokenTree. When there are subrelmaps in token trees, constructs HalfRelmap recursively.

[HalfRelmap] -> HalfRelmap

Wrap list of HalfRelmap into one HalfRelmap that has these relmaps in halfSubmap.