Koshucode.Baala.Core.Relmap.Construct

Description

Construction of relmaps.

Source

relmapConst :: Intmed' h c -> Rel c -> Relmap' h c

Make a constant relmap.

relmapSource :: Intmed' h c -> JudgeClass -> [TermName] -> Relmap' h c

Relmap for retrieving relation from dataset.

Calculation

relmapFlow :: Intmed' h c -> RelkitFlow c -> Relmap' h c

Make a flow relmap. Flow relmaps take no submaps.

relmapHook :: Intmed' h c -> RelkitHook' h c -> Relmap' h c

Relmap with hook data.

relmapBinary :: Intmed' h c -> RelkitBinary c -> Relmap' h c -> Relmap' h c

Make a binary relmap. Binary relmaps take one submap.

relmapConfl :: Intmed' h c -> RelkitConfl c -> [Relmap' h c] -> Relmap' h c

Make a confluent relmap. Confluent relmaps take multiple submaps.

Variable

relmapLink :: Intmed' h c -> Relmap' h c

Link-to-other relmap.

relmapNest :: Intmed' h c -> Map (Relmap' h c)

Reference base of nested relation references. This base is refered by relmapLocalNest.

relmapLocalNest :: Intmed' h c -> TermName -> Relmap' h c

Relation reference for nested relation.

relmapCopy :: Intmed' h c -> RopName -> Map (Relmap' h c)

Reference base of copy-of-input relation reference. This base is refered by relmapLocalSymbol.

relmapLocalSymbol :: Intmed' h c -> RopName -> Relmap' h c

Relation reference for locally introduced relation.

Append relmaps

This picture represents calculation of mapping input relation to output relation.

input -[ relmap ]- output

Relmap A maps relation R1 to relation R2. Another relmap B maps R2 to R3.

R1 -[ A ]- R2
R2 -[ B ]- R3

Two relmaps A and B are jointed with intermidiate relation R2.

R1 -[ A ]- R2 -[ B ]- R3

Or, we can draw a directly jointed picture. This whole structure is also RelmapAppend A B.

R1 -[ A ]--[ B ]- R3

Bundle of constructors

type ConsRelmap' h c = Lexmap -> Ab (Relmap' h c)

Second step of constructing relmap, make relmap from lexmap.

consRelmap :: FindRop' h c -> h c -> ConsRelmap' h c

Construct relmap from lexmap.