Koshucode.Baala.Core.Relmap.Construct

Description

Operations on Relmap.

Generic relmap

type ConsRelmap' h c = Lexmap -> Ab (Relmap' h c)

Second step of constructing relmap, make relmap from lexmap.

relmapCons :: GetGlobal h => h c -> (ConsLexmap, ConsRelmap' h c)

Make a constructor pair of lexmap and relmap.

Constructor

relmapSource :: RopUse' h c -> JudgePat -> [TermName] -> Relmap' h c

Retrieve relation from dataset.

relmapConst :: RopUse' h c -> Rel c -> Relmap' h c

Make a constant relmap.

relmapFlow :: RopUse' h c -> RelkitFlow c -> Relmap' h c

Make a flow relmap. Flow relmaps take no subrelmaps.

relmapHook :: RopUse' h c -> RelkitHook' h c -> Relmap' h c

relmapBinary :: RopUse' h c -> RelkitBinary c -> Relmap' h c -> Relmap' h c

Make a binary relmap. Binary relmaps take one subrelmap.

relmapConfl :: RopUse' h c -> RelkitConfl c -> [Relmap' h c] -> Relmap' h c

Make a confluent relmap. Confluent relmaps take multiple subrelmaps.

relmapCopy :: RopUse' h c -> String -> Map (Relmap' h c)

relmapNest :: RopUse' h c -> [Terminal String] -> Map (Relmap' h c)

relmapNestVar :: RopUse' h c -> String -> Relmap' h c

relmapLink :: RopUse' h c -> Relmap' h c

Select from relmap

relmapSourceList :: Relmap' h c -> [Relmap' h c]

List of RelmapSource

relmapNameList :: Relmap' h c -> [String]

List of name in RelmapLink

Append relmaps

This picture represents calculation of mapping input relation to output relation.

input -[ relmap ]- output

Relmap A maps relation R1 to relation R2. Another relmap B maps R2 to R3.

R1 -[ A ]- R2
R2 -[ B ]- R3

Two relmaps A and B are jointed with intermidiate relation R2.

R1 -[ A ]- R2 -[ B ]- R3

Or, we can draw a directly jointed picture. This whole structure is also RelmapAppend A B.

R1 -[ A ]--[ B ]- R3