Koshucode.Baala.Core.Relmap.Relmap

Description

Data structures for relation-to-relation mappings

Append relmaps

This picture represents calculation of mapping input relation to output relation.

 input -[ relmap ]- output

Relmap A maps relation R1 to relation R2. Another relmap B maps R2 to R3.

 R1 -[ A ]- R2
 R2 -[ B ]- R3

Two relmaps A and B are jointed with intermidiate relation R2.

 R1 -[ A ]- R2 -[ B ]- R3

Or, we can draw a directly jointed picture. This whole structure is also RelmapAppend A B.

 R1 -[ A ]--[ B ]- R3

Constructors

relmapSource :: RopUse c -> JudgePattern -> [Termname] -> Relmap c

Retrieve relation from dataset.

relmapConst :: RopUse c -> Rel c -> Relmap c

Constant relmap.

relmapAlias :: RopUse c -> Relmap c -> Relmap c

Alias for relmap.

relmapCalc :: RopUse c -> RelkitCalc c -> Relmap c

Make a non-confluent relmap.

relmapBinary :: RopUse c -> RelkitBinary c -> Relmap c -> Relmap c

relmapConfl :: RopUse c -> RelkitConfl c -> [Relmap c] -> Relmap c

Make a confluent relmap.

relmapGlobal :: RopUse c -> RelkitGlobal c -> Relmap c

Selectors

relmapSourceList :: Relmap c -> [Relmap c]

List of RelmapSource

relmapNameList :: Relmap c -> [String]

List of name in RelmapName

Linker

relmapLinker

Arguments

:: [Named (Relmap c)]
Relmap and its linking name

-> Relmap c
Relmap before linking

-> Relmap c
Linked relmap

Link relmaps by its name.