Koshucode.Baala.Core.Relmap.Relmap

Description

Data structures for relation-to-relation mappings

Data

data Relmap v

Relation-to-relation mapping. A Relmap is correspond to a use of relational operator.

Constructors

RelmapSource HalfRelmap String [String]
Retrieve a relation from a dataset

RelmapConst HalfRelmap String (Rel v)
Constant relation

RelmapAlias HalfRelmap (Relmap v)
Equavalent relmap

RelmapCalc HalfRelmap String (RelmapSub v) [Relmap v]
Relmap that maps relations to a relation

RelmapAppend (Relmap v) (Relmap v)
Connect two relmaps

RelmapName HalfRelmap String
Relmap reference

type RelmapSub v

Arguments

= [Rel v]
Relations in operand

-> Rel v
Main input relation

-> AbOr (Rel v)
Output relation

Function of relmap.

Append relmaps

This picture represents calculation of mapping input relation to output relation.

 input -[ relmap ]- output

Relmap A maps relation R1 to relation R2. Another relmap B maps R2 to R3.

 R1 -[ A ]- R2
 R2 -[ B ]- R3

Two relmaps A and B are jointed with intermidiate relation R2.

 R1 -[ A ]- R2 -[ B ]- R3

Or, we can draw a directly jointed picture. This whole structure is also RelmapAppend A B.

 R1 -[ A ]--[ B ]- R3

Selectors

relmapSourceList :: Relmap v -> [Relmap v]

List of RelmapSource

relmapNameList :: Relmap v -> [String]

List of name in RelmapName

relmapAppendList :: Relmap v -> [Relmap v]

Expand RelmapAppend to list of Relmap

Linker

relmapLinker

Arguments

:: [Named (Relmap v)]
Relmap and its linking name

-> Relmap v
Relmap before linking

-> Relmap v
Linked relmap

Link relmaps by its name.