Koshucode.Baala.Core.Resource.Resource

Description

Internal representation of relational data resource. Data resource bundles data themselves and calculation stuffs. In Koshucode, data is represented as set of judgements, and calculation is as relation-to-relation mapping. This internal representation is usually constructed from computer files or web resources addressed by URI. Result of calculation on some resource becomes also data resource.

Data type

data Resource c

Relational data resource

Constructors

Resource

Fields

resGlobal :: Global c

Global parameter

resImport :: [Resource c]

Importing resources

resExport :: [String]

Exporting names

resSlot :: [NamedTrees]

Global slots

resRelmap :: [RelmapSource]

Source of relmaps

resAssert :: [ShortAssert c]

Assertions of relmaps

resJudge :: [Judge c]

Affirmative or denial judgements

resArticle :: ([CodeName], [CodeName], [CodePiece])

Codes of resource

resMessage :: [String]

Collection of messages

resLastSecNo :: SecNo

Last section number

resSelect :: RelSelect c

type AbResource c = Ab (Resource c)

Abort or resource.

resEmpty :: Resource c

Resource that has no contents.

resIncluded :: Resource c -> [CodePiece]

addMessage :: String -> Map (Resource c)

addMessages :: [String] -> Map (Resource c)

Hook

type Assert c = Assert' Resource c

type ConsRelmap c = ConsRelmap' Resource c

type Global c = Global' Resource c

type RelkitHook c = RelkitHook' Resource c

type Relmap c = Relmap' Resource c

type RelmapLinkTable c = RelmapLinkTable' Resource c

type Rop c = Rop' Resource c

type RopCons c = RopCons' Resource c

type Intmed c = Intmed' Resource c

type ShortAssert c = ShortAssert' Resource c

type ShortAsserts c = ShortAsserts' Resource c

global :: Global c