Koshucode.Baala.Core.Resource.Resource

Description

Internal representation of relational data resource. Data resource bundles data themselves and calculation stuffs. In Koshucode, data is represented as set of judgements, and calculation is as relation-to-relation mapping. This internal representation is usually constructed from computer files or web resources addressed by URI. Result of calculation on some resource becomes also data resource.

Data type

data Resource c

Relational data resource

Constructors

Resource

Fields

resGlobal :: Global c

Global: Global parameter

resLastSecNo :: SecNo

Global: Last section number

resCacheT :: CacheT Chars

Global: Term name cache

resOption :: Option c

Setting: Options

resExport :: [String]

Setting: Exporting names

resImport :: [Resource c]

Input: Importing resources

resInputQueue :: InputQueue

Input: Input points

resJudge :: [Judge c]

Input: Affirmative or denial judgements

resDataset :: Dataset c

Input: Dataset

resSlot :: [GlobalSlot Chars]

Calc: Global slots

resLexmap :: [LexmapClause]

Calc: Source of relmaps

resAssert :: [ShortAssert c]

Calc: Assertions of relmaps

resOutput :: IOPoint

Output: Output point

resEcho :: [[TokenLine Chars]]

Output: Echo text

resLicense :: [(SecNo, String)]

Output: License text

resMessage :: [String]

Output: Collection of messages

type AbResource c = Ab (Resource c)

Abort or resource.

resIncluded :: Resource c -> [IxIOPoint]

Included sources, i.e., done-part of input queue.

resInput :: Resource c -> [IOPoint]

IOPoint of all inputs.

resInputPoint :: Resource c -> [InputPoint Chars]

All input points.

resClass :: Resource c -> [JudgeClass]

List of all judgement classes.

resFeature :: Resource c -> Feature

Get feature from resource.

resMessagesAdd :: [String] -> Map (Resource c)

Add messages.

Input queue

type InputQueue = (Queue (InputPoint Chars), [IxIOPoint])

Queue for input code: todo, ready and done.

resQueueTodo :: InputPoint Chars -> Map (Resource c)

Add input to todo-part of input queue.

resQueueDone :: IxIOPoint -> Map (Resource c)

Add input to done-part of input queue.

Hook

type Assert c = Assert' Resource c

Assertion of relation.

type ConsRelmap c = ConsRelmap' Resource c

Relmap construction.

type Global c = Global' Resource c

Global parameter.

type RelkitHook c = RelkitHook' Resource c

Make relkit with data resource.

type Relmap c = Relmap' Resource c

Generic relmap.

type RelmapLinkTable c = RelmapLinkTable' Resource c

Lexmap-to-relmap table.

type Rop c = Rop' Resource c

Relmap operator.

type RopCons c = RopCons' Resource c

Constructor of relmap operator.

type Intmed c = Intmed' Resource c

Intermediate relmap.

type ShortAssert c = ShortAssert' Resource c

Assertion with short signs.

type ShortAsserts c = ShortAsserts' Resource c

Assertion list with short signs.

global :: CContent c => Global c

Default global parameter.