Koshucode.Baala.Data.Church.Cox

Description

Term-content expression.

Data types

data Cox c

Term-content expressions.

Constructors

CoxLit [CodePos] c
1. Literal content

CoxTerm [CodePos] [TermName] [TermIndex]
2. Term reference, its name and index

CoxCalc [CodePos] BlankName (CopCalc c)
3. Content calculator

CoxLocal [CodePos] String Int
4. Local blank, its name and De Bruijn index

CoxBlank [CodePos] BlankName
5. Blank in form

CoxFill [CodePos] (Cox c) [Cox c]
6. Fill args in a form

CoxForm1 [CodePos] CoxTag String (Cox c)
7. Form with single blank

CoxForm [CodePos] CoxTag [String] (Cox c)
8. Form with multiple blanks

CoxWith [CodePos] [NamedCox c] (Cox c)
9. Cox with outside arguments

type Cox2 c = (Cox c, Cox c)

Expression pair.

type Cox3 c = (Cox c, Cox c, Cox c)

Expression triple.

type Cox4 c = (Cox c, Cox c, Cox c, Cox c)

Expression quadruple.

type MaybeCox c = Maybe (Cox c)

Maybe content expression.

type AbCox c = Ab (Cox c)

Abortable content expression.

type NamedCox c = Named (Cox c)

Named content expression.

type TermCox c = Term (Cox c)

Term name and content expression.

type CopCalc c = [Ab c] -> Ab c

Term-content calculator.

type CoxTag = Maybe String

Expression tag.

Functions

coxLit :: c -> Cox c

Create literal expression.

coxSyntacticArity :: Cox c -> Int

Get arity of expression.

coxMap :: Map (Map (Cox c))

Map expression form.

coxCall :: Cox c -> Map (Cox c) -> Cox c

Fill expression form.

checkIrreducible :: AbMap (Cox c)

Check irreducible expression.

Process

Phase 1. Convert input string into token trees.

1. String -> [Token@\

@] Parse input string into a list of tokens.

2. [Token@\

-> [TTree]@] Analyze token list structure.

3. [TTree@\

-> @TTree] Wrap the list of token trees in a tree.

Phase 2. Transform token trees, and convert into abstract syntax trees. See Koshucode.Baala.Data.Content.Build module.

4. TTree -> TTree

Expand tree-level syntax operators.

5. TTree -> TTree

Transform binary operators from infix to prefix.

6. TTree -> Cox c

Convert token tree into an abstract syntax tree.

7. Cox c -> Cox c

Attach De Bruijn indecies for bound variables.

8. Cox c -> Cox c

Expand expression-level syntax operators.

Phase 3. Calculate content expressions. See Koshucode.Baala.Data.Content.Run module.

9. Relhead -> Cox c -> Cox c

Attach term positions using actural heading of relation.

10. CopSet c -> [NamedCox@ c\

-> Cox c -> Cox c@] Replace base and derived operators into its bodies.

11. Cox c -> Beta c

Reduce derived expressions into beta expressions.

12. @Beta c -> [c\

-> c@] Run beta expression for each association of relation.