Koshucode.Baala.Data.Content.Tree

Description

Utilities for token trees.

Single content

treeToText :: Bool -> TTreeToAb String

Get text from token tree.

treesToDigits :: TTreesToAb String

Get digits from token trees.

>>> S.tt "-123 450.00" >>= treesToDigits
Right "-123450.00"

tokenClock :: Token -> Ab Clock

Get clock from token.

treesToTime :: TTreesToAb Time

Get time from token trees.

>>> S.tt "2013-04-18 12:00" >>= treesToTime
Right (TimeYmdc (Monthly 2013-04-18) (ClockDhm 0 43200))

>>> S.tt "2013-04-18" >>= treesToTime
Right (TimeYmd (Monthly 2013-04-18))

>>> S.tt "2013-#16" >>= treesToTime
Right (TimeYw 2013-04-15)

treesToInterp :: TTreesToAb Interp

Get interpretation from token trees.

>>> S.tt "term /a" >>= treesToInterp
Right (Interp { interpWords = [InterpText "term", InterpTerm "a"],
                interpTerms = ["a"] })

treeToFlatTerm :: TTreeToAb TermName

Get flat term name from token tree. If the token tree contains nested term name, this function failed.

Multiple contents

treesToTerms :: TTreesToAb [NamedTrees]

Get list of named token trees from token trees.

>>> S.tt "/a 'A3 /b 10" >>= treesToTerms
Right [("a", [TreeL ...]),
       ("b", [TreeL ...])]

treesToTerms1 :: TTreesToAb [NamedTree]

Get list of named token trees from token trees. This function wraps long branches into group.