Koshucode.Baala.Data.Decode.Term

Description

Decode terms.

Decode terms

Term name cache

type CacheT = CacheS TermName

Term name cache.

cacheT :: CacheT

Empty term name cache.

Term name

treeFlatName :: Tree -> Ab TermName

Read flat term name from token tree. If the token tree contains nested term name, this function failed.

>>> S.toTree "/a" >>= treeFlatName
Right (TermName EQ "a")

>>> S.toTree "+/a" >>= treeFlatName
Right (TermName GT "a")

>>> S.toTree "/a/x" >>= treeFlatName
Left ...

treeFlatNameCached :: CacheT -> Tree -> Ab (CacheT, TermName)

Cached version of treeFlatName.

treesFlatNames :: [Tree] -> Ab [TermName]

Read flat term names.

>>> S.toTrees "/a +/b -/c" >>= treesFlatNames
Right [TermName EQ "a",TermName GT "b",TermName LT "c"]

treesFlatNamesCo :: [Tree] -> Ab (Bool, [TermName])

Decode term names with optional complement symbol.

>>> S.toTrees "~ /a /b" >>= treesFlatNamesCo
Right (True, [TermName EQ "a",TermName EQ "b"])

>>> S.toTrees "/a /b" >>= treesFlatNamesCo
Right (False, [TermName EQ "a",TermName EQ "b"])

treesFlatNamePairs :: [Tree] -> Ab [TermName2]

Decode a list of name-and-name pairs.

>>> S.toTrees "/a /x /b /y" >>= treesFlatNamePairs
Right [(TermName EQ "a", TermName EQ "x"), (TermName EQ "b", TermName EQ "y")]

treesNamesByColon :: [Tree] -> Ab [[TermName]]

Decode term names grouped by colons.

>>> S.toTrees "/a /b : /p /q : /x /y" >>= treesNamesByColon
Right [ [TermName EQ "a", TermName EQ "b"]
      , [TermName EQ "p", TermName EQ "q"]
      , [TermName EQ "x", TermName EQ "y"] ]

Term

treesTerms :: [Tree] -> Ab [Term [Tree]]

Read list of named token trees from token trees.

>>> S.toTrees "/a 'A3 /b 10" >>= treesTerms
Right [ (TermName EQ "a", [TreeL ...]),
        (TermName EQ "b", [TreeL ...]) ]

treesTermsCached :: CacheT -> [Tree] -> Ab (CacheT, [Term [Tree]])

Cached version of treesTerms.

treesTerms1 :: [Tree] -> Ab [Term Tree]

Read list of named token trees from token trees. This function wraps long branches into group.