Koshucode.Baala.Data.Type.Judge

Description

Judgements: a symbolic representations of affirmed or denied statements.

Datatype

data Judge c

Judgement on type c.

Judgement (or judge for short) is divided into three parts: logical quality, name of class, and argument. Boolean values True or False of logical quality corresponds to affirmed or denied judge. A name of judgement class represents certain sentence class that gives intepretation of data. Sentence class has placeholders filled by Named c in argument.

Constructors

JudgeAffirm JudgeClass [Term c]
|-- P /x 10 /y 20

JudgeDeny JudgeClass [Term c]
|-x P /x 10 /y 20

JudgeMultiDeny JudgeClass [Term c]
|-xx P /x 10 /y 20

JudgeChange JudgeClass [Term c] [Term c]
|-c P /x 10 +/y 20

JudgeMultiChange JudgeClass [Term c] [Term c]
|-cc P /x 10 +/y 20

JudgeViolate JudgeClass [Term c]
|-v P /x 10 /y 20

judgeTermsMap :: ([Term a] -> [Term b]) -> Judge a -> Judge b

Map function to terms.

judgeCons :: Term c -> Map (Judge c)

Prepend a term into judgement.

Logical quality

type JudgeOf c = JudgeClass -> [Term c] -> Judge c

Construct judgement from its class and terms.

affirm :: JudgeOf c

Construct affirmative judgement.

deny :: JudgeOf c

Construct denial judgement.

affirmJudge :: Map (Judge c)

Affirm judgement, i.e., change logical quality to affirmative.

denyJudge :: Map (Judge c)

Deny judgement, i.e., change logical quality to denial.

isAffirmative :: Judge c -> Bool

Test which judgement is affirmed.

isDenial :: Judge c -> Bool

Test which judgement is denied.

isViolative :: Judge c -> Bool

Test which judgement is for violation.

Encode

judgeBreak :: LineBreak

Conventional line-break setting for judges: 4-spaces indent and 120-columns line.

termNameToMix :: TermName -> MixText

Encode term name.

>>> termNameToMix "foo"
MixText "/foo"

termsToMix1 :: MixShortEncode c => Shorten -> [Term c] -> MixText

Encode term list with one-space separator.

termsToMix2 :: MixShortEncode c => Shorten -> [Term c] -> MixText

Encode term list with two-spaces separator.