Koshucode.Baala.Overture.Shorthand

Description

Shorthand types and functions.

Derived types

type Pair a b = (a, b)

Nominal pair type.

type Twin a = (a, a)

Twin pair.

type Bs = ByteString

Strict bytestring.

type Bz = ByteString

Lazy bytestring.

type Tx = Text

Strict text.

type Tz = Text

Lazy text.

pattern Jp :: t -> t -> Maybe (t, t)

Just pair: Jp a b is equal to Just (a, b).

pattern Jp2 :: t -> t -> t -> Maybe (t, Maybe (t, t))

Double just pairs: Jp2 a b c is equal to Just (a, Just (b, c)).

pattern Jp3 :: t -> t -> t -> t -> Maybe (t, Maybe (t, Maybe (t, t)))

Triple just pairs: Jp3 a b c d is equal to Just (a, Just (b, Just (c, d))).

Functions

ordEq :: Ord a => a -> a -> Bool

Test equality using Ord method. This function can be used for implementing Eq instance.

instance Eq X where
  (==) = ordEq

compareOn :: Ord b => (a -> b) -> a -> a -> Ordering

compareOn f x y comapres f x and f y instead of x and y.

>>> compareOn length "foo" "bar"
EQ

int :: Map Int

Int shorthand.

>>> int 12
12

This is same as:

>>> 12 :: Int
12

integer :: Map Integer

Integer shorthand.

>>> integer 12
12

This is same as:

>>> 12 :: Integer
12

nothing :: a -> Maybe b

Function which always returns Nothing.

>>> nothing True :: Maybe String
Nothing