Koshucode.Baala.Rop.Flat.Subtext

Description

Text decomposition by subtext.

Relmap

type SubtextPara = (TermName, [NameDepth], CharMatch, Bool)

Type of subtext parameters.

consSubtext :: CContent c => RopCons c

subtext /P E

Construct subtext relmap. Decompose text content of /P by matcing with subtext expression E. Terms contained in E are added to output relation. Expression E is either single or bundle:

( E1 )

Decompose text by single expression E.

{ N1 = E1 | N2 = E2 | ... }

Decompose text by bundle of named expressions. Text decomposition starts from N1.

relmapSubtext :: CContent c => Intmed c -> SubtextPara -> Relmap c

Create subtext relmap.

relkitSubtext :: CContent c => SubtextPara -> Maybe Head -> Ab (Relkit c)

Create subtext relkit.

Subtext

Subtext is a domain-specific language for text decomposition. The subtext relmap operator provides the functionality of this language. In the following descrition, we use a symbol E for subtext expression, /P for present term name, /N for new term name, N for symbolic name, T for text, C for character (single-letter text).

Relmap operator

The subtext operator decompose text using subtext expression. There are two ways for describing expression: single or bundle.

subtext /P ( E )

Decompose text content of /P by single expression E.

subtext /P { N1 = E1 | N2 = E2 | ... }

Decompose text by bundle of named expressions. Text decomposition starts from N1.

Basic pattern

?

Any character.

??

Context-dependent pattern.

begin

Beginning of text.

end

End of text.

( E )

Group expression E.

Gathering subtext

/N E

Submatch for E whose name is /N.

/N

Named submatch, equivalent to /N ??.

off E

Turn off gathering text in E.

on E

Turn on gathering text in E.

Text

T

Text literal as T.

char T

Alternative occurences of characters included in T.

word T

Alternative occurences of words included in T.

cat T ...

Characters in the Unicode general categories, T is one-letter or two-letter category name. See categoryLookup.

C1 to C2

Characters betweeen C1 and C2.

sp

Unicode space character or space-like control character including tabs (HT/VT), newlines (CR/LF), or form feed (FF). See isSpace.

012

Unicode number character. See categoryNumber.

digit

ASCII digits. See isDigit.

abc

Unicode letter or mark character. See categoryLetter or categoryMark.

+-

Unicode punctuation or symbol character. See categoryPunctuation or categorySymbol.

open

Open punctuations. See categoryOpen.

close

Close punctuations. See categoryClose.

ascii

ASCII characters. See isAscii.

latin-1

ISO 8859-1 (Latin-1) characters. See isLatin1.

koshu-symbol

Symbol character in Koshucode.

koshu-general

General character in Koshucode.

koshu-plain

Plain character in Koshucode.

koshu-numeric

Numeric character in Koshucode.

Combination

E1 | E2

Alternative occurence of E1 or E2.

E1 ++ E2

Sequence of E1 followed by E2.

E1 or E2

Alternative occurence of E1 or E2.

E1 and E2

E1 and additional condition E2.

not E

Inverted condition.

last E

Find last E.

before E

Sequence of not E.

stay E

Match E but input position is not changed.

Repetition

{ E }

Zero-or-more occurences of E. (Zero-repeat)

{ E } ( N )

N-times occurences of E.

{ E } ( L to )

Repetition of E with lower bound L.

{ E } ( to U )

Repetition of E with upper bound U.

{ E } ( L to U )

Repetition of E with lower bound L and upper bound U.

[ E ]

Zero-or-one occurences of E. (Option)

{- E -}

One-or-more occurences of E. (One-repeat)

E1 sep E2

E1-separated E2. (Zero-repeat)

Modification

E as T

Replace to text T when E is matched.

E as wrap T1 T2

Wrap matched text in T1 and T2.

E as prepend T

Prepend T to matched text.

E as append T

Append T to matched text.

E as lower

Convert text to lower case when E is matched.

E as upper

Convert text to upper case when E is matched.